% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_uhc.R
\name{accelerate_anc4}
\alias{accelerate_anc4}
\title{Accelerate anc4}
\usage{
accelerate_anc4(
  df,
  ind_ids = billion_ind_codes("uhc"),
  scenario_col = "scenario",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{...}{additional parameters to be passed to scenario function}
}
\value{
data frame with acceleration scenario binded to \code{df}. \code{scenario_col} is
set to \code{acceleration}
}
\description{
Accelerate anc4 by first dividing countries into those with reported data and
those without.
\itemize{
\item For countries without reported data, the acceleration scenario_col is the same
as business as usual.
\item For countries with reported data, scenarios with both a \strong{fixed target of 95\%
by 2030} and a \strong{linear change of 2.6 per year till 2025} are tried, with the easiest
to achieve of the two selected. The selected scenario is then compared against
the business as usual scenario for reported data, and the best of the two chosen
as the acceleration scenario.
}
}
