% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_hpop.R
\name{accelerate_child_viol}
\alias{accelerate_child_viol}
\title{Accelerate child_viol}
\usage{
accelerate_child_viol(
  df,
  ind_ids = billion_ind_codes("hpop"),
  end_year = 2025,
  scenario_col = "scenario",
  start_year = 2018,
  value_col = "value",
  default_scenario = "default",
  scenario_name = "acceleration",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{start_year}{Year from which the acceleration scenario begins, inclusive.}

\item{value_col}{Name of the column containing indicator value in \code{df}.}

\item{default_scenario}{name of the default scenario.}

\item{scenario_name}{name of scenario}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate child_viol by targeting 0 by 2030.
}
\details{
Runs:
\itemize{
\item \code{scenario_fixed_target(df, target_value = 0, target_year = 2030, small_is_best = TRUE,...)},
}
}
\seealso{
HPOP acceleration scenarios
\code{\link{accelerate_adult_obese}()},
\code{\link{accelerate_alcohol}()},
\code{\link{accelerate_child_obese}()},
\code{\link{accelerate_devontrack}()},
\code{\link{accelerate_fuel}()},
\code{\link{accelerate_hpop_sanitation_rural}()},
\code{\link{accelerate_hpop_sanitation_urban}()},
\code{\link{accelerate_hpop_sanitation}()},
\code{\link{accelerate_hpop_tobacco}()},
\code{\link{accelerate_ipv}()},
\code{\link{accelerate_overweight}()},
\code{\link{accelerate_pm25}()},
\code{\link{accelerate_road}()},
\code{\link{accelerate_stunting}()},
\code{\link{accelerate_suicide}()},
\code{\link{accelerate_transfats}()},
\code{\link{accelerate_wasting}()},
\code{\link{accelerate_water_rural}()},
\code{\link{accelerate_water_urban}()},
\code{\link{accelerate_water}()}
}
\concept{hpop_acceleration}
