% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_hep.R
\name{accelerate_cholera_campaign}
\alias{accelerate_cholera_campaign}
\title{Accelerate cholera_campaign}
\usage{
accelerate_cholera_campaign(
  df,
  value_col = "value",
  end_year = 2025,
  target_year = end_year,
  start_year = 2018,
  ind_ids = billion_ind_codes("hep"),
  scenario_col = "scenario",
  default_scenario = "default",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{value_col}{Column name of column with indicator values.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{target_year}{Year by which the scenario should eventually be
achieved. Defaults to \code{end_year}}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{default_scenario}{name of the default scenario to be used.}

\item{...}{additional parameters to be passed to scenario function}
}
\value{
data frame with acceleration scenario binded to \code{df}. \code{scenario_col} is
set to \code{acceleration}
}
\description{
Accelerate cholera_campaign by adding planned cholera campaigns to the provided
values in \code{df}. When a value is reported for a year and country, then this
value is kept, even after 2018. Planned values are provided only for the
denominator. For some planned campaigns only the denominator is provided. When
this is the case, the numerator is calculated by taking the best historical
vaccination coverage achieved, or if not available that the best regional
historical coverage.
}
\details{
Planned campaigns are a mix between planned campaigns and the targets outlined
in the \href{https://www.gtfcc.org/about-gtfcc/roadmap-2030/}{roadmap 2030}
of the Global Task Force on Cholera Control.
}
