% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_hpop.R
\name{accelerate_devontrack}
\alias{accelerate_devontrack}
\title{Accelerate devontrack}
\usage{
accelerate_devontrack(
  df,
  ind_ids = billion_ind_codes("hpop"),
  end_year = 2025,
  scenario = "scenario",
  ind = "ind",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{scenario}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{ind}{Column name of column with indicator names.}

\item{...}{additional parameters to to pass to \code{recycle_data}}
}
\value{
data frame with acceleration scenario binded to \code{df}. \code{scenario} is
set to \code{acceleration}
}
\description{
Accelerate devontrack by targeting 80 by 2030.
}
\details{
Runs:
\itemize{
\item \code{scenario_fixed_target(df, target_value = 80, target_year = 2030, small_is_best = FALSE,...)},
}
}
