% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_hep.R
\name{accelerate_detect}
\alias{accelerate_detect}
\alias{accelerate_respond}
\alias{accelerate_notify}
\alias{accelerate_detect_respond}
\title{Accelerate Detect, Notify, Respond, and Detect and Respond}
\usage{
accelerate_detect(
  df,
  ind_ids = billion_ind_codes("hep"),
  scenario_col = "scenario",
  bau_scenario = "historical",
  scenario_name = "acceleration",
  ...
)

accelerate_respond(df, scenario_name = "acceleration", ...)

accelerate_notify(df, scenario_name = "acceleration", ...)

accelerate_detect_respond(df, scenario_name = "acceleration", ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{bau_scenario}{name of scenario to be used for business as usual.
Default is \code{historical}.}

\item{scenario_name}{name of scenario}

\item{...}{additional parameters to be passed to scenario function}
}
\value{
data frame with acceleration scenario binded to \code{df}. \code{scenario_col} is
set to \code{acceleration}
}
\description{
\code{accelerate_detect()}, \code{accelerate_respond()}, \code{accelerate_notify()}, and
\code{accelerate_detect_respond()} accelerate by returning to business as usual
values.
}
\seealso{
HEP acceleration scenarios
\code{\link{accelerate_cholera_campaign}()},
\code{\link{accelerate_espar}()},
\code{\link{accelerate_measles_routine}()},
\code{\link{accelerate_meningitis_campaign}()},
\code{\link{accelerate_polio_routine}()},
\code{\link{accelerate_yellow_fever_campaign}()}
}
\concept{hep_acceleration}
