% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_uhc.R
\name{accelerate_fp}
\alias{accelerate_fp}
\title{Accelerate fp}
\usage{
accelerate_fp(
  df,
  ind_ids = billion_ind_codes("uhc"),
  scenario_col = "scenario",
  value_col = "value",
  default_scenario = "default",
  bau_scenario = "historical",
  scenario_name = "acceleration",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{value_col}{Name of the column containing indicator value in \code{df}.}

\item{default_scenario}{name of the default scenario.}

\item{bau_scenario}{name of scenario to be used for business as usual.
Default is \code{historical}.}

\item{scenario_name}{name of scenario}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate fp by dividing the countries into two groups:
\itemize{
\item For BRN, CYP, FSM, ISL, LUX, and SYC, return business as usual.
\item For all other countries, take the best of business as usual and the quantile
target for quantile_year = 2018 and 5 quantiles (capped by the maximum regional
value in 2018).
}
}
\seealso{
UHC acceleration scenarios
\code{\link{accelerate_anc4}()},
\code{\link{accelerate_art}()},
\code{\link{accelerate_beds}()},
\code{\link{accelerate_bp}()},
\code{\link{accelerate_doctors}()},
\code{\link{accelerate_dtp3}()},
\code{\link{accelerate_fh}()},
\code{\link{accelerate_fpg}()},
\code{\link{accelerate_hwf}()},
\code{\link{accelerate_itn}()},
\code{\link{accelerate_nurses}()},
\code{\link{accelerate_pneumo}()},
\code{\link{accelerate_tb}()},
\code{\link{accelerate_uhc_sanitation}()},
\code{\link{accelerate_uhc_tobacco}()}
}
\concept{uhc_acceleration}
