% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_uhc.R
\name{accelerate_fpg}
\alias{accelerate_fpg}
\title{Accelerate fpg}
\usage{
accelerate_fpg(
  df,
  ind_ids = billion_ind_codes("uhc"),
  scenario_col = "scenario",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate fpg by halting the rise to 2010 value.
}
\seealso{
UHC acceleration scenarios
\code{\link{accelerate_anc4}()},
\code{\link{accelerate_art}()},
\code{\link{accelerate_beds}()},
\code{\link{accelerate_bp}()},
\code{\link{accelerate_doctors}()},
\code{\link{accelerate_dtp3}()},
\code{\link{accelerate_fh}()},
\code{\link{accelerate_fp}()},
\code{\link{accelerate_hwf}()},
\code{\link{accelerate_itn}()},
\code{\link{accelerate_nurses}()},
\code{\link{accelerate_pneumo}()},
\code{\link{accelerate_tb}()},
\code{\link{accelerate_uhc_sanitation}()},
\code{\link{accelerate_uhc_tobacco}()}
}
\concept{uhc_acceleration}
