% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_hpop.R
\name{accelerate_fuel}
\alias{accelerate_fuel}
\title{Accelerate fuel}
\usage{
accelerate_fuel(
  df,
  ind_ids = billion_ind_codes("hpop"),
  scenario_col = "scenario",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{...}{additional parameters to be passed to scenario function}
}
\value{
data frame with acceleration scenario binded to \code{df}. \code{scenario_col} is
set to \code{acceleration}
}
\description{
Accelerate fuel by keeping business as usual for high income countries
(HIC) (according to the World Bank), and by aiming at the best rate of change
between 2013 and 2015 in the region for Lower-middle income countries (LMC),
Low-income countries (LIC), High-income countries(HIC), and for unclassified
countries.
}
\details{
Runs:
\itemize{
\item \code{scenario_bau(df, small_is_best = FALSE,...)} for HIC,
\item \code{scenario_best_in_region(df, target_year = 2018, baseline_year = 2013, small_is_best = FALSE,...)} for other income groups.
}
}
