% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_hpop.R
\name{accelerate_hpop_sanitation_urban}
\alias{accelerate_hpop_sanitation_urban}
\title{Accelerate hpop_sanitation_urban}
\usage{
accelerate_hpop_sanitation_urban(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate hpop_sanitation_urban by aiming at best value in quintile.
}
