% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_hpop.R
\name{accelerate_hpop_tobacco}
\alias{accelerate_hpop_tobacco}
\title{Accelerate hpop_tobacco}
\usage{
accelerate_hpop_tobacco(
  df,
  ind_ids = billion_ind_codes("hpop"),
  scenario_col = "scenario",
  value_col = "value",
  start_year = 2018,
  start_year_trim = start_year + 1,
  end_year = 2025,
  default_scenario = "default",
  bau_scenario = "historical",
  scenario_name = "acceleration",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{value_col}{Name of the column containing indicator value in \code{df}.}

\item{start_year}{Year from which the acceleration scenario begins, inclusive.}

\item{start_year_trim}{(integer) year to start trimming from.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{default_scenario}{name of the default scenario.}

\item{bau_scenario}{name of scenario to be used for business as usual.
Default is \code{historical}.}

\item{scenario_name}{name of scenario}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate hpop_tobacco by picking the best value between business as usual,
halt the rise in 2018, or a custom version of \code{scenario_percent_baseline()}. The
custom function is taking similar parameters to
\code{scenario_percent_baseline()}'s \code{percent_change} = -30, \code{baseline_year} = 2010,
but values are added to the \code{start_year} value, rather than the \code{baseline_year}
values.
}
\details{
Runs:
\itemize{
\item custom scenario_percent_baseline (see above).
\item \code{scenario_bau(df, small_is_best = TRUE,...)}
\item \code{scenario_halt_rise(df, baseline_year= 2018, small_is_best = TRUE,...)}
}

Then picks the best result between the three scenarios.
}
\seealso{
scenario_percent_baseline

HPOP acceleration scenarios
\code{\link{accelerate_adult_obese}()},
\code{\link{accelerate_alcohol}()},
\code{\link{accelerate_child_obese}()},
\code{\link{accelerate_child_viol}()},
\code{\link{accelerate_devontrack}()},
\code{\link{accelerate_fuel}()},
\code{\link{accelerate_hpop_sanitation_rural}()},
\code{\link{accelerate_hpop_sanitation_urban}()},
\code{\link{accelerate_hpop_sanitation}()},
\code{\link{accelerate_ipv}()},
\code{\link{accelerate_overweight}()},
\code{\link{accelerate_pm25}()},
\code{\link{accelerate_road}()},
\code{\link{accelerate_stunting}()},
\code{\link{accelerate_suicide}()},
\code{\link{accelerate_transfats}()},
\code{\link{accelerate_wasting}()},
\code{\link{accelerate_water_rural}()},
\code{\link{accelerate_water_urban}()},
\code{\link{accelerate_water}()}
}
\concept{hpop_acceleration}
