% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_hpop.R
\name{accelerate_ipv}
\alias{accelerate_ipv}
\title{Accelerate ipv}
\usage{
accelerate_ipv(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate ipv by targeting 0 by 2030.
}
\details{
Runs:
\itemize{
\item \code{scenario_fixed_target(df, target_value = 0, target_year = 2030, small_is_best = TRUE,...)},
}
}
