% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_hep.R
\name{accelerate_meningitis_campaign}
\alias{accelerate_meningitis_campaign}
\alias{accelerate_meningitis_routine}
\title{Accelerate meningitis}
\usage{
accelerate_meningitis_campaign(
  df,
  ind_ids = billion_ind_codes("hep"),
  scenario_col = "scenario",
  value_col = "value",
  start_year = 2018,
  end_year = 2025,
  years_best_performance = 2015:2018,
  default_scenario = "default",
  scenario_name = "scenario_name",
  ...
)

accelerate_meningitis_routine(
  df,
  ind_ids = billion_ind_codes("hep"),
  value_col = "value",
  scenario_col = "scenario",
  start_year = 2018,
  default_scenario = "default",
  scenario_name = "acceleration",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{value_col}{Column name of column with indicator values.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{years_best_performance}{vector of years with the years in which the
best performance should be found.}

\item{default_scenario}{name of the default scenario.}

\item{scenario_name}{name of scenario}

\item{...}{additional parameters to be passed to scenario function}
}
\value{
data frame with acceleration scenario binded to \code{df}. \code{scenario_col} is
set to \code{acceleration}
}
\description{
\code{accelerate_meningitis_campaign()} accelerate meningitis_campaign by adding
planned meningitis campaigns to the
provided values in \code{df}. When a value is reported for a year and country, then this
value is kept, even after 2018. Some planned values are provided only for the
denominator. For some planned campaigns only the denominator is provided. When
this is the case, the numerator is calculated by taking the best historical
vaccination coverage achieved, or if not available by taking the best historical
coverage across all countries.
}
\details{
Planned campaigns are the planned campaigns targets provided by WHO technical
programs based on member states planification.

\code{accelerate_meningitis_routine()} accelerate routine vaccination by aiming at
a 90\% 2030, only when value is >= 0 (removes cases where value is absent).

Runs:
\itemize{
\item \verb{scenario_fixed_target_col(df, target_col = "target, target_year = 2025, small_is_best = FALSE, upper_limit = 99)}
}
}
\seealso{
HEP acceleration scenarios
\code{\link{accelerate_cholera_campaign}()},
\code{\link{accelerate_detect}()},
\code{\link{accelerate_espar}()},
\code{\link{accelerate_measles_routine}()},
\code{\link{accelerate_polio_routine}()},
\code{\link{accelerate_yellow_fever_campaign}()}
}
\concept{hep_acceleration}
