% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_hep.R
\name{accelerate_meningitis_routine}
\alias{accelerate_meningitis_routine}
\title{Accelerate meningitis_routine}
\usage{
accelerate_meningitis_routine(
  df,
  ind_ids = billion_ind_codes("hep"),
  value_col = "value",
  scenario_col = "scenario",
  start_year = 2018,
  default_scenario = "default",
  scenario_name = "acceleration",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{value_col}{Column name of column with indicator values.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{default_scenario}{name of the default scenario.}

\item{scenario_name}{name of scenario}

\item{...}{additional parameters to be passed to scenario function}
}
\value{
data frame with acceleration scenario binded to \code{df}. \code{scenario_col} is
set to \code{acceleration}
}
\description{
Accelerate meningitis_routine by aiming at a 90\% 2030, only when value is >= 0
(removes cases where value is absent).
}
\details{
Runs:
\itemize{
\item \verb{scenario_fixed_target_col(df, target_col = "target, target_year = 2025, small_is_best = FALSE, upper_limit = 99)}
}
}
