% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_hpop.R
\name{accelerate_pm25}
\alias{accelerate_pm25}
\title{Accelerate pm25}
\usage{
accelerate_pm25(
  df,
  ind_ids = billion_ind_codes("hpop"),
  scenario_col = "scenario",
  value_col = "value",
  default_scenario = "default",
  bau_scenario = "historical",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{value_col}{Name of the column containing indicator value in \code{df}.}

\item{default_scenario}{name of the default scenario.}

\item{bau_scenario}{name of scenario to be used for business as usual.
Default is \code{historical}.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate pm25 by picking the best value between business as usual, and
linear change of -2\% * 2018 \code{value_col} per year.
}
\details{
Runs:
\itemize{
\item \code{scenario_bau(df, small_is_best = TRUE,...)}
\item \code{scenario_linear_change(df, linear_value = df$value_col[df$"year" == 2018] * -0.02, small_is_best = TRUE,...)}
}

Then picks the best result between the two scenarios.
}
