% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_uhc_target.R
\name{accelerate_target_anc4}
\alias{accelerate_target_anc4}
\title{Accelerate anc4 to 95\% by 2030}
\usage{
accelerate_target_anc4(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\value{
data frame with acceleration scenario binded to \code{df}. \code{scenario} is
set to \code{acceleration}
}
\description{
Accelerate \code{anc4} by aiming at \strong{fixed target of 95\% by 2030}.
}
