% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_target_uhc.R
\name{accelerate_target_art}
\alias{accelerate_target_art}
\title{Accelerate art to 90.25 by 2025}
\usage{
accelerate_target_art(
  df,
  ind_ids = billion_ind_codes("uhc"),
  scenario_col = "scenario",
  start_year = 2018,
  default_scenario = "default",
  bau_scenario = "default",
  scenario_name = "acceleration_target",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{start_year}{Year from which the acceleration scenario begins, inclusive.}

\item{default_scenario}{name of the default scenario.}

\item{bau_scenario}{name of scenario to be used for business as usual.
Default is \code{historical}.}

\item{scenario_name}{name of scenario}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate art by first dividing countries into those with reported data and
those without.
\itemize{
\item For countries without reported data, business as usual is returned.
\item For countries with reported data, the best of business as usual and \strong{fixed
target of 90.25\% by 2025} is chosen.
}
}
\seealso{
Target scenario
\code{\link{accelerate_target_anc4}()},
\code{\link{accelerate_target_beds}()},
\code{\link{accelerate_target_bp}()},
\code{\link{accelerate_target_doctors}()},
\code{\link{accelerate_target_dtp3}()},
\code{\link{accelerate_target_fh}()},
\code{\link{accelerate_target_fpg}()},
\code{\link{accelerate_target_fp}()},
\code{\link{accelerate_target_hwf}()},
\code{\link{accelerate_target_itn}()},
\code{\link{accelerate_target_nurses}()},
\code{\link{accelerate_target_pneumo}()},
\code{\link{accelerate_target_tb}()},
\code{\link{accelerate_target_uhc_sanitation}()},
\code{\link{accelerate_target_uhc_tobacco}()}
}
\concept{accelerate_target}
