% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_target_uhc.R
\name{accelerate_target_beds}
\alias{accelerate_target_beds}
\title{Accelerate beds to 18 by 2025}
\usage{
accelerate_target_beds(
  df,
  ind_ids = billion_ind_codes("uhc"),
  scenario_col = "scenario",
  value_col = "value",
  start_year = 2018,
  default_scenario = "default",
  bau_scenario = "historical",
  scenario_name = "acceleration_target",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{value_col}{Name of the column containing indicator value in \code{df}.}

\item{start_year}{Year from which the acceleration scenario begins, inclusive.}

\item{default_scenario}{name of the default scenario.}

\item{bau_scenario}{name of scenario to be used for business as usual.
Default is \code{historical}.}

\item{scenario_name}{name of scenario}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate beds by first dividing countries into two groups:
\itemize{
\item For countries with 18 or more beds for all years after 2018, business
as usual is returned.
\item For countries which have less than 18 beds for any of the years after 2018 (inclusive),
18 beds is targeted by 2025
}
}
\seealso{
Target scenario
\code{\link{accelerate_target_anc4}()},
\code{\link{accelerate_target_art}()},
\code{\link{accelerate_target_bp}()},
\code{\link{accelerate_target_doctors}()},
\code{\link{accelerate_target_dtp3}()},
\code{\link{accelerate_target_fh}()},
\code{\link{accelerate_target_fpg}()},
\code{\link{accelerate_target_fp}()},
\code{\link{accelerate_target_hwf}()},
\code{\link{accelerate_target_itn}()},
\code{\link{accelerate_target_nurses}()},
\code{\link{accelerate_target_pneumo}()},
\code{\link{accelerate_target_tb}()},
\code{\link{accelerate_target_uhc_sanitation}()},
\code{\link{accelerate_target_uhc_tobacco}()}
}
\concept{accelerate_target}
