% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_target_uhc.R
\name{accelerate_target_dtp3}
\alias{accelerate_target_dtp3}
\title{Accelerate dtp3}
\usage{
accelerate_target_dtp3(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate dtp3 using a customised version of scenario_fixed_target with the
following peculiarities:
\itemize{
\item baseline_year = 2019;
\item the 2020 value is kept identical to the 2019 (baseline) value;
\item the target_year is 2030; and
\item the scenario is then a straight line to the target_value and target_year
\item the target values for each country are provided by the technical program.
}
}
\details{
Runs:
\itemize{
\item \code{scenario_halt_rise(df, baseline_year = 2018, small_is_best = TRUE,...)},
\item \code{scenario_bau(df, small_is_best = TRUE,...)}, and
\item \code{scenario_percent_baseline(df, percent_change = -10, baseline_year = 2010, small_is_best = TRUE, ...)}
}

Then picks the best result between the three scenarios.
}
