% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_target_uhc.R
\name{accelerate_target_fp}
\alias{accelerate_target_fp}
\title{Accelerate fp}
\usage{
accelerate_target_fp(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate fp by returning to business as usual, as there are no globally agreed target.
}
