% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_target_uhc.R
\name{accelerate_target_fpg}
\alias{accelerate_target_fpg}
\title{Accelerate fpg}
\usage{
accelerate_target_fpg(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate fpg by halting the rise to 2010 value.
}
