% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_target_uhc.R
\name{accelerate_target_itn}
\alias{accelerate_target_itn}
\title{Accelerate itn}
\usage{
accelerate_target_itn(
  df,
  ind_ids = billion_ind_codes("uhc"),
  scenario_col = "scenario",
  default_scenario = "default",
  bau_scenario = "historical",
  scenario_name = "acceleration",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{default_scenario}{name of the default scenario.}

\item{bau_scenario}{name of scenario to be used for business as usual.
Default is \code{historical}.}

\item{scenario_name}{name of scenario}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate itn by taking the best of business as usual and a \strong{fixed target of
80 by 2030}.
}
