% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_target_uhc.R
\name{accelerate_target_pneumo}
\alias{accelerate_target_pneumo}
\title{Accelerate pneumo}
\usage{
accelerate_target_pneumo(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate pneumo by taking the best of business as usual and a \strong{fixed target
of 90 by 2025} for countries with two or more data points since 2000. Otherwise,
the business as usual scenario is used.
}
