% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_target_uhc.R
\name{accelerate_target_tb}
\alias{accelerate_target_tb}
\title{Accelerate tb}
\usage{
accelerate_target_tb(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate tb by using a \strong{fixed target of 90 by 2025}.
}
