% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_hpop.R
\name{accelerate_transfats}
\alias{accelerate_transfats}
\title{Accelerate transfats}
\usage{
accelerate_transfats(
  df,
  ind_ids = billion_ind_codes("hpop"),
  scenario_col = "scenario",
  default_scenario = "default",
  scenario_name = "acceleration",
  start_year = 2018,
  start_year_trim = start_year + 1,
  end_year = 2025,
  value_col = "value",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{default_scenario}{name of the default scenario.}

\item{scenario_name}{name of scenario}

\item{start_year}{Year from which the acceleration scenario begins, inclusive.}

\item{start_year_trim}{(integer) year to start trimming from.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{value_col}{Column name of column with indicator values. This column will be
used to return the results.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate transfats by targeting 100 by 2025.
}
\seealso{
HPOP acceleration scenarios
\code{\link{accelerate_adult_obese}()},
\code{\link{accelerate_alcohol}()},
\code{\link{accelerate_child_obese}()},
\code{\link{accelerate_child_viol}()},
\code{\link{accelerate_devontrack}()},
\code{\link{accelerate_fuel}()},
\code{\link{accelerate_hpop_sanitation_rural}()},
\code{\link{accelerate_hpop_sanitation_urban}()},
\code{\link{accelerate_hpop_sanitation}()},
\code{\link{accelerate_hpop_tobacco}()},
\code{\link{accelerate_ipv}()},
\code{\link{accelerate_overweight}()},
\code{\link{accelerate_pm25}()},
\code{\link{accelerate_road}()},
\code{\link{accelerate_stunting}()},
\code{\link{accelerate_suicide}()},
\code{\link{accelerate_wasting}()},
\code{\link{accelerate_water_rural}()},
\code{\link{accelerate_water_urban}()},
\code{\link{accelerate_water}()}
}
\concept{hpop_acceleration}
