% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_uhc.R
\name{accelerate_uhc_sanitation}
\alias{accelerate_uhc_sanitation}
\title{Accelerate uhc_sanitation}
\usage{
accelerate_uhc_sanitation(
  df,
  ind_ids = billion_ind_codes("uhc"),
  scenario_col = "scenario",
  default_scenario = "default",
  scenario_name = "acceleration",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{default_scenario}{name of the default scenario.}

\item{scenario_name}{name of scenario}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate uhc_sanitation by encouraging the country to reach the mean (or upper
threshold) of the quantile it belongs to in 2017, with n = 5 quantiles. Lower
and upper limits of 0 and 99, respectively, are also imposed on the results.
}
\seealso{
UHC acceleration scenarios
\code{\link{accelerate_anc4}()},
\code{\link{accelerate_art}()},
\code{\link{accelerate_beds}()},
\code{\link{accelerate_bp}()},
\code{\link{accelerate_doctors}()},
\code{\link{accelerate_dtp3}()},
\code{\link{accelerate_fh}()},
\code{\link{accelerate_fpg}()},
\code{\link{accelerate_fp}()},
\code{\link{accelerate_hwf}()},
\code{\link{accelerate_itn}()},
\code{\link{accelerate_nurses}()},
\code{\link{accelerate_pneumo}()},
\code{\link{accelerate_tb}()},
\code{\link{accelerate_uhc_tobacco}()}
}
\concept{uhc_acceleration}
