% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_hpop.R
\name{accelerate_water_rural}
\alias{accelerate_water_rural}
\title{Accelerate water_rural}
\usage{
accelerate_water_rural(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate water_rural by aiming at best value in quintile.
}
