% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_hep.R
\name{accelerate_yellow_fever_routine}
\alias{accelerate_yellow_fever_routine}
\title{Accelerate yellow_fever_routine}
\usage{
accelerate_yellow_fever_routine(df, scenario_name = "acceleration", ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{scenario_name}{name of scenario}

\item{...}{additional parameters to be passed to scenario function}
}
\value{
data frame with acceleration scenario binded to \code{df}. \code{scenario_col} is
set to \code{acceleration}
}
\description{
Accelerate yellow_fever_routine by aiming at a +20\% percent change between 2015 and 2025
AROC.
}
\details{
Runs:
\itemize{
\item \code{scenario_aroc(df, aroc_type = "percent_change", percent_change = 20, baseline_year = 2015, target_year = 2025, small_is_best = FALSE)}
}
}
