% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpop.R
\name{add_hpop_populations}
\alias{add_hpop_populations}
\title{Add Population Figures for HPOP Billion}
\usage{
add_hpop_populations(
  df,
  iso3 = "iso3",
  ind = "ind",
  ind_ids = billion_ind_codes("hpop")
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{ind}{Column name of column with indicator names.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
Data frame in long format.
}
\description{
\code{add_hpop_populations()} adds relevant populations to each HPOP Billion indicator
and country, so these can be used to calculate indicator-level contributions
to the HPOP Billion.
}
\details{
For more details on the HPOP Billion calculation process and how this function
ties in with the rest, see the vignette:

\href{../doc/hpop.html}{\code{vignette("hpop", package = "billionaiRe")}}
}
