% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_populations.R
\name{add_populations}
\alias{add_populations}
\title{Add Population Figures for all billions}
\usage{
add_populations(
  df,
  population = "population",
  pop_year = 2025,
  scenario_col = NULL,
  transform_value_col = "transform_value",
  ind_ids = billion_ind_codes("all", include_calculated = TRUE)
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{population}{Column name of column to create with population figures.}

\item{pop_year}{Year used to pull in HPOP populations, defaults to 2025.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{transform_value_col}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
Data frame in long format.
}
\description{
\code{add_populations()} adds relevant populations to each indicator and country,
so these can be used to calculate indicator-level aggregations of the
billions. The column specified by \code{population} will be generated and
filled with relevant populations for that country and indicator. If the column
already exists, only missing values will be replaced by the function.
}
\details{
Essentially, the function wraps around \code{add_hep_populations()} and
\code{add_hpop_populations()} and add the country population to UHC indicators.

As HEP indicators where population is relevant are all generated by billionaiRe,
the \code{transform_value_col} is required.
}
