% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_scenarios.R
\name{add_scenario}
\alias{add_scenario}
\title{Add scenario to data frame}
\usage{
add_scenario(
  df,
  scenario_function = c("aroc", "halt_rise", "percent_baseline",
    "linear_percent_change", "linear_percent_change_col", "quantile", "best_in_region",
    "fixed_target", "fixed_target_col", "accelerate"),
  ind_ids = billion_ind_codes("all"),
  ind = "ind",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{scenario_function}{character identifier of scenario to use. Can be any
of the following:
\itemize{
\item \code{aroc}: calls \code{\link{scenario_aroc}}
\item \code{halt_rise}: calls \code{\link{scenario_halt_rise}}
\item \code{percent_baseline}: calls \code{\link{scenario_percent_baseline}}
\item \code{linear_percent_change}: calls \code{\link{scenario_linear_percent_change}}
\item \code{linear_percent_change_col}: calls \code{\link{scenario_linear_percent_change_col}}
\item \code{quantile}: calls \code{\link{scenario_quantile}}
\item \code{best_in_region}: calls \code{\link{scenario_best_in_region}}
\item \code{fixed_target}: calls \code{\link{scenario_fixed_target}}
\item \code{fixed_target_col}: calls \code{\link{scenario_fixed_target_col}}
\item \code{accelerate}: calls indicator accelerate function.
}}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{ind}{Column name of column with indicator names.}

\item{...}{additional arguments passed to \code{add_scenario_indicator}}
}
\value{
data frame with additional rows with scenario values.
}
\description{
\code{add_scenario} wraps around \code{add_scenario_indicator} to add the scenario
specified in \code{scenario_function} to all the valid billionaiRe indicators
present in \code{df}.
}
