% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios_hpop.R
\name{add_scenario_adult_obese}
\alias{add_scenario_adult_obese}
\title{Add scenario to adult obese indicator}
\usage{
add_scenario_adult_obese(
  df,
  scenario_function,
  ind = "ind",
  small_is_best = TRUE,
  ind_ids = billion_ind_codes("hpop"),
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{scenario_function}{character identifier of scenario to use. Can be any
of the following:
\itemize{
\item \code{aroc}: calls \code{\link{scenario_aroc}}
\item \code{halt_rise}: calls \code{\link{scenario_halt_rise}}
\item \code{percent_baseline}: calls \code{\link{scenario_percent_baseline}}
\item \code{linear_percent_change}: calls \code{\link{scenario_linear_percent_change}}
\item \code{linear_percent_change_col}: calls \code{\link{scenario_linear_percent_change_col}}
\item \code{quantile}: calls \code{\link{scenario_quantile}}
\item \code{best_in_region}: calls \code{\link{scenario_best_in_region}}
\item \code{fixed_target}: calls \code{\link{scenario_fixed_target}}
\item \code{fixed_target_col}: calls \code{\link{scenario_fixed_target_col}}
}}

\item{ind}{Column name of column with indicator names.}

\item{small_is_best}{Logical to identify if a lower value is better than a higher
one (e.g. lower obesity in a positive public health outcome, so obesity rate
should have small_is_best = TRUE).}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{...}{additional parameters pass}
}
\value{
data frame with additional rows with scenario values.
}
\description{
Essentially passes \code{...} values to \code{add_scenario_dispatch} while forcing
\code{small_is_best} = TRUE.
}
