% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_scenarios.R
\name{add_scenario_dispatch}
\alias{add_scenario_dispatch}
\title{Add scenario to the data frame}
\usage{
add_scenario_dispatch(
  df,
  scenario_function = c("aroc", "halt_rise", "percent_baseline",
    "linear_percent_change", "linear_percent_change_col", "quantile", "best_in_region",
    "fixed_target", "fixed_target_col"),
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{scenario_function}{character identifier of scenario to use. Can be any
of the following:
\itemize{
\item \code{aroc}: calls \code{\link{scenario_aroc}}
\item \code{halt_rise}: calls \code{\link{scenario_halt_rise}}
\item \code{percent_baseline}: calls \code{\link{scenario_percent_baseline}}
\item \code{linear_percent_change}: calls \code{\link{scenario_linear_percent_change}}
\item \code{linear_percent_change_col}: calls \code{\link{scenario_linear_percent_change_col}}
\item \code{quantile}: calls \code{\link{scenario_quantile}}
\item \code{best_in_region}: calls \code{\link{scenario_best_in_region}}
\item \code{fixed_target}: calls \code{\link{scenario_fixed_target}}
\item \code{fixed_target_col}: calls \code{\link{scenario_fixed_target_col}}
\item \code{accelerate}: calls indicator accelerate function.
}}

\item{...}{additional parameters passed to \code{scenario_function}}
}
\value{
data frame with additional rows with scenario values.
}
\description{
\code{add_scenario_dispatch} adds the right scenario_function
}
