% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_scenarios.R
\name{add_scenario_indicator}
\alias{add_scenario_indicator}
\title{Add scenario for a specific indicator}
\usage{
add_scenario_indicator(
  df,
  scenario_function,
  indicator,
  ind_ids = billion_ind_codes("all"),
  start_year = 2018,
  end_year = 2025,
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{scenario_function}{character identifier of scenario to use. Can be any
of the following:
\itemize{
\item \code{aroc}: calls \code{\link{scenario_aroc}}
\item \code{halt_rise}: calls \code{\link{scenario_halt_rise}}
\item \code{percent_baseline}: calls \code{\link{scenario_percent_baseline}}
\item \code{linear_percent_change}: calls \code{\link{scenario_linear_percent_change}}
\item \code{linear_percent_change_col}: calls \code{\link{scenario_linear_percent_change_col}}
\item \code{quantile}: calls \code{\link{scenario_quantile}}
\item \code{best_in_region}: calls \code{\link{scenario_best_in_region}}
\item \code{fixed_target}: calls \code{\link{scenario_fixed_target}}
\item \code{fixed_target_col}: calls \code{\link{scenario_fixed_target_col}}
\item \code{accelerate}: calls indicator accelerate function.
}}

\item{indicator}{name of indicator to be passed to \code{scenario_function}}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{...}{additional parameters to be passed to the add_scenario_indicator
function (e.g. \code{add_scenario_adult_obese}).}
}
\description{
\code{add_scenario_indicator} gets the right \code{add_scenario_} function for the
indicator specified and adds the scenario.
}
