% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_style.R
\name{add_style_wrapper}
\alias{add_style_wrapper}
\title{Generate a style with variable color or bolding}
\usage{
add_style_wrapper(wb, b_sheet, rows, cols, fontColour, textDecoration)
}
\arguments{
\item{b_sheet}{}

\item{rows}{Rows to apply style to, passed to \code{\link[openxlsx:add_style]{openxlsx::add_style()}}.}

\item{cols}{Columns to apply style to, passed to \code{\link[openxlsx:add_style]{openxlsx::add_style()}}.}

\item{fontColour}{Font colour, passed to \code{\link[openxlsx:createStyle]{openxlsx::createStyle()}}.}

\item{textDecoration}{Font decoration, passed to \code{\link[openxlsx:createStyle]{openxlsx::createStyle()}}.}
}
\description{
\code{add_style_wrapper()} adds a style for use within \code{\link[=timeseries_style]{timeseries_style()}},
where the font color and decoration needs to vary based on data type. Wraps
around \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}} and \code{\link[openxlsx:createStyle]{openxlsx::createStyle()}}.
}
