% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_indicator_list.R
\name{add_style_wrapper_billion}
\alias{add_style_wrapper_billion}
\title{Generate a style with variable color or bolding}
\usage{
add_style_wrapper_billion(wb, sheet_name, rows, cols, fgFill)
}
\arguments{
\item{wb}{\code{\link[openxlsx:createWorkbook]{openxlsx::createWorkbook()}} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{rows}{Rows to apply style to, passed to \code{openxlsx::add_style()}.}

\item{cols}{Columns to apply style to, passed to \code{openxlsx::add_style()}.}

\item{fgFill}{Background color, passed to \code{\link[openxlsx:createStyle]{openxlsx::createStyle()}}.}
}
\description{
\code{add_style_wrapper()} adds a style for use within \code{\link[=timeseries_style]{timeseries_style()}},
where the font color and decoration needs to vary based on data type. Wraps
around \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}} and \code{\link[openxlsx:createStyle]{openxlsx::createStyle()}}.
}
