% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_class}
\alias{assert_class}
\title{Assert that an object is (or is not) of a given (range of) class(es)}
\usage{
assert_class(x, expected, reverse = FALSE, how = c("any", "all"))
}
\arguments{
\item{x}{The input object}

\item{expected}{(character) The expected class(es) of x}

\item{reverse}{Invert the test (i.e., the class of x is not)}

\item{how}{One of "any" and "all". When \code{expected} is a vector, and
\itemize{
\item \code{how = "any"}, the test is passed if \code{x} inherits from any of the elements
of \code{expected}
\item \code{how = "all"}, the test is passed only is \code{x} inherits from all the elements
of \code{expected}.
}}
}
\description{
Assert that an object is (or is not) of a given (range of) class(es)
}
