% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_col_paired_with}
\alias{assert_col_paired_with}
\title{Assert that paired columns have values}
\usage{
assert_col_paired_with(df, col_name, pair_cols)
}
\arguments{
\item{df}{a data frame}

\item{col_name}{(string) the name of a column in the data frame}

\item{pair_cols}{(character vectors) the names of the columns which are paired
with the \code{col_name} column}
}
\description{
Ensures that if rows of a given column has a value (i.e., is not NA), then
other columns in those rows also have a value (i.e., are not NA). Useful for
enforcing dependencies between columns.
}
