% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_col_types}
\alias{assert_col_types}
\title{Assert that the column types of a data frame are as expected}
\usage{
assert_col_types(df, expected)
}
\arguments{
\item{df}{a data frame}

\item{expected}{a \strong{named vector}, whose names are the names of the columns (i.e.,
\code{names(df) == names(expected)}) and whose elements are the expected class/type
of the column.}
}
\description{
Assert that the column types of a data frame are as expected
}
