% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_columns}
\alias{assert_columns}
\title{Assert that columns exist in a data frame}
\usage{
assert_columns(df, ...)
}
\arguments{
\item{df}{Data frame}

\item{...}{Column names}
}
\description{
Assert that columns exist in a data frame
}
