% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_distinct_rows}
\alias{assert_distinct_rows}
\title{Assert unique rows by key columns}
\usage{
assert_distinct_rows(df, key_cols)
}
\arguments{
\item{df}{a data frame}

\item{key_cols}{(character vector) the names of the key columns}
}
\description{
Ensures that all the rows of a given data frame are unique for each combination
of a set of key columns.
}
