% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_equals}
\alias{assert_equals}
\title{Assert that x and y are (or are not) equal/identical}
\usage{
assert_equals(x, y, identical = FALSE, reverse = FALSE, msg_suffix = NULL)
}
\arguments{
\item{x}{(vector)}

\item{y}{(vector)}

\item{identical}{(logical) whether to use the \code{identical()} function for the test}

\item{reverse}{(logical) whether to reverse the condition (i.e., the two vectors are
not equal/identical)}

\item{msg_suffix}{(string) A string to be appended to the end of the error message}
}
\description{
Assert that x and y are (or are not) equal/identical
}
