% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_fileext}
\alias{assert_fileext}
\title{Assert that a file's extension is one of a few options}
\usage{
assert_fileext(file_names, valid_exts)
}
\arguments{
\item{file_names}{(character vector) The file names.}

\item{valid_exts}{(character vector) A list of the valid extensions.}
}
\description{
Assert that a file's extension is one of a few options
}
