% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_homogeneous_col}
\alias{assert_homogeneous_col}
\title{Assert that the data frame only has one value (is homogeneous) for a given column}
\usage{
assert_homogeneous_col(df, col_name)
}
\arguments{
\item{df}{input data frame}

\item{col_name}{string specifying the column to check}
}
\description{
Assert that the data frame only has one value (is homogeneous) for a given column
}
