% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_in_list_or_null}
\alias{assert_in_list_or_null}
\title{Assert that x is in list or is NULL}
\usage{
assert_in_list_or_null(x, list)
}
\description{
@param x value to be checked
@param list list of values to be checked against
}
