% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_ind_ids}
\alias{assert_ind_ids}
\title{Assert that ind_ids is the correct named vector}
\usage{
assert_ind_ids(ind_ids, billion)
}
\arguments{
\item{ind_ids}{Indicator ids to check}

\item{billion}{Billion which we're checking for}
}
\description{
Assert that ind_ids is the correct named vector
}
