% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_iso3_not_empty}
\alias{assert_iso3_not_empty}
\title{Asserts that iso3 (and scenario if provided) is not only NAs}
\usage{
assert_iso3_not_empty(df, scenario_col = NULL, value_col = "value")
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{value_col}{Column name of column with indicator values.}
}
\description{
Asserts that iso3 (and scenario if provided) is not only NAs
}
