% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_length}
\alias{assert_length}
\title{Assert that a vector is of length n}
\usage{
assert_length(x, n)
}
\arguments{
\item{x}{(vector)}

\item{n}{(integer) the expected length of x}
}
\description{
Assert that a vector is of length n
}
