% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_mart_columns}
\alias{assert_mart_columns}
\title{Assert that xMart4 columns for billions exist in data frame}
\usage{
assert_mart_columns(
  df,
  mart_cols = c("iso3", "year", "ind", "upload_date", "value", "lower", "upper",
    "use_dash", "use_calc", "source", "type", "type_detail", "other_detail",
    "upload_detail")
)
}
\arguments{
\item{df}{Data frame following structure exported by \code{billionaiRe::load_billion_data()}}

\item{mart_cols}{Name of xmart4 columns to be checked}
}
\description{
Assert that xMart4 columns for billions exist in data frame
}
