% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_min_length}
\alias{assert_min_length}
\title{Assert that a vector has a minimum length n}
\usage{
assert_min_length(x, n)
}
\arguments{
\item{x}{(vector)}

\item{n}{(integer) the minimum allowed size of the vector}
}
\description{
Assert that a vector has a minimum length n
}
