% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_numeric}
\alias{assert_numeric}
\title{Assert that the given data frame columns are numeric}
\usage{
assert_numeric(df, ...)
}
\arguments{
\item{df}{Data frame}

\item{...}{Column names}
}
\description{
Assert that the given data frame columns are numeric
}
