% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_same_length}
\alias{assert_same_length}
\title{Assert that two vectors are the same length}
\usage{
assert_same_length(...)
}
\arguments{
\item{...}{Arguments to pass two vectors that should be the same length.}
}
\description{
Assert that two vectors are the same length
}
