% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_start_end_year}
\alias{assert_start_end_year}
\title{Asserts start and end year are present}
\usage{
assert_start_end_year(
  df,
  value_col = "value",
  start_year = 2018,
  end_year = 2025,
  scenario_col = "scenario"
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{value_col}{Column name of column with indicator values.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}
}
\description{
Asserts that there are values at the start and end year by iso3 (and
scenarios if provided).
}
