% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_string}
\alias{assert_string}
\title{Assert that \code{x} is a character vector of length n}
\usage{
assert_string(x, n)
}
\arguments{
\item{x}{Supposed string to test}

\item{n}{Length to test}
}
\description{
Assert that \code{x} is a character vector of length n
}
