% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_string_l1}
\alias{assert_string_l1}
\title{Assert that \code{x} is a character vector of length 1}
\usage{
assert_string_l1(x)
}
\arguments{
\item{x}{Supposed string to test}
}
\description{
Assert that \code{x} is a character vector of length 1
}
