% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_strings}
\alias{assert_strings}
\title{Assert that arguments passed in are length 1 character vectors}
\usage{
assert_strings(...)
}
\arguments{
\item{...}{Character vectors to check}
}
\description{
Assert that arguments passed in are length 1 character vectors
}
