% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_style_param}
\alias{assert_style_param}
\title{Assert that \code{params} are valid formal argument to \code{\link[openxlsx:createStyle]{openxlsx::createStyle()}}}
\usage{
assert_style_param(...)
}
\arguments{
\item{...}{character vector of parameters to \code{\link[openxlsx:createStyle]{openxlsx::createStyle()}}}
}
\description{
Assert that \code{params} are valid formal argument to \code{\link[openxlsx:createStyle]{openxlsx::createStyle()}}
}
