% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_summarize}
\alias{assert_summarize}
\title{Assert that summarize_hpop_country_data output data are present}
\usage{
assert_summarize(
  df,
  dfs = c("ind_df", "latest_reported", "baseline_proj", "hpop_contrib", "hpop_billion",
    "df_iso", "transformed_time_series")
)
}
\arguments{
\item{df}{Data frame following structure exported by \code{summarize_hpop_country_data()}}

\item{dfs}{character vector with names of data frame to be found in output}
}
\description{
Assert that summarize_hpop_country_data output data are present
}
