% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_timestamp}
\alias{assert_timestamp}
\title{Assert that x is a valid timestamp string}
\usage{
assert_timestamp(x)
}
\arguments{
\item{x}{(string)}
}
\description{
Assert that x is a valid timestamp string
}
