% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_type}
\alias{assert_type}
\title{Assert that an object is of a given type}
\usage{
assert_type(x, expected_type)
}
\arguments{
\item{x}{The input object}

\item{expected_type}{The expected type of x}
}
\description{
Assert that an object is of a given type
}
