% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_unique_rows}
\alias{assert_unique_rows}
\title{Assert unique rows of df}
\usage{
assert_unique_rows(df, ind, iso3, year, ind_ids)
}
\description{
Makes sure there are distinct rows for each ind, iso3, and year
}
