% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_unique_vector}
\alias{assert_unique_vector}
\title{Assert that the elements of the vector are unique}
\usage{
assert_unique_vector(x)
}
\arguments{
\item{x}{(vector)}
}
\description{
Assert that the elements of the vector are unique
}
