% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_styles_modify.R
\name{assert_valid_style_parameters}
\alias{assert_valid_style_parameters}
\title{Assert that style parameters are valid}
\usage{
assert_valid_style_parameters(...)
}
\arguments{
\item{...}{list of parameters to be passed to \code{\link[openxlsx:createStyle]{openxlsx::createStyle()}}}
}
\description{
Assert that style parameters are valid
}
