% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_who_iso3}
\alias{assert_who_iso3}
\title{Asserts that provided ISO is valid}
\usage{
assert_who_iso3(iso3)
}
\arguments{
\item{iso3}{Single ISO3 code}
}
\description{
Checks that provided ISO code is a valid ISO3 code for a WHO member state,
using \code{\link[whoville:is_who_member]{whoville::is_who_member()}}.
}
