% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_x_in_y}
\alias{assert_x_in_y}
\title{Assert that all elements in x are members of y}
\usage{
assert_x_in_y(x, y)
}
\arguments{
\item{x}{(vector)}

\item{y}{(vector)}
}
\description{
In other words, assert that x is a subset of y. Useful for ensuring that an
argument is one of a given set of options.
}
