% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_years}
\alias{assert_years}
\title{Assert that end years are always later than start year}
\usage{
assert_years(start_year, end_year)
}
\arguments{
\item{start_year}{Start year}

\item{end_year}{End year(s)}
}
\description{
Assert that end years are always later than start year
}
