% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking_uhc.R
\name{benchmarking_bp}
\alias{benchmarking_bp}
\title{Accelerate bp to benchmarking scenario}
\usage{
benchmarking_bp(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate \code{bp} by aiming at 100 by 2030 if there are 2 values or more
reported. Otherwise, business as usual.
}
