% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking_uhc.R
\name{benchmarking_dtp3}
\alias{benchmarking_dtp3}
\title{Accelerate dtp3 to benchmarking scenario}
\usage{
benchmarking_dtp3(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate dtp3 to benchmarking scenario
}
