% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking_uhc.R
\name{benchmarking_fpg}
\alias{benchmarking_fpg}
\title{Accelerate fpg to benchmarking scenario}
\usage{
benchmarking_fpg(df, start_year = 2018, scenario_col = "scenario", ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{start_year}{Year from which the acceleration scenario begins, inclusive.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate fpg to benchmarking scenario
}
