% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking_uhc.R
\name{benchmarking_itn}
\alias{benchmarking_itn}
\title{Accelerate itn to benchmarking scenarios}
\usage{
benchmarking_itn(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate itn to benchmarking scenarios
}
