% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking_uhc.R
\name{benchmarking_uhc_tobacco}
\alias{benchmarking_uhc_tobacco}
\title{Accelerate uhc_tobacco to benchmarking scenario}
\usage{
benchmarking_uhc_tobacco(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate uhc_tobacco to benchmarking scenario
}
