% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_uhc_billion.R
\name{billion_group_mean}
\alias{billion_group_mean}
\title{Calculate average per Billion group}
\usage{
billion_group_mean(ind, transform_value, ind_ids)
}
\arguments{
\item{ind}{Vector of inds}

\item{transform_value}{Vector of transformed values}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\description{
Calculates average per Billion group, taking into account number of
indicators available for use in the calculation. It returns NA (and thus NA when
calculated as ASC) if any indicator is missing. The only exception is if
insecticide treated nets is missing.
}
\keyword{internal}
