% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_columns.R
\name{billionaiRe_add_columns}
\alias{billionaiRe_add_columns}
\title{Add columns to data frame if not already existing}
\usage{
billionaiRe_add_columns(df, cols, fill)
}
\arguments{
\item{df}{A data frame.}

\item{cols}{A character vector of column names to add.}

\item{fill}{A list of fill to be used for each column. If length 1, same fill
is used for each column. Can also be a vector, but lists allow different types
of fill, e.g. \code{fill = list(NA_character_, NA_real_)}.}
}
\value{
A data frame.
}
\description{
The function takes in a data frame, a vector of column names,
and a vector of base fill to use for the column. If the column doesn't already
exist, it is added to the data frame and filled with the relevant fill value.
}
\keyword{internal}
