% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scenarios.R
\name{calculate_aarc}
\alias{calculate_aarc}
\title{Calcualte Average annual rate of change}
\usage{
calculate_aarc(baseline_year, baseline_value, end_year, end_value)
}
\arguments{
\item{baseline_year}{year where the AARC starts}

\item{baseline_value}{value at start_year}

\item{end_year}{year where the AARC ends}

\item{end_value}{value at end_year}
}
\value{
numeric with AARC
}
\description{
Calculates annual average rate of change (AARC or AROC) with the compounded
formula:
((\code{end_value} / \code{baseline_value}))^(1/(\code{end_year} - \code{baseline_year})) - 1
}
