% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scenarios.R
\name{calculate_aroc}
\alias{calculate_aroc}
\title{Calculate Average rate of change}
\usage{
calculate_aroc(baseline_year, baseline_value, end_year, end_value)
}
\arguments{
\item{baseline_year}{year where the AARC starts}

\item{baseline_value}{value at start_year}

\item{end_year}{year where the AARC ends}

\item{end_value}{value at end_year}
}
\value{
numeric with AARC
}
\description{
Calculates Average rate of change (AARC or AROC) \emph{without} the compounded
formula:
(\code{end_value} / \code{baseline_value})/(\code{end_year} - \code{baseline_year})
}
\details{
This formula provides an absolute value change and not a proportional one
like \code{calculate_aarc}. To forecast change it should be add to the baseline
value:
\code{baseline_value} + (\code{aroc}*(\code{current_year} - \code{start_year}))
}
