% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios_target.R
\name{calculate_fixed_target}
\alias{calculate_fixed_target}
\title{Calculate fixed target from a baseline year by a target year}
\usage{
calculate_fixed_target(
  target_value,
  small_is_best,
  year,
  baseline_year,
  target_year,
  baseline_value
)
}
\arguments{
\item{target_value}{vector of values to use as targets}

\item{small_is_best}{Logical to identify if a lower value is better than a higher
one (e.g. lower obesity in a positive public health outcome, so obesity rate
should have small_is_best = TRUE).}

\item{year}{Column name of column with years.}

\item{baseline_year}{Year from which the scenario is measured.
Defaults to \code{start_year}}

\item{target_year}{Year by which the scenario should eventually be
achieved. Defaults to \code{end_year}}

\item{baseline_value}{value at baseline_year}
}
\value{
Dataframe with additional scenario rows
}
\description{
The returned values are a portion of the straight line drawn from the
\code{baseline_year} value to the \code{target_year}. Only values for years between
\code{start_year} and \code{end_year} will be returned.
}
\details{
The \code{percent_change} parameter is understood as a percentage change,
and not a percentage point change, as this is usually what intended by those
formulations. If it is indeed the percentage change that is required, please
use \code{THIS_FUNCTION}. For instance, to calculate the scenario "reduce the 2018
value (90\%) by 30\% by 2025", will results to a 2025 value of 63\% and not 60\%.

The returned scenario is a portion of the straight line drawn from the
\code{baseline_year} value to the \code{target_year}. Only values for years between
\code{start_year} and \code{end_year} will be returned.
}
