% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hep_billion.R
\name{calculate_hep_billion}
\alias{calculate_hep_billion}
\title{Calculate HEP Billion}
\usage{
calculate_hep_billion(
  df,
  iso3 = "iso3",
  ind = "ind",
  year = "year",
  scenario = NULL,
  transform_value = "transform_value",
  level = stringr::str_replace(transform_value, "transform_value", "level"),
  contribution = stringr::str_replace(transform_value, "transform_value",
    "contribution"),
  start_year = 2018,
  end_year = 2019:2023,
  pop_year = 2023,
  ind_ids = billion_ind_codes("hep")
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{ind}{Column name of column with indicator names.}

\item{year}{Column name of column with years.}

\item{scenario}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{transform_value}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{level}{Column name of column(s) with indicator levels. Should be same length
as \code{transform_value}.}

\item{contribution}{Glue specification used to transform \code{transform_value} column
names into contribution column names. Defaults to replacing the string
\code{"transform_value"} with \code{"contribution"}.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2023.}

\item{pop_year}{Year used to pull in HPOP populations, defaults to 2023.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
Data frame in long format.
}
\description{
\code{calculate_hep_billion()} calculates country-level HEP Billion based on
the component indicators. It calculates the change for Prevent and Prepare between
a start year and end year, and estimated the change for Detect and Respond based on
its level in the end year. If data is not available for the end year for Detect and
Respond, the latest year of observed data is used. Details are available in
the methods report for the exact method applied.
}
\details{
For more details on the HEP Billion calculation process and how this function
ties in with the rest, see the vignette:

\href{../doc/hep.html}{\code{vignette("hep", package = "billionaiRe")}}
}
