% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hep.R
\name{calculate_hep_components}
\alias{calculate_hep_components}
\title{Calculate HEP component indicators}
\usage{
calculate_hep_components(
  df,
  iso3 = "iso3",
  year = "year",
  ind = "ind",
  transform_value = "transform_value",
  type = "type",
  level = "level",
  hepi_start_year = 2018,
  ind_ids = billion_ind_codes("hep")
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{year}{Column name of column with year data.}

\item{ind}{Column name of column with indicator names.}

\item{transform_value}{Column name of column with transformed values.}

\item{type}{Column name of column with type data.}

\item{level}{Column name to create to hold levels data.}

\item{hepi_start_year}{First year to calculate HEP index for.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
Data frame in long format.
}
\description{
Currently, the Prepare and DNR indicators are already calculated in the input
data, so only Prevent is calculated in this function. It takes the numerator
and denominator data from the inputs and calculates the vaccination coverage
per year and country for each relevant pathogen, as well as the overall Prevent
score. For DNR and its components, the level is calculated in this function.
}
