% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hep_components.R
\name{calculate_hepi}
\alias{calculate_hepi}
\title{Calculate HEPI}
\usage{
calculate_hepi(
  df,
  iso3,
  year,
  scenario,
  ind,
  type_col,
  source_col,
  source,
  transform_value,
  earliest_year,
  ind_ids
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{year}{Column name of column with year data.}

\item{scenario}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{ind}{Column name of column with indicator names.}

\item{type_col}{Column name of column with type data.}

\item{source_col}{Column name of column to store source information.}

\item{source}{Source to use if no unique source available for the calculation.}

\item{transform_value}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{earliest_year}{Earliest year for HEPI calculation.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\description{
Function to calculate HEPI as the average of DNR, Prepare, and Prevent.
Used within \code{calculate_hep_components()}.
}
