% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hpop_billion.R
\name{calculate_hpop_billion}
\alias{calculate_hpop_billion}
\title{Calculate HPOP Billion}
\usage{
calculate_hpop_billion(
  df,
  year = "year",
  start_year = 2018,
  end_year = 2019:2023,
  iso3 = "iso3",
  ind = "ind",
  population = "population",
  pop_year = 2023,
  transform_value = "transform_value",
  contribution = stringr::str_replace(transform_value, "transform_value",
    "contribution"),
  contribution_pct = paste0(contribution, "_percent"),
  scenario = NULL,
  ind_ids = billion_ind_codes("hpop")
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{year}{Column name of column with years.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2023.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{ind}{Column name of column with indicator names.}

\item{population}{Column name of column with population figures.}

\item{pop_year}{Year used to pull in HPOP populations, defaults to 2023.}

\item{transform_value}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{contribution}{Column name of column(s) to store contribution (population)
values. Must be the same length as \code{transform_value}.}

\item{contribution_pct}{Column name of column(s) to store contribution (percent)
values. Must be the same length as \code{transform_value}.}

\item{scenario}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
Data frame in long format.
}
\description{
\code{calculate_hpop_billion()} calculates country-level HPOP Billion based on
indicator level changes.
}
\details{
For more details on the HPOP Billion calculation process and how this function
ties in with the rest, see the vignette:

\href{../doc/hpop.html}{\code{vignette("hpop", package = "billionaiRe")}}
}
