% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hpop_billion.R
\name{calculate_hpop_change_vector}
\alias{calculate_hpop_change_vector}
\title{Calculate the change for vectors, used in \code{\link[=calculate_hpop_billion]{calculate_hpop_billion()}}}
\usage{
calculate_hpop_change_vector(transform_value_col, year, start_year)
}
\arguments{
\item{transform_value_col}{Vector of transform values}

\item{year}{Vector of years}

\item{start_year}{Start year}
}
\description{
Calculate the change for vectors, used in \code{\link[=calculate_hpop_billion]{calculate_hpop_billion()}}
}
