% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hpop_contributions.R
\name{calculate_hpop_contribution_vector}
\alias{calculate_hpop_contribution_vector}
\title{Calculate HPOP contribution in a vector}
\usage{
calculate_hpop_contribution_vector(
  contribution,
  ind,
  ind_ids,
  year,
  start_year,
  end_year,
  transform_value,
  pop
)
}
\arguments{
\item{contribution}{Contribution vector}

\item{ind}{Ind vector}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{year}{Year vector}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2023.}

\item{transform_value}{Transformed value vector}

\item{pop}{Population vector}
}
\description{
Done this to avoid using \code{\link[dplyr:case_when]{dplyr::case_when()}}  because that evaluates all
RHS arguments, which produces errors for indicators without sufficient data.
}
