% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpop.R
\name{calculate_hpop_ind_contributions}
\alias{calculate_hpop_ind_contributions}
\title{Calculate HPOP Indicator Contributions}
\usage{
calculate_hpop_ind_contributions(
  df,
  year = "year",
  start_year = 2018,
  end_year = 2023,
  iso3 = "iso3",
  ind = "ind",
  population = "population",
  transform_value = "transform_value"
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year for contribution calculation, defaults to 2023.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{ind}{Column name of column with indicator names.}

\item{population}{Column name of column with population figures.}

\item{transform_value}{Column name of column with transformed indicator values.}
}
\value{
Data frame in long format.
}
\description{
\code{calculate_hpop_contributions()} calculates indicator-level contributions and
changes for the HPOP Billion.
}
\details{
For more details on the HPOP Billion calculation process and how this function
ties in with the rest, see the vignette:

\href{../doc/hpop.html}{\code{vignette("hpop", package = "billionaiRe")}}
}
