% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios_percent_change.R
\name{calculate_percent_change_baseline}
\alias{calculate_percent_change_baseline}
\title{Calculate percent change}
\usage{
calculate_percent_change_baseline(
  baseline_value,
  goal_value,
  year,
  baseline_year,
  target_year
)
}
\arguments{
\item{baseline_value}{vector with the baseline value to be used}

\item{goal_value}{vector with the goal value to be used}

\item{year}{Column name of column with years.}

\item{baseline_year}{Year from which the scenario is measured.
Defaults to \code{start_year}}

\item{target_year}{Year by which the scenario should eventually be
achieved. Defaults to \code{end_year}}
}
\description{
Calculate percent change
}
