% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_uhc_billion.R
\name{calculate_uhc_billion}
\alias{calculate_uhc_billion}
\title{Calculate UHC Billion}
\usage{
calculate_uhc_billion(
  df,
  transform_value_col = "transform_value",
  value_col = "value",
  scenario_col = NULL,
  source = sprintf("WHO DDI calculation, \%s", format(Sys.Date(), "\%B \%Y")),
  projected_year = lubridate::year(lubridate::now()),
  ind_ids = billion_ind_codes("uhc")
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{transform_value_col}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{value_col}{Column name of column with indicator values.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{source}{Source to provide for calculated average service coverage and
single measure.}

\item{projected_year}{Year that calculated data will be regarded as projected.
Years prior reported as estimated.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
Data frame in long format.
}
\description{
\code{calculate_uhc_billion()} calculates country-level UHC Billion based on
indicator level data. Calculates it for each country-year combination in the provided data.
In order to calculate average service coverage, and thus the overall Billion,
all UHC indicators must be present for a given year. The only exception is
insecticide treated nets, which need not be present.
}
\details{
For more details on the UHC Billion calculation process and how this function
ties in with the rest, see the vignette:

\href{../doc/uhc.html}{\code{vignette("uhc", package = "billionaiRe")}}
}
