% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_uhc_contribution.R
\name{calculate_uhc_contribution}
\alias{calculate_uhc_contribution}
\title{Calculate UHC contribution}
\usage{
calculate_uhc_contribution(
  df,
  year = "year",
  iso3 = "iso3",
  ind = "ind",
  value = "value",
  start_year = 2018,
  end_year = 2019:2025,
  pop_year = 2025,
  transform_value = "transform_value",
  contribution = stringr::str_replace(transform_value, "transform_value",
    "contribution"),
  contribution_pct = paste0(contribution, "_percent"),
  scenario = NULL,
  scenario_reported_estimated = "routine",
  scenario_covid_shock = "covid_shock",
  scenario_reference_infilling = "reference_infilling",
  ind_ids = billion_ind_codes("uhc", include_calculated = TRUE)
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{year}{Column name of column with years.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{ind}{Column name of column with indicator names.}

\item{value}{Column name of column with indicator values.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{pop_year}{Year used to pull in HPOP populations, defaults to 2025.}

\item{transform_value}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{contribution}{Column name of column(s) to store contribution (population)
values. Must be the same length as \code{transform_value}.}

\item{contribution_pct}{Column name of column(s) to store contribution (percent)
values. Must be the same length as \code{transform_value}.}

\item{scenario}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{scenario_reported_estimated}{name of the reported/estimated scenario.}

\item{scenario_covid_shock}{name of the scenario with the COVID-19 shock years.}

\item{scenario_reference_infilling}{name of the WHO technical programs projections/imputations scenario.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
Data frame in long format.
}
\description{
\code{calculate_uhc_contribution()} calculates country-level UHC contributions, in
total population and percents, based on indicator level data. Calculates it
for each country-year combination in the provided data, and for specific
scenarios if specified.
}
\details{
For more details on the UHC Billion calculation process and how this function
ties in with the rest, see the vignette:

\href{../doc/uhc.html}{\code{vignette("uhc", package = "billionaiRe")}}
}
