% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_styles_modify.R
\name{col2hex}
\alias{col2hex}
\title{Converts rgb colour to hex colour}
\usage{
col2hex(colour)
}
\arguments{
\item{colour}{colour as in \code{colours()}}
}
\description{
Converts rgb colour to hex colour
}
\author{
Philipp Schauberger
}
