% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{convert_ind_codes}
\alias{convert_ind_codes}
\title{Convert indicator codes between types}
\usage{
convert_ind_codes(
  ind_codes,
  from = c("dashboard_id", "ind", "gho_code"),
  to = c("dashboard_id", "ind", "gho_code")
)
}
\arguments{
\item{ind_codes}{Character vector of indicator codes}

\item{from}{Type of code being passed to \code{convert_ind_codes()}.}

\item{to}{Type of code to return.}
}
\value{
A character vector.
}
\description{
\code{convert_ind_codes()} allows easy conversion of indicator codes from dashboard
IDs, xMart IDs, and GHO codes.
}
