% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{convert_wash_name}
\alias{convert_wash_name}
\title{Convert WASH names to remove the urban/rural element}
\usage{
convert_wash_name(ind_list)
}
\arguments{
\item{ind_list}{character vector with the indicators to be changed}
}
\description{
Convert WASH names to remove the urban/rural element
}
