% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_shares}
\alias{country_shares}
\title{Country shares data}
\format{
A data frame with 582 rows and 4 variables:
\describe{
\item{iso3}{Country ISO3 codes.}
\item{billion}{Relevant billion}
\item{share_n}{Share, in number of people.}
\item{share_perc}{Share, as percent of total projected population in 2023.}
}
}
\usage{
country_shares
}
\description{
Country shares data for UHC and HPOP Billions for all 194 WHO member
states.
}
\keyword{datasets}
