% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_styles_modify.R
\name{excel_hex_colour}
\alias{excel_hex_colour}
\title{Converts colour to Excel hex colour}
\usage{
excel_hex_colour(colour, errorMsg = "Invalid colour")
}
\arguments{
\item{colour}{colour as in \code{colours()}}

\item{errorMsg}{error message to be passed.}
}
\description{
Converts colour to Excel hex colour
}
