% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_styles.R
\name{excel_styles}
\alias{excel_styles}
\title{Create basic Excel Styles and provides a way to modify them}
\usage{
excel_styles(
  billion = NULL,
  style_category = NULL,
  type_data = NULL,
  fontName = "Calibri",
  billion_fgFill = NULL,
  hide = FALSE,
  ...
)
}
\arguments{
\item{billion}{Billion to be used for billion styling when relevant: either
"hep", "hpop", "uhc", or NULL when no billion to be applied.}

\item{style_category}{category of style to be used. Can be either:
\itemize{
\item "title"
\item "subtitle"
\item "datatable_header": for column headers of data tables
\item "sub_datatable_header": for sub-headers of data tables
\item "data" : for standard data points
\item "normal_text"
\item NULL when no category applies
}}

\item{type_data}{type of data. Can be either "date", "numeric", or "integer".}

\item{fontName}{character name with the name of the font to use.}

\item{billion_fgFill}{character with the type of billion color to be used. Passed
to \code{\link[openxlsx:createStyle]{openxlsx::createStyle()}} fgFill parameter.}

\item{hide}{bolean. Set fontColor to white and hidden to TRUE in \code{\link[openxlsx:createStyle]{openxlsx::createStyle()}}}

\item{...}{additional parameters to be passed to \code{\link[openxlsx:createStyle]{openxlsx::createStyle()}}.
Overrides any of the parameter passed by previous parameters.}
}
\description{
\code{excel_styles} creates a series of styles used across the package and provides
a way to modify them to increase the performance of the styling functions.
}
\seealso{
\code{\link[openxlsx]{createStyle}}
}
