% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scenarios.R
\name{exec_scenario}
\alias{exec_scenario}
\title{Execute a scenario}
\usage{
exec_scenario(df, fn, parameters)
}
\arguments{
\item{df}{(data.frame) containing the data.}

\item{fn}{function to be executed}

\item{parameters}{parameters to be passed to \code{fn}}
}
\description{
Simple wrapper around \link[rlang:exec]{rlang::exec} to pass \code{parameters} and \code{df} to \code{fn}
}
