% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_country_summary.R
\name{export_all_country_summary_xls}
\alias{export_all_country_summary_xls}
\title{Export country summary to Excel for all billions.
\code{export_all_country_summary_xls} Export a country-specific for all billions.}
\usage{
export_all_country_summary_xls(df, iso3)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{iso3}{ISO3 codes of country to export.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_fldr}.
}
\description{
Export country summary to Excel for all billions.
\code{export_all_country_summary_xls} Export a country-specific for all billions.
}
