% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_country_summary.R
\name{export_country_summary_xls}
\alias{export_country_summary_xls}
\title{Export country summary to Excel}
\usage{
export_country_summary_xls(
  df,
  iso,
  start_year = 2018,
  end_year = 2019:2023,
  billion = c("hep", "hpop", "uhc", "all"),
  output_fldr = "outputs",
  xls_template = "data-raw/CountrySummary_template.xlsx",
  ...
)
}
\arguments{
\item{billion}{Billion to export to return, either "hep", "hpop", "uhc" or "all".}

\item{output_fldr}{Folder path to where the Excel files should be written}

\item{xls_template}{Path to the Excel template to be used.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_fldr}.
}
\description{
\code{export_country_summary_xls} Export a country-specific for all three
billions or for a specific billion.
}
\examples{
\dontrun{
data <- load_billion_data("hpop")
export_country_summary_xls(data, iso = "AFG")
}
}
