% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_country_summary.R
\name{export_hep_country_summary_xls}
\alias{export_hep_country_summary_xls}
\title{Export country summary to Excel for HEP billion$}
\usage{
export_hep_country_summary_xls(
  df,
  iso,
  start_year = 2018,
  end_year = 2019:2023,
  ...
)
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_fldr}.
}
\description{
\code{export_hep_country_summary_xls} Export a country-specific for HEP billion.
}
