% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_country_summary.R
\name{export_hpop_country_summary_xls}
\alias{export_hpop_country_summary_xls}
\title{Export country summary to Excel for HPOP billion}
\usage{
export_hpop_country_summary_xls(
  df,
  iso,
  iso3 = "iso3",
  year = "year",
  ind = "ind",
  value = "value",
  transform_value = "transform_value",
  scenario = NULL,
  type_col = "type",
  source_col = "source",
  population = "population",
  contribution = "contribution",
  start_year = 2018,
  end_year = 2019:2023,
  sheet_prefix = "HPOP",
  output_folder = "outputs",
  ind_ids = billion_ind_codes("hpop"),
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{iso}{ISO3 code of country to summarize.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{ind}{Column name of column with indicator names.}

\item{value}{Column name of column with indicator values.}

\item{sheet_prefix}{Character prefix to add in front of export sheets}

\item{output_folder}{Folder path to where the Excel files should be written}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{export_hpop_country_summary_xls} Export a country-specific for HPOP billion.
}
