% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_country_summary.R
\name{export_uhc_country_summary_xls}
\alias{export_uhc_country_summary_xls}
\title{Export country summary to Excel for UHC billion
\code{export_uhc_country_summary_xls} Export a country-specific for UHC billion.}
\usage{
export_uhc_country_summary_xls(
  df,
  iso,
  start_year = 2018,
  end_year = 2019:2023,
  output_folder = "outputs",
  xls_template = "data-raw/CountrySummary_template.xlsx"
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{iso}{ISO3 code of country to summarize.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2023.}

\item{output_folder}{Folder path to where the Excel files should be written}

\item{xls_template}{Path to the Excel template to be used.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
Export country summary to Excel for UHC billion
\code{export_uhc_country_summary_xls} Export a country-specific for UHC billion.
}
