% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate_campaign_data.R
\name{extrapolate_campaign_data}
\alias{extrapolate_campaign_data}
\title{Extrapolate campaign data for the transform value, type, and source}
\usage{
extrapolate_campaign_data(
  pathogen,
  pathogen_year,
  df,
  transform_value_col,
  source
)
}
\arguments{
\item{pathogen}{Indicator IDs of the pathogens}

\item{pathogen_year}{Latest year of data to use for flat extrapolation (extrapolating after that year)}

\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{transform_value_col}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{source}{Source to use if no unique source available for the calculation.}
}
\description{
Works across multiple \code{transform_value_col} columns.
}
