% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate_campaign_data.R
\name{extrapolate_campaign_vector}
\alias{extrapolate_campaign_vector}
\title{Rolling sum and extrapolate campaigns data}
\usage{
extrapolate_campaign_vector(x, n)
}
\arguments{
\item{x}{Vector of campaigns data, either numerator or denominator}

\item{n}{Years to take rolling average for}
}
\description{
Takes rolling sum over observed data and then flat extrapolates from last reported
}
