% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hpop_billion.R
\name{generate_hpop_populations}
\alias{generate_hpop_populations}
\title{Generate HPOP Population Table}
\usage{
generate_hpop_populations(pop_year)
}
\arguments{
\item{pop_year}{Year used to pull in HPOP populations, defaults to 2023.}
}
\value{
Data frame in long format.
}
\description{
\code{generate_hpop_populations()} creates the country/population data frame necessary
for correcting double counting within the HPOP Billion calculation. This is used
within \code{calculate_hpop_billion()}.
}
\details{
Because the underlying data from the wppdistro package is not publicly distributed
for specific small WHO member states, this is not exported as an already created
(and thus publicly available) table within the billionaiRe package, but instead
dynamically generated each R session.
}
