% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{get_col_width}
\alias{get_col_width}
\title{Set column width based on type of column in data frame for openxlsx export}
\usage{
get_col_width(
  df,
  value,
  transform_value,
  type_col,
  source_col,
  start_year,
  end_year
)
}
\arguments{
\item{df}{data frame to be written}

\item{value}{Column name of column with indicator values.}

\item{transform_value}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{type_col}{Column name of column to store type.}

\item{source_col}{Column name of column to store source information.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2023.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_fldr}.
}
\description{
\code{transform_hpop_data()} applies transformations on HPOP Billion indicators so
that transformed indicator values can be used within Billions calculations.
Details on the specific transformations applied can be found within the
Billions methods report. Values in the transform column, if it already exists,
are replaced for HPOP indicators that have data in the value column, otherwise
the column keeps its original data.
}
\details{
For more details on the HPOP Billion calculation process and how this function
ties in with the rest, see the vignette:

\href{../doc/hpop.html}{\code{vignette("hpop", package = "billionaiRe")}}
}
