% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{get_country_shares}
\alias{get_country_shares}
\title{Get country shares data}
\usage{
get_country_shares(
  iso3,
  billion = c("hep", "hpop", "uhc"),
  share_type = c("n", "perc")
)
}
\arguments{
\item{iso3}{Country ISO3 codes.}

\item{billion}{Relevant Billion, one of "hep", "hpop", or "uhc".}

\item{share_type}{Type of share to return, either in total numbers of people ("n"),
or as a percent of total projected population in 2023 ("perc").}
}
\description{
Matches country ISO3 codes to data in the \code{country_shares} data frame.
}
\seealso{
Functions to access metadata
\code{\link{billion_ind_codes}()},
\code{\link{convert_ind_codes}()},
\code{\link{get_ind_billion}()},
\code{\link{get_ind_metadata}()},
\code{\link{get_sdi_ratio}()},
\code{\link{indicator_df}}
}
\concept{metadata}
