% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_whdh.R
\name{get_data_lake_name}
\alias{get_data_lake_name}
\title{Return the name of the 3B data lake}
\usage{
get_data_lake_name()
}
\value{
A string. The name of the Triple Billions WHDH data lake.
}
\description{
Return the name of the 3B data lake
}
