% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_data.R
\name{get_data_type}
\alias{get_data_type}
\title{Gets data type of data frame for styling}
\usage{
get_data_type(df)
}
\arguments{
\item{df}{data frame to be tested}
}
\value{
character vector with the data_type of data frame columns
}
\description{
Gets the type of data for styling purposes. Can be either:
\itemize{
\item numeric
\item integer
\item Date
\item character
\item character_centered
Used to pass appropriate parameters to \code{style_data()}
}
}
