% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_data.R
\name{get_data_type_single}
\alias{get_data_type_single}
\title{Gets data type for styling}
\usage{
get_data_type_single(vec)
}
\arguments{
\item{vec}{vector to be tested}
}
\description{
Gets the type of data for styling purposes. Can be either:
\itemize{
\item numeric
\item integer
\item Date
\item character
\item character_centered
\item formula
Used to pass appropriate parameters to \code{style_data()}
}
}
