% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{get_ind_metadata}
\alias{get_ind_metadata}
\title{Get metadata for a Triple Billions indicator(s)}
\usage{
get_ind_metadata(ind_codes, metadata_col = names(billionaiRe::indicator_df))
}
\arguments{
\item{ind_codes}{A character vector with indicator (analysis) codes}

\item{metadata_col}{The name of the indicator_df column with the desired metadata.}
}
\value{
A character vector with the metadata. The positions correspond to the
order of the ind_codes input.
}
\description{
This is a convenience function that looks up the relevant metadata for one or
more indicators in the indicator_df table.
}
\examples{
# Find the gho_code for UHC Tobacco
get_ind_metadata("uhc_tobacco", "gho_code")

# Find the indicator type for multiple indicators
get_ind_metadata(c("alcohol", "hwf"), "ind_type")
}
\seealso{
Functions to access metadata
\code{\link{billion_ind_codes}()},
\code{\link{convert_ind_codes}()},
\code{\link{get_country_shares}()},
\code{\link{get_ind_billion}()},
\code{\link{get_sdi_ratio}()},
\code{\link{indicator_df}}
}
\concept{metadata}
