% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-summary.R
\name{get_ind_order}
\alias{get_ind_order}
\title{Get order of indicator}
\usage{
get_ind_order(ind)
}
\arguments{
\item{ind}{character vector of indicators}
}
\value{
character vector
}
\description{
Get order of indicator
}
