% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scenarios.R
\name{get_last_value}
\alias{get_last_value}
\title{Get last value for the filter}
\usage{
get_last_value(df, type_filter = c("reported", "estimated"))
}
\arguments{
\item{df}{(data.frame) containing the data. Needs to have at least the \code{year} and \code{type} columns.}

\item{type_filter}{(character vector) vector of types to be filtered for.}
}
\description{
Get last value for the filter
}
