% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scenarios.R
\name{get_last_year_scenario}
\alias{get_last_year_scenario}
\title{Get last value for the filter in default scenario}
\usage{
get_last_year_scenario(
  df,
  indicator,
  start_year = 2018,
  scenario = "default",
  scenario_col = "scenario",
  type_filter = c("reported", "estimated", "projected", "estimated")
)
}
\arguments{
\item{df}{(data.frame) containing the data. Needs to have at least the \code{year}, \code{scenario_col} and \code{type} columns.}

\item{indicator}{(character) identifying the indicator for which to get the last year of \code{type_filter} data.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{scenario}{name of scenario to recycle for.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{type_filter}{(character vector) vector of types to be filtered for.}
}
\description{
Get last value for the filter in default scenario
}
\seealso{
General scenario functions
\code{\link{add_scenario}()},
\code{\link{calculate_aarc}()},
\code{\link{calculate_aroc}()},
\code{\link{exec_scenario}()},
\code{\link{fill_cols_scenario}()},
\code{\link{flat_extrapolation}()},
\code{\link{get_aarr}()},
\code{\link{get_baseline_value}()},
\code{\link{get_baseline_year}()},
\code{\link{get_last_value}()},
\code{\link{get_latest_aarc}()},
\code{\link{get_percent_change_aarc}()},
\code{\link{get_target_aarc}()},
\code{\link{remove_unwanted_scenarios}()},
\code{\link{trim_values}()},
\code{\link{trim_years}()}
}
\concept{scenarios}
\keyword{internal}
