% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scenarios.R
\name{get_latest_aarc}
\alias{get_latest_aarc}
\title{Get the latest AARC for data frame}
\usage{
get_latest_aarc(
  df,
  baseline_year,
  value = "value",
  year = "year",
  iso3 = "iso3",
  ind = "ind"
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{baseline_year}{Year from which the scenario is measured.
Defaults to \code{start_year}}

\item{value}{Column name of column with indicator values.}

\item{year}{Column name of column with years.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{ind}{Column name of column with indicator names.}
}
\value{
dataframe with AROC (in \code{aroc} column) for every combination of
\code{iso3} and \code{ind}
}
\description{
Gets the annual average rate of change (AARC or AROC) between the
baseline year and the year immediately preceding for every combination of
\code{iso3} and \code{ind} present in \code{df}.
}
