% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-summary.R
\name{get_latest_reported_df}
\alias{get_latest_reported_df}
\title{Get latest reported data frame}
\usage{
get_latest_reported_df(
  df,
  iso3,
  ind,
  type_col,
  year,
  value,
  transform_value = NULL,
  source_col,
  ind_df
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{ind}{Column name of column with indicator names.}

\item{type_col}{Column name of column to store type.}

\item{year}{Column name of column with years.}

\item{value}{Column name of column with indicator values.}

\item{transform_value}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{source_col}{Column name of column to store source information.}

\item{ind_df}{data frame containing information on indicators}
}
\description{
\code{get_latest_reported()} gets the latest reported data available for estimated
or reported data. Used in write functions.
}
