% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hep.R
\name{get_latest_year}
\alias{get_latest_year}
\title{Get latest years for pathogens}
\usage{
get_latest_year(
  df,
  ind,
  year,
  cholera_latest_year,
  meningitis_latest_year,
  yellow_fever_latest_year
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind}{Column name of column with indicator names.}

\item{year}{Column name of column with year data.}

\item{cholera_latest_year}{Latest year to calculate rolling average for cholera.
If \code{NULL}, uses the latest year with observations in the data.}

\item{meningitis_latest_year}{Latest year to calculate rolling average for meningitis
If \code{NULL}, uses the latest year with observations in the data.}

\item{yellow_fever_latest_year}{Latest year to calculate rolling average for meningitis
If \code{NULL}, uses the latest year with observations in the data.}
}
\description{
Since campaign data is not projected, we only want to do rolling sums on the
variables until the latest year that data is available. A user can either
provide a latest year for each pathogen, or it will be determined by the available
data. This function processes that for all three pathogens to get the latest
available year.
}
