% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{get_sdi_ratio}
\alias{get_sdi_ratio}
\title{Get SDI ratio data}
\usage{
get_sdi_ratio(iso3)
}
\arguments{
\item{iso3}{Country ISO3 codes.}
}
\value{
Numeric vector of SDI ratios.
}
\description{
Matches country ISO3 codes to data in the \code{sdi_ratio} data frame.
}
\seealso{
Functions to access metadata
\code{\link{billion_ind_codes}()},
\code{\link{convert_ind_codes}()},
\code{\link{get_country_shares}()},
\code{\link{get_ind_billion}()},
\code{\link{get_ind_metadata}()},
\code{\link{indicator_df}}
}
\concept{metadata}
