% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{get_small_is_best}
\alias{get_small_is_best}
\title{Gets the direction of indicator}
\usage{
get_small_is_best(ind_codes)
}
\arguments{
\item{ind_codes}{Character vector of indicator codes}
}
\value{
A logical vector.
}
\description{
\code{get_small_is_best()} pulls from \code{billionaiRe::indicator_df} the value stored
in the \code{small_is_best} column. If \code{small_is_best} is TRUE, a lower value means
a better public health outcome for this indicator. For instance, lower
child obesity prevalence is a positive public health outcome, so \code{small_is_best}
is TRUE for \code{child_obese}.
}
