% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_excel_formulas.R
\name{get_transform_formula_single}
\alias{get_transform_formula_single}
\title{Get the transformation formula for a specified indicator and cell reference}
\usage{
get_transform_formula_single(ind, raw_col, raw_row, ind_ids, iso3)
}
\arguments{
\item{ind}{character with the indicator code (\code{ind} in \code{billionaiRe::indicator_df()})}

\item{raw_col}{integer identifying the column where the raw value (pre-transformation)
is located.}

\item{raw_row}{integer identifying the row where the raw value (pre-transformation)
is located.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{iso3}{Country ISO3 code.}
}
\value{
character vector with the formula for the indicator, column, row combination
specified. If indicator is not found, "" is returned.
}
\description{
\code{get_transform_formula_single()} gets the right transformation Excel formula
for the specified indicator at cell reference. Used in \code{get_transform_formula_single}.
}
