% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_whdh.R
\name{get_whdh_path}
\alias{get_whdh_path}
\title{Generate file paths for the Triple Billions WHDH data lake}
\usage{
get_whdh_path(
  operation = c("download", "upload"),
  data_type = c("wrangled_data", "projected_data", "final_data", "ingestion_data"),
  billion = c("all", "hep", "hpop", "uhc"),
  ind_codes = "all",
  file_names = NULL,
  experiment = "unofficial"
)
}
\arguments{
\item{operation}{(string) Either "download" or "upload".}

\item{data_type}{(string) The type of data to load.
\itemize{
\item \code{wrangled_data} (default): raw data that has been wrangled into a suitable
form for analysis.
\item \code{projected_data}:  data that has been fully projected to the target year but
has not yet been transformed or calculated upon.
\item \code{final_data}: the complete set of billions data with transformed values, contributions,
and all calculations available.
\item \code{ingestion_data}: raw data in its original form as received from the technical
program, GHO, or other sources. These files have not been wrangled or modified
in any way.
}}

\item{billion}{(string) One of "all" (default), "hep", "hpop", or "uhc". If "all",
the file paths for all indicators in all three bilions are returned.}

\item{ind_codes}{(character vector) The name of the indicator (or indicators) to load data for.
If \code{all}, returns paths for all indicators for a given billion. If \code{billion = "all"},
this argument is ignored and the file paths for all indicators in all three bilions
are returned.}

\item{file_names}{(character vector) The name(s) of the file(s) to download.
NULL by default. Ignored if either \code{billion = "all"} or \code{ind_codes = "all"}.}

\item{experiment}{(string) Either \code{NULL} or a string ("unofficial" by default).
\itemize{
\item If \code{NULL}, the root folder for the data layers is the 3B folder (i.e., where
the "official" data is stored (e.g., \verb{3B/...}).
\item If a string, the root folder for the data layers is a sub-folder within the
Sandbox layer of the 3B data lake (e.g., if \code{experiment = "my_exp"}, then
paths would be of the form \verb{3B/Sandbox/my_exp/Silver/...})
}}
}
\value{
A character vector.
}
\description{
\code{get_whdh_path()} simplifies the process of generating accurate file paths for
downloading/uploading files from the Triple Billions WHDH data lake.
}
\details{
Using this function when working with the data lake is highly recommended as
it ensures file paths abide by the established standards and directory structure
for the data lake.
}
\seealso{
Wrangle data functions
\code{\link{add_missing_xmart_rows}()},
\code{\link{get_data_lake_name}()},
\code{\link{has_xmart_cols}()},
\code{\link{save_gho_backup_to_whdh}()},
\code{\link{save_wrangled_output}()},
\code{\link{wrangle_gho_data}()},
\code{\link{wrangle_gho_rural_urban_data}()},
\code{\link{wrangle_unsd_data}()},
\code{\link{xmart_col_types}()},
\code{\link{xmart_cols}()}
}
\concept{whdh}
