% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hep_df}
\alias{hep_df}
\title{HEP generated example data}
\format{
A data frame with 21 rows and 5 variables:
\describe{
\item{iso3}{Country ISO3 codes.}
\item{year}{Year.}
\item{ind}{HPOP indicator code.}
\item{value}{Raw indicator value.}
\item{type}{Data type.}
}
}
\usage{
hep_df
}
\description{
Generated (fake) HEP data used to test the Billions calculations code within the billionaiRe
package.
}
\details{
See the HEP vignette for its example use:

\href{../doc/hep.html}{\code{vignette("hep", package = "billionaiRe")}}
}
\keyword{datasets}
