% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_covid_shock.R
\name{impute_covid_shock}
\alias{impute_covid_shock}
\title{Impute a COVID-19 shock}
\usage{
impute_covid_shock(
  df,
  region = c("who_region", "un_region", "un_subregion", "un_intermediate_region",
    "sdg_region", "sdg_subregion", "gbd_region", "gbd_subregion", "un_desa_region",
    "un_desa_subregion", "wb_region", "wb_ig"),
  start_year = 2018,
  covid_year = 2020,
  recovery_year = 2022,
  end_year = 2025,
  value_col = "value",
  scenario_col = "scenario",
  default_scenario = "default",
  scenario_reported_estimated = "routine",
  scenario_covid_shock = "covid_shock",
  scenario_reference_infilling = "reference_infilling",
  source = sprintf("WHO DDI interim infilling and projections, \%s", format(Sys.Date(),
    "\%B \%Y"))
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{region}{(character) Type of region to use for regional average. See \link[whoville]{iso3_to_regions} for more details.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{covid_year}{(integer) year where the values are impacted by COVID.}

\item{recovery_year}{integer year from which the AROC will be applied. Default to 2022.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{value_col}{Column name of column with indicator values.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{default_scenario}{name of the default scenario.}

\item{scenario_reported_estimated}{name of the reported/estimated scenario.}

\item{scenario_covid_shock}{name of the scenario with the COVID-19 shock years.}

\item{scenario_reference_infilling}{name of the WHO technical programs projections/imputations scenario.}

\item{source}{Source to provide for calculated average service coverage and
single measure.}
}
\description{
Impute a COVID-19 shock for each \code{iso3} and \code{ind} combination by the imputation flow chart:
\itemize{
\item if there is a 2020 value, then 2021 = 2020
\item If no 2020 values, but more than 30\% of countries in \code{region} have one:
\itemize{
\item Use mean regional chock (difference between 2020 and 2019) to infer a 2020 value
}
}
}
