% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicator_order}
\alias{indicator_order}
\title{Dataset of indicators order and meta-information}
\format{
A data frame with 61 rows and 9 variables:
\describe{
\item{sdg}{Indicator code from SDG and attributed if not in SDG}
\item{short_name}{Short name of the indicator}
\item{unit_raw}{Unit of raw values for indicator}
\item{unit_transformed}{Unit of raw values for indicator}
\item{medium_name}{Medium length name of indicator (typically used for tables)}
\item{transformed_name}{Name of indicator after transformation}
\item{ind}{Code of the indicator in billionaiRe}
\item{order}{Order of the indicators}
\item{pillar}{Pillar in which the indicators sits for UHC and HEP (e.g. Prevent, Non communicable diseases (NCDs), etc.)}
}
}
\usage{
indicator_order
}
\description{
A dataset containing dashboard and analysis indicator codes and their uses
within the GPW13 Billions. It is joinded to \code{indicator_df}.
}
\keyword{datasets}
