% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_billion_data}
\alias{load_billion_data}
\title{Load Billions indicator data}
\usage{
load_billion_data(
  data_type = c("wrangled_data", "projected_data", "final_data"),
  billion = c("all", "hep", "hpop", "uhc"),
  ind_codes = "all",
  date_filter = "latest",
  na_rm = TRUE,
  experiment = NULL,
  silent = TRUE,
  data_source = c("whdh", "xmart"),
  ...
)
}
\arguments{
\item{data_type}{(string) The type of data to load.
\itemize{
\item \code{wrangled_data} (default): raw data that has been wrangled into a suitable
form for analysis.
\item \code{projected_data}:  data that has been fully projected to the target year but
has not yet been transformed or calculated upon.
\item \code{final_data}: the complete set of billions data with transformed values, contributions,
and all calculations available.
}}

\item{billion}{(string) One of "all" (default), "hep", "hpop", or "uhc". If "all",
downloads all indicators for all three billions.}

\item{ind_codes}{(character vector) The name of the indicator (or indicators) to load data for.
If \code{all}, downloads data for all indicators for a given billion. Ignored if
billion = "all".}

\item{date_filter}{(string) Either "latest", or a single date string. The date
string needs to be in ISO6801 format, such as "1989-4-4" or "1988-06-21".
If a date is provided, the returned data frame is a snapshot of the data as
stored on that date (if no updates were made on the given date, then the most
recent update before that date is used).}

\item{na_rm}{(logical) Specifies whether to filter the data to only rows
where \code{value} is not missing. Defaults to \code{TRUE}.}

\item{experiment}{(string) Either NULL (default) or a string.
\itemize{
\item If NULL, the path returned is in the "official" data layers in the 3B data lake
(e.g., 3B/Silver).
\item If a string, the path returned is in a sub-folder within the Sandbox layer of the
3B data lake. The Bronze/Silver/Gold structure is replicated within this sub-folder
(e.g., 3B/Sandbox/my_exp/Silver)
}}

\item{silent}{(logical) Specifies whether to show authentication messages and
a download progress bar. Defaults to \code{TRUE}.}

\item{data_source}{(string) Ether "whdh" or "xmart". Indicates where to download
the data from}

\item{...}{Additional arguments passed to \code{xmart4::xmart4_table()}. Use if
you need to provide additional token specifications for Azure authentication.}
}
\value{
A data frame.
}
\description{
\code{load_billion_data()} provides access to the raw inputs to the Billions stored
within the World Health Data Hub or the World Health Organization's xMart4 database.
By default, it filters the loaded data to take the most recently uploaded data
for each indicator, country, and year. The user can also specify to take the
latest data before a certain time period, or to extract all data from the database.
}
\details{
This function requires that the user have the \href{https://github.com/gpw13/xmart4}{xmart4}
and \href{https://github.com/WorldHealthOrganization/whdh}{whdh} packages installed
and setup to access the data. For quetions about getting the relevant permissions,
please contact \href{mailto:kanjim@who.int}{kanjim@who.int} or \href{mailto:messeillere@who.int}{messeillere@who.int}.
}
