% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_billion_data_xmart}
\alias{load_billion_data_xmart}
\title{Load Billions indicator data from xMart}
\usage{
load_billion_data_xmart(
  data_type = c("wrangled_data", "projected_data", "final_data"),
  billion = c("all", "hep", "hpop", "uhc"),
  ind_codes = "all",
  version = "latest",
  na_rm = TRUE,
  experiment = NULL,
  silent = TRUE
)
}
\arguments{
\item{data_type}{(string) The type of data to load.
\itemize{
\item \code{wrangled_data} (default): raw data that has been wrangled into a suitable
form for analysis.
\item \code{projected_data}:  data that has been fully projected to the target year but
has not yet been transformed or calculated upon.
\item \code{final_data}: the complete set of billions data with transformed values, contributions,
and all calculations available.
}}

\item{billion}{(string) One of "all" (default), "hep", "hpop", or "uhc". If "all",
downloads all indicators for all three billions.}

\item{ind_codes}{(character vector) The name of the indicator (or indicators) to load data for.
If \code{all}, downloads data for all indicators for a given billion. Ignored if
billion = "all".}

\item{version}{(string) Either \code{latest}  or a \code{yyyy-MM-dd} or \code{yyyy-mm-ddTHH-MM-SS}
formatted string.
\itemize{
\item If \code{latest}, the latest version of the data will be downloaded.
\item If a \code{yyyy-MM-dd} formatted string, the latest version of the data on or
before the provided date will be downloaded.
\item If a \code{yyyy-mm-ddTHH-MM-SS} formatted string, an exact match for the given
time stamp is sought, if it exists; otherwise, raises an error.
}}

\item{na_rm}{(logical) Specifies whether to filter the data to only rows
where \code{value} is not missing. Defaults to \code{TRUE}.}

\item{experiment}{(string) Either \code{NULL} or a string ("unofficial" by default).
Identifies where the Bronze/Silver/Gold data layers from which data is downloaded
are located.
\itemize{
\item If \code{NULL}, the root folder for the data layers is the 3B folder (i.e., where
the "official" data is stored). For example, \verb{3B/Silver/...}.
\item If a string, the root folder for the data layers is a sub-folder within the
Sandbox layer of the 3B data lake (e.g., if \code{experiment = "my_exp"}, then
data is download from \verb{3B/Sandbox/my_exp/\{data_layer\}/...})
\item If an empty string, the root folder for the data layers is the Sandbox itself
(i.e., if \code{experiment = ""}, then data is download from \verb{3B/Sandbox/\{data_layer\}/...})
}}

\item{silent}{(logical) Specifies whether to show authentication messages and
a download progress bar. Defaults to \code{TRUE}.}
}
\value{
A data frame.
}
\description{
Load Billions indicator data from xMart
}
