% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_whdh_billion_data}
\alias{load_whdh_billion_data}
\title{Load Billions indicator data}
\usage{
load_whdh_billion_data(
  billion = c("hep", "hpop", "uhc", "all"),
  ind_code,
  data_type = c("wrangled_data", "projected_data", "final_data"),
  na_rm = TRUE,
  silent = TRUE
)
}
\arguments{
\item{billion}{String. Billions data to load, one of "hep" (default), "hpop",
"uhc", or "all". If \code{all}, downloads data for all indicators for each of the
three billions.}

\item{data_type}{The type of data to load.
\itemize{
\item \code{wrangled_data} (default): raw data that has been wrangled into a suitable
form for analysis.
\item \code{projected_data}:  data that has been fully projected to the target but has
not yet been transformed or calculated upon.
\item \code{final_data}: the complete set of billions data with transformed values, contributions,
and all calculations available.
}}

\item{na_rm}{Logical. Specifies whether to filter the data to only rows
where \code{value} is not missing. Defaults to \code{TRUE}.}

\item{silent}{Logical. Specifies whether to show authentication messages and
a download progress bar. Defaults to \code{TRUE}.}

\item{ind_codes}{Character vector. The name of the indicator to load data for.
If \code{all}, downloads data for all indicators for a given billion. This argument
is ignored if \code{billion = "all"}.}
}
\value{
A data frame.
}
\description{
Load Billions indicator data
}
