% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_styles_modify.R
\name{modifyStyle}
\alias{modifyStyle}
\title{Modify Excel Style}
\usage{
modifyStyle(style, ...)
}
\arguments{
\item{style}{a \code{\link[openxlsx:createStyle]{openxlsx::createStyle()}} object to be modified}

\item{...}{any of  the parameters passed to \code{\link[openxlsx:createStyle]{openxlsx::createStyle()}}}
}
\description{
\code{modifyStyle} modifies a \code{\link[openxlsx:createStyle]{openxlsx::createStyle()}} object to add the
parameters
}
