% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hep_components.R
\name{pathogen_calc}
\alias{pathogen_calc}
\title{Calculate the vaccine coverage for a specific pathogen}
\usage{
pathogen_calc(
  df,
  name,
  numerators,
  denominators,
  transform_value_col,
  source,
  ind_ids,
  multiply_surviving_infs,
  max_value
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{name}{Name of pathogen to provide in the data frame.}

\item{numerators}{Indicator names for numerators.}

\item{denominators}{Indicator names for denominators.}

\item{transform_value_col}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{source}{Source to use if no unique source available for the calculation.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{multiply_surviving_infs}{Logical, multiple surviving infant population by
number of routine vaccines in numerator.}

\item{max_value}{Maximum value the calculated pathogen value can take.}
}
\description{
Using numerators and denominators pre-supplied for each ISO3 and year,
the overall vaccine coverage for a pathogen is calculated. This function
currently counts the number of routine vaccinations included in the numerator,
and multiplies the surviving infants denominator by that number.
}
\details{
This function is currently called from the \code{prevent_calculations()} function
that sits within \code{calculate_hep_components()}.
}
\keyword{internal}
