% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_links}
\alias{pop_links}
\title{HPOP Billion population links}
\format{
A data frame with 130 rows and 2 variables:
\describe{
\item{ind}{HPOP indicator code.}
\item{pop_group}{Population group.}
}
}
\usage{
pop_links
}
\description{
A dataset linking each HPOP Billion indicator to relevant population groups
to be used for double counting correction. Used within \code{generate_hpop_populations()}.
}
\keyword{datasets}
