% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hep_components.R
\name{prevent_calculations}
\alias{prevent_calculations}
\title{Calculate prevent indicators}
\usage{
prevent_calculations(df, scenario_col, source, transform_value_col, ind_ids)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{source}{Source to use if no unique source available for the calculation.}

\item{transform_value_col}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\description{
This function calculates the overall vaccine coverage scores for each pathogen,
including the overall prevent indicator. It is used within \code{calculate_hep_components()}
and primarily relies on using \code{purrr::pmap_dfr()} to apply the \code{pathogen_calc()}
function for each prevent component and overall score.
}
