% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recycle_data.R
\name{recycle_data_scenario_single}
\alias{recycle_data_scenario_single}
\title{Recycle data between scenarios for a single scenario}
\usage{
recycle_data_scenario_single(
  df,
  scenario,
  billion = c("hep", "hpop", "uhc"),
  value_col = "value",
  start_year = 2018,
  end_year = 2025,
  scenario_col = "scenario",
  default_scenario = "default",
  scenario_reported_estimated = "routine",
  scenario_covid_shock = "covid_shock",
  scenario_reference_infilling = "reference_infilling",
  include_projection = TRUE,
  recycle_campaigns = TRUE,
  ind_ids = NULL,
  trim_years = FALSE,
  start_year_trim = start_year,
  end_year_trim = end_year,
  assert_data_calculations = TRUE
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{scenario}{name of scenario to recycle for.}

\item{billion}{name of billion to recycle data for.}

\item{value_col}{Column name of column with indicator values.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{scenario_col}{Column name of column with scenario identifiers.}

\item{default_scenario}{name of the default scenario.}

\item{scenario_reported_estimated}{name of the reported/estimated scenario.}

\item{scenario_covid_shock}{name of the scenario with the COVID-19 shock years.}

\item{scenario_reference_infilling}{name of the WHO technical programs projections/imputations scenario.}

\item{include_projection}{Boolean to include or not projections in recycling}

\item{recycle_campaigns}{Boolean to include or not campaigns in recycling}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{trim_years}{logical to indicate if years before \code{start_year} and after
\code{end_year} should be removed}

\item{start_year_trim}{(integer) year to start trimming from.}

\item{end_year_trim}{(integer) year to end trimming.}

\item{assert_data_calculations}{Boolean if true then output data frame will
be tested to see if it contains the minimal required data to run the
calculations.}
}
\value{
Data frame in long format.
}
\description{
\code{recycle_data_scenario_single } reuses values present in the specified
scenarios in \code{default_scenario}, \code{scenario_reported_estimated},
\code{scenario_covid_shock} and \code{scenario_reference_infilling} for the specified
scenarios.
}
\details{
To do so, it looks at:
\enumerate{
\item values in \code{default_scenario} but not in the scenario specified
\item values in \code{scenario_reported_estimated} or \code{scenario_covid_shock} but not
in the scenario specified or \code{default_scenario}.
\item values in \code{scenario_reference_infilling} but not in the scenario specified,
\code{scenario_reported_estimated}, \code{scenario_covid_shock}, or
\code{scenario_reference_infilling}
}

For more information see:

\code{vignette("scenarios", package = "billionaiRe")}
}
