% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_type.R
\name{reduce_type}
\alias{reduce_type}
\title{Helper to reduce type to unique value}
\usage{
reduce_type(value, type)
}
\arguments{
\item{value}{Vector of values}

\item{type}{Vector of types}
}
\description{
Used with \code{pathogen_calc()}
}
