% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{remove_unused_wash_ind}
\alias{remove_unused_wash_ind}
\title{Removes unused WASH indicator from indicator list}
\usage{
remove_unused_wash_ind(ind_list, correct_wash)
}
\arguments{
\item{ind_list}{character vector containing the list of indicator to be treated}

\item{correct_wash}{character vector with the exact indicators to keep for WASH}
}
\description{
\code{remove_unused_wash_ind} removes unused WASH indicators from indicator list to
provide a clean list to be printed.
}
