% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_whdh.R
\name{save_gho_backup_to_whdh}
\alias{save_gho_backup_to_whdh}
\title{Backup GHO data to the WHDH Data Lake}
\usage{
save_gho_backup_to_whdh(df, billion, ind_code)
}
\arguments{
\item{df}{The response from the GHO API, typically the result of calling
\code{ghost::gho_data()}.}

\item{billion}{The billion to which the indicator belongs}

\item{ind_code}{The GPW13 code for the indicator}
}
\value{
The original data frame
}
\description{
This function simplifies the process of creating backups snapshots of GHO data
and saving these to the WHDH data lake.
}
\details{
It returns the input data frame, without any changes.
}
