% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_wrangling.R
\name{save_wrangled_output}
\alias{save_wrangled_output}
\title{Save the output to disk after ensuring column specs}
\usage{
save_wrangled_output(df, ind)
}
\arguments{
\item{df}{data frame the output}

\item{ind}{string name of the indicator}
}
\description{
Helper functions that saves a wrangled output data frame to the disk if it has
the correct columns as required by the Triple Billions xMart tables.
}
