% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios_specific_scenarios.R
\name{scenario_bau}
\alias{scenario_bau}
\title{Scenario establish a business as usual scenario}
\usage{
scenario_bau(
  df,
  only_reported_estimated = FALSE,
  value_col = "value",
  start_year = 2018,
  end_year = 2025,
  scenario_name = glue::glue("business_as_usual"),
  scenario_col = "scenario",
  avoid_worstening = FALSE,
  trim = TRUE,
  small_is_best = FALSE,
  keep_better_values = TRUE,
  upper_limit = 100,
  lower_limit = 0,
  trim_years = TRUE,
  start_year_trim = start_year,
  end_year_trim = end_year,
  ind_ids = billion_ind_codes("all"),
  bau_scenario = "historical",
  default_scenario = "default"
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{only_reported_estimated}{(logical) if TRUE only the last \code{reported} and \code{estimated}
values are imputed.}

\item{value_col}{Column name of column with indicator values.}

\item{start_year}{Start year for scenario, defaults to 2018.}

\item{end_year}{End year for scenario, defaults to 2025}

\item{scenario_name}{Name of the scenario. Defaults to scenario_{percent_change}_{baseline_year}}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{avoid_worstening}{If TRUE, the value at \code{start_year} is kept if it is
better (see \code{small_is_best}).}

\item{trim}{logical to indicate if the data should be trimmed between
\code{upper_limit} and \code{lower_limit}.}

\item{small_is_best}{Logical to identify if a lower value is better than a higher
one (e.g. lower obesity in a positive public health outcome, so obesity rate
should have small_is_best = TRUE).}

\item{keep_better_values}{logical to indicate if "better" values should be
kept from \code{value_col} if they are present. Follows the direction set in
\code{small_is_best}.  For instance, if small_is_best is TRUE, then \code{value_col} lower
than \code{col} will be kept.}

\item{upper_limit}{limit at which the indicator should be caped.
Can take any of "guess", or any numeric. \code{guess} (default) will take 100 as
the limit if \code{percent_change} is positive, and 0 if negative.}

\item{lower_limit}{limit at which the indicator should be caped.
Can take any of "guess", or 0 to 100. \code{guess} (default) will take 0 as the
limit if \code{percent_change} is positive, and 100 if negative.}

\item{trim_years}{logical to indicate if years before \code{start_year_trim} and after
\code{end_year_trim} should be removed}

\item{start_year_trim}{(integer) year to start trimming from.}

\item{end_year_trim}{(integer) year to end trimming.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{bau_scenario}{name of scenario to be used for business as usual.
Default is \code{historical}.}

\item{default_scenario}{name of the default scenario to be used.}
}
\value{
Dataframe with scenario rows
}
\description{
\code{scenario_bau} filters for values between start_year and end_year for \code{bau_scenario} and
returns values in value. If values are missing for years between \code{start_year} and \code{end_year},
the last available value will be imputed.
}
\details{
The returned scenario is a portion of the straight line drawn from the
\code{baseline_year} value to the \code{target_year}. Only values for years between
\code{start_year} and \code{end_year} will be returned.

If \code{value_col} has values that are higher or lower than the scenario values, then
only \code{value_col} will be kept, depending on \code{small_is_best}. For instance,
if the scenario value is 80 and the value 75 and small_is_best is TRUE, then
75 will be kept.
}
