% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios_specific_scenarios.R
\name{scenario_best_of}
\alias{scenario_best_of}
\title{Scenario to pick the best scenario out of a list of scenarios}
\usage{
scenario_best_of(
  df,
  scenario_names,
  value_col = "value",
  maximize_end_year = FALSE,
  start_year = 2018,
  end_year = 2025,
  target_year = end_year,
  scenario_name = glue::glue("best_of_{paste0(scenario_names, collapse = '_')}"),
  scenario_col = "scenario",
  trim = TRUE,
  small_is_best = FALSE,
  keep_better_values = TRUE,
  upper_limit = 100,
  lower_limit = 0,
  trim_years = TRUE,
  ind_ids = billion_ind_codes("all")
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{scenario_names}{names of the scenario to pick from.}

\item{value_col}{Column name of column with indicator values.}

\item{maximize_end_year}{(Boolean) if TRUE, the best scenario is
picked on the best value at \code{end_year}. Default to FALSE.}

\item{start_year}{Start year for scenario, defaults to 2018.}

\item{end_year}{End year for scenario, defaults to 2025}

\item{target_year}{Year by which the scenario should eventually be
achieved. Defaults to \code{end_year}}

\item{scenario_name}{Name of the scenario. Defaults to scenario_{percent_change}_{baseline_year}}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{trim}{logical to indicate if the data should be trimmed between
\code{upper_limit} and \code{lower_limit}.}

\item{small_is_best}{Logical to identify if a lower value is better than a higher
one (e.g. lower obesity in a positive public health outcome, so obesity rate
should have small_is_best = TRUE).}

\item{keep_better_values}{logical to indicate if "better" values should be
kept from \code{value_col} if they are present. Follows the direction set in
\code{small_is_best}.  For instance, if small_is_best is TRUE, then \code{value_col} lower
than \code{col} will be kept.}

\item{upper_limit}{limit at which the indicator should be caped.
Can take any of "guess", or any numeric. \code{guess} (default) will take 100 as
the limit if \code{percent_change} is positive, and 0 if negative.}

\item{lower_limit}{limit at which the indicator should be caped.
Can take any of "guess", or 0 to 100. \code{guess} (default) will take 0 as the
limit if \code{percent_change} is positive, and 100 if negative.}

\item{trim_years}{logical to indicate if years before \code{start_year} and after
\code{end_year} should be removed}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
Dataframe with scenario rows
}
\description{
\code{scenario_best_of} picks the best value between multiple scenarios specified
in \code{scenario_names}. \code{small_is_best} allows to pick if higher value or
smaller values are best.
}
\details{
If multiple scenario are tied for the best scenario, the scenario with the best time
series for each \code{iso3} and \code{ind} is picked. If they are still tied scenarios for
the best value, the first in alphabetical order is picked.
}
