% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios_covid.R
\name{scenario_covid_delayed_return}
\alias{scenario_covid_delayed_return}
\title{Scenario COVID delayed return to normal}
\usage{
scenario_covid_delayed_return(
  df,
  start_year = 2018,
  covid_year = 2020,
  recovery_year = 2023,
  end_year = 2025,
  value_col = "value",
  scenario_col = "scenario",
  scenario_name = "covid_delayed_return",
  ind_ids = billion_ind_codes("all"),
  default_scenario = "default",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{covid_year}{(integer) year where the values are impacted by COVID.}

\item{recovery_year}{integer year from which the AROC will be applied. Default to 2023.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{value_col}{Column name of column with indicator values. This column will be
used to return the results.}

\item{scenario_col}{name of scenario column to be created}

\item{scenario_name}{name of scenario}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{default_scenario}{name of the default scenario.}

\item{...}{additional parameters to be passed to
\code{scenario_dip_recover_iso3()}}
}
\value{
a data frame with scenario values in \code{value_col} with a \code{scenario_col} column.
}
\description{
\code{scenario_covid_delayed_return()} creates a scenario where there is a
a delayed by number for years between \code{covid_year} and \code{recovery_year} before
returning to the pre-pandemic situation.
}
\details{
In details, the AROC  between the \code{start_year}  and \code{covid_year} - 1 is
applied to the last reported value to \code{recovery_year} onward. If there are
missing values between \code{covid_year} and \code{recovery_year}, the last value from
\code{covid_year} is carried forward. This applies only to countries where the
indicator value for \code{covid_year} is \code{reported} or \code{estimated}. Otherwise, the
value is carried with \code{scenario_bau}.
}
\seealso{
COVID scenarios
\code{\link{scenario_covid_never_return}()},
\code{\link{scenario_covid_rapid_return}()},
\code{\link{scenario_covid_sustained_disruption}()},
\code{\link{scenario_dip_recover}()},
\code{\link{scenario_return_previous_trajectory}()}
}
\concept{scenario_covid}
\keyword{internal}
