% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios_covid.R
\name{scenario_covid_rapid_return}
\alias{scenario_covid_rapid_return}
\title{Scenario COVID rapid return to normal}
\usage{
scenario_covid_rapid_return(
  df,
  year = "year",
  ind = "ind",
  iso3 = "iso3",
  start_year = 2018,
  covid_year = 2020,
  recovery_year = 2022,
  end_year = 2025,
  value = "value",
  scenario = "scenario",
  scenario_name = "covid_rapid_return",
  ind_ids = billion_ind_codes("all"),
  default_scenario = "default",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{year}{Column name of column with years.}

\item{ind}{Column name of column with indicator names.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{start_year}{Start year for scenario, defaults to 2018.}

\item{covid_year}{(integer) year where the values are impacted by COVID.}

\item{recovery_year}{integer year from which the AROC will be applied. Default to 2022.}

\item{end_year}{End year for scenario, defaults to 2025}

\item{value}{Column name of column with indicator values. This column will be
used to return the results.}

\item{scenario}{name of scenario column to be created}

\item{scenario_name}{name of scenario}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{default_scenario}{name of the default scenario.}

\item{...}{additional parameters to be passed to
\code{scenario_dip_recover_iso3()}}
}
\value{
a data frame with scenario values in \code{value} with a \code{scenario} column.
}
\description{
\code{scenario_covid_rapid_return} creates a scenario where there is a rapid
return to the pre-pandemic situation after a dip due to COVID-19.
}
\details{
In details, the AROC  between the \code{start_year}  and \code{covid_year} - 1 is
applied to the last reported value to \code{recovery_year} onward. If there are
missing values between \code{covid_year} and \code{recovery_year}, the last value from
\code{covid_year} is carried forward. This applies only to countries where the
indicator value for \code{covid_year} is \code{reported} or \code{estimated}. Otherwise,
the value is carried with \code{scenario_bau}.
}
