% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios_dip_recover.R
\name{scenario_dip_recover}
\alias{scenario_dip_recover}
\title{Scenario dip and recover}
\usage{
scenario_dip_recover(
  df,
  start_year = 2018,
  dip_year = 2020,
  recovery_year = 2021,
  progressive_recovery = FALSE,
  aroc_type = c("lastest_year", "average_years_in_range"),
  aroc_start_year = start_year,
  end_year = 2025,
  value_col = "value",
  scenario_col = "scenario",
  scenario_name = "dip_recover",
  ind_ids = billion_ind_codes("all"),
  default_scenario = "default",
  source = sprintf("WHO DDI, \%s", format(Sys.Date(), "\%B \%Y")),
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{start_year}{Start year for scenario, defaults to 2018.}

\item{dip_year}{(integer) year where the dip appends}

\item{recovery_year}{(integer) year from which the AROC will be applied}

\item{progressive_recovery}{(logical) TRUE if the recovery after dip
should be progressive.}

\item{aroc_type}{(character) name of the type of AROC to be used. Can be either:
\itemize{
\item \code{lastest_year}: only the latest \code{reported}/\code{estimated} year between \code{start_year}
and \code{dip_year} is used to calculate
\item \code{average_years_in_range}:
}}

\item{aroc_start_year}{(integer) year}

\item{end_year}{End year for scenario, defaults to 2025}

\item{value_col}{Column name of column with indicator values. This column will be
used to return the results.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{scenario_name}{name of scenario}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{default_scenario}{name of the default scenario.}

\item{source}{Source to use if no unique source available for the calculation.}

\item{...}{additional parameters to be passed to
\code{scenario_dip_recover_iso3()}}

\item{scenario}{name of scenario column to be created}
}
\value{
a data frame with scenario values in \code{value_col} with a \code{scenario} column.
}
\description{
\code{scenario_dip_recover} creates a scenario where there is a rapid
return to the previous situation after a dip. The same annual rate of change (AROC) as
before dip is applied from the \code{recovery_year}.
}
\details{
Two types of AROC are supported:
\itemize{
\item \code{lastest_year}: the AROC is calculated the values between the \code{start_year} and the last \code{reported} or
\code{estimated} value before \code{dip_year} is applied to the last reported value to
\code{recovery_year} onward.
\item \code{average_years_in_range}: all AROC between the \code{aroc_start_year} and the last \code{reported} or
\code{estimated} value before \code{dip_year} are calculated. The average AROC for that period is then
applied to the last reported value to \code{recovery_year} onward.
}

If there are missing values between \code{dip_year} and
\code{recovery_year}, the last value from \code{dip_year} is carried forward. This
applies only to countries where the indicator value for \code{dip_year} is
\code{reported} or \code{estimated}. Otherwise, the value is carried with
\code{scenario_bau}.

If \code{progressive_recovery} is TRUE, then the recovery is
spread between the years between \code{recovery_year} and \code{end_year}. For
instance, if \code{recovery_year} is 2021 and \code{end_year} 2025, then 2021 will have 0\%
of AROC, 2022 25\%, 2023 50\%, 2024 75\%, and 2025 100\%.
}
