% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios_dip_recover.R
\name{scenario_dip_recover_iso3}
\alias{scenario_dip_recover_iso3}
\title{Scenario dip and recover to specific iso3}
\usage{
scenario_dip_recover_iso3(
  df,
  iso3,
  start_year = 2018,
  dip_year = 2020,
  recovery_year = 2021,
  progressive_recovery = FALSE,
  aroc_type = c("lastest_year", "average_years_in_range"),
  aroc_start_year = start_year,
  end_year = 2025,
  value_col = "value",
  scenario_col = "scenario",
  scenario_name = "dip_recover",
  trim = TRUE,
  small_is_best = FALSE,
  keep_better_values = FALSE,
  upper_limit = 100,
  lower_limit = 0,
  trim_years = TRUE,
  start_year_trim = start_year,
  end_year_trim = end_year,
  ind_ids = billion_ind_codes("all"),
  source = sprintf("WHO DDI, \%s", format(Sys.Date(), "\%B \%Y")),
  default_scenario = "default"
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{iso3}{(character) ISO3 code of country to scenario}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{dip_year}{(integer) year where the dip appends}

\item{recovery_year}{(integer) year from which the AROC will be applied}

\item{progressive_recovery}{(logical) TRUE if the recovery after dip
should be progressive.}

\item{aroc_type}{(character) name of the type of AROC to be used. Can be either:
\itemize{
\item \code{lastest_year}: only the latest \code{reported}/\code{estimated} year between \code{start_year}
and \code{dip_year} is used to calculate
\item \code{average_years_in_range}:
}}

\item{aroc_start_year}{(integer) year}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{value_col}{Column name of column with indicator values. This column will be
used to return the results.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{scenario_name}{name of scenario}

\item{trim}{logical to indicate if the data should be trimmed between
\code{upper_limit} and \code{lower_limit}.}

\item{small_is_best}{Logical to identify if a lower value is better than a higher
one (e.g. lower obesity in a positive public health outcome, so obesity rate
should have small_is_best = TRUE).}

\item{keep_better_values}{logical to indicate if "better" values should be
kept from \code{value_col} if they are present. Follows the direction set in
\code{small_is_best}.  For instance, if small_is_best is TRUE, then \code{value_col} lower
than \code{col} will be kept.}

\item{upper_limit}{upper limit at which the indicator should be caped.}

\item{lower_limit}{lower_limit limit at which the indicator should be caped.}

\item{trim_years}{logical to indicate if years before \code{start_year_trim} and after
\code{end_year_trim} should be removed}

\item{start_year_trim}{(integer) year to start trimming from.}

\item{end_year_trim}{(integer) year to end trimming.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{source}{Source to use if no unique source available for the calculation.}

\item{default_scenario}{name of the default scenario.}
}
\description{
Applies \code{scenario_dip_recover()} to a specific iso3.
}
