% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios_percent_change.R
\name{scenario_linear_percent_change_col}
\alias{scenario_linear_percent_change_col}
\title{Scenario to add a linear percentage point change stored in a column}
\usage{
scenario_linear_percent_change_col(
  df,
  linear_value,
  value = "value",
  ind = "ind",
  iso3 = "iso3",
  year = "year",
  start_year = 2018,
  end_year = 2025,
  baseline_year = start_year,
  target_year = end_year,
  scenario_name = glue::glue("linear_percent_change"),
  upper_limit = 100,
  lower_limit = 0
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{linear_value}{name of column with linear values}

\item{value}{Column name of column with indicator values.}

\item{ind}{Column name of column with indicator names.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{year}{Column name of column with years.}

\item{start_year}{Start year for scenario, defaults to 2018.}

\item{end_year}{End year for scenario, defaults to 2025}

\item{baseline_year}{Year from which the scenario is measured.
Defaults to \code{start_year}}

\item{target_year}{Year by which the scenario should eventually be
achieved. Defaults to \code{end_year}}

\item{scenario_name}{Name of the scenario. Defaults to scenario_{percent_change}_{baseline_year}}
}
\value{
Dataframe with additional scenario rows
}
\description{
\code{scenario_linear_percent_change_col} wraps around
\code{scenario_linear_percent_change} to provide linear values from a column
specified in \code{linear_value} rather than a single value.
}
\details{
The returned scenario is a portion of the straight line drawn from the
\code{baseline_year} value to the \code{target_year}. Only values for years between
\code{start_year} and \code{end_year} will be returned.
}
