% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_hpop.R
\name{sdg_adult_obese}
\alias{sdg_adult_obese}
\title{Accelerate adult_obese to SDG target}
\usage{
sdg_adult_obese(
  df,
  ind_ids = billion_ind_codes("hpop"),
  scenario_col = "scenario",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{...}{additional parameters to be passed to scenario function}
}
\value{
data frame with acceleration scenario binded to \code{df}. \code{scenario} is
set to \code{sdg}
}
\description{
Accelerate adult_obese by halting upwards trend in the data to the 2010 value.
Runs:
\itemize{
\item \code{scenario_halt_rise(df, baseline_year = 2010, small_is_best = TRUE,...)}.
}
}
\seealso{
SDG scenarios
\code{\link{sdg_alcohol}()},
\code{\link{sdg_child_obese}()}
}
\concept{sdg}
