% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_uhc.R
\name{sdg_art}
\alias{sdg_art}
\title{Accelerate art to SDG target}
\usage{
sdg_art(df, end_year = 2030, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate \code{art} by aiming at 100 by 2030 if there are 2 values or more
reported. Otherwise, business as usual.
}
