% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_hpop.R
\name{sdg_child_obese}
\alias{sdg_child_obese}
\title{Accelerate child_obese to SDG target}
\usage{
sdg_child_obese(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate child_obese by halting upwards trend in the data to the 2010 value.
}
\seealso{
SDG scenarios
\code{\link{sdg_adult_obese}()},
\code{\link{sdg_alcohol}()}

Other hpop: 
\code{\link{accelerate_adult_obese}()}
}
\concept{hpop}
\concept{sdg}
