% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_hpop.R
\name{sdg_devontrack}
\alias{sdg_devontrack}
\title{Accelerate devontrack to SDG target}
\usage{
sdg_devontrack(
  df,
  ind_ids = billion_ind_codes("hpop"),
  end_year = 2025,
  scenario_col = "scenario",
  start_year = 2018,
  value_col = "value",
  default_scenario = "default",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{value_col}{Name of the column containing indicator value in \code{df}.}

\item{default_scenario}{name of the default scenario.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Put devontrack on SDG trajectory by targeting 100 by 2030.
}
\seealso{
HPOP SDG scenarios
\code{\link{sdg_adult_obese}()},
\code{\link{sdg_alcohol}()},
\code{\link{sdg_child_obese}()},
\code{\link{sdg_child_viol}()},
\code{\link{sdg_fuel}()},
\code{\link{sdg_hpop_sanitation_rural}()},
\code{\link{sdg_hpop_sanitation_urban}()},
\code{\link{sdg_hpop_sanitation}()},
\code{\link{sdg_hpop_tobacco}()},
\code{\link{sdg_ipv}()},
\code{\link{sdg_overweight}()},
\code{\link{sdg_pm25}()},
\code{\link{sdg_road}()},
\code{\link{sdg_stunting}()},
\code{\link{sdg_suicide}()},
\code{\link{sdg_transfats}()},
\code{\link{sdg_wasting}()},
\code{\link{sdg_water_rural}()},
\code{\link{sdg_water_urban}()},
\code{\link{sdg_water}()}
}
\concept{hpop_sdg}
