% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_uhc.R
\name{sdg_doctors}
\alias{sdg_doctors}
\title{Accelerate doctors to SDG target}
\usage{
sdg_doctors(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate doctors using the business as usual scenario.
}
