% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_uhc.R
\name{sdg_fh}
\alias{sdg_fh}
\title{Accelerate fh to SDG target}
\usage{
sdg_fh(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate \code{fh} by aiming at 0 by 2030 if there are 2 values or more
reported. Otherwise, business as usual.
}
