% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_uhc.R
\name{sdg_fp}
\alias{sdg_fp}
\title{Accelerate fp to SDG target}
\usage{
sdg_fp(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate \code{dtp3} by aiming at 100 by 2030 if there are 2 values or more
reported. Otherwise, business as usual.
}
