% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_uhc.R
\name{sdg_fpg}
\alias{sdg_fpg}
\title{Accelerate fpg to SDG target}
\usage{
sdg_fpg(
  df,
  start_year = 2018,
  scenario_name = "sdg",
  scenario_col = "scenario",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{start_year}{Year from which the acceleration scenario begins, inclusive.}

\item{scenario_name}{name of scenario}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate \code{fpg} by aiming at 0 by 2030 if there are 2 values or more
reported. Otherwise, business as usual.
}
