% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_hpop.R
\name{sdg_hpop_tobacco}
\alias{sdg_hpop_tobacco}
\title{Accelerate hpop_tobacco to SDG target}
\usage{
sdg_hpop_tobacco(
  df,
  ind_ids = billion_ind_codes("hpop"),
  scenario_col = "scenario",
  start_year = 2018,
  end_year = 2025,
  start_year_trim = start_year,
  end_year_trim = end_year,
  default_scenario = "default",
  value_col = "value",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{start_year_trim}{(integer) year to start trimming from.}

\item{end_year_trim}{(integer) year to end trimming.}

\item{default_scenario}{name of the default scenario.}

\item{value_col}{Column name of column with indicator values.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate hpop_tobacco by running a custom version of
scenario_percent_baseline. The custom \code{scenario_percent_baseline} is
taking similar parameters to \code{scenario_percent_baseline}'s
\code{percent_change} = -30, \code{baseline_year} = 2010, but values are added to the
\code{start_year} value, rather than the \code{baseline_year} values.
}
