% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_uhc.R
\name{sdg_hwf}
\alias{sdg_hwf}
\title{Accelerate hwf to SDG target}
\usage{
sdg_hwf(
  df,
  ind_ids = billion_ind_codes("uhc"),
  scenario_col = "scenario",
  start_year = 2018,
  default_scenario = "default",
  bau_scenario = "historical",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{default_scenario}{name of the default scenario.}

\item{bau_scenario}{name of scenario to be used for business as usual.
Default is \code{historical}.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate \code{hwf} by returning to business as usual, as there are no globally
agreed target.
}
