% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_hpop.R
\name{sdg_ipv}
\alias{sdg_ipv}
\title{Accelerate ipv to SDG target}
\usage{
sdg_ipv(df, scenario_col = "scenario", ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Put hpop_tobacco on SDG trajectory by targeting 0 by 2030.
}
\seealso{
HPOP SDG scenarios
\code{\link{sdg_adult_obese}()},
\code{\link{sdg_alcohol}()},
\code{\link{sdg_child_obese}()},
\code{\link{sdg_child_viol}()},
\code{\link{sdg_devontrack}()},
\code{\link{sdg_fuel}()},
\code{\link{sdg_hpop_sanitation_rural}()},
\code{\link{sdg_hpop_sanitation_urban}()},
\code{\link{sdg_hpop_sanitation}()},
\code{\link{sdg_hpop_tobacco}()},
\code{\link{sdg_overweight}()},
\code{\link{sdg_pm25}()},
\code{\link{sdg_road}()},
\code{\link{sdg_stunting}()},
\code{\link{sdg_suicide}()},
\code{\link{sdg_transfats}()},
\code{\link{sdg_wasting}()},
\code{\link{sdg_water_rural}()},
\code{\link{sdg_water_urban}()},
\code{\link{sdg_water}()}
}
\concept{hpop_sdg}
