% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_uhc.R
\name{sdg_uhc_sanitation}
\alias{sdg_uhc_sanitation}
\title{Accelerate uhc_sanitation to SDG target}
\usage{
sdg_uhc_sanitation(df, ind_ids = billion_ind_codes("uhc"), ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate \code{uhc_sanitation} by aiming at 100 by 2030 if there are 2 values or more
reported. Otherwise, business as usual.
}
