% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_hpop.R
\name{sdg_wasting}
\alias{sdg_wasting}
\title{Accelerate wasting to SDG targets}
\usage{
sdg_wasting(
  df,
  ind_ids = billion_ind_codes("hpop"),
  start_year = 2018,
  scenario_col = "scenario",
  default_scenario = "default",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{start_year}{Year from which the acceleration scenario begins, inclusive.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{default_scenario}{name of the default scenario.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate wasting by aiming at 0\% by 2030.
}
