% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_hpop.R
\name{sdg_water}
\alias{sdg_water}
\title{Accelerate water to SDG target}
\usage{
sdg_water(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Put water on SDG trajectory by aiming at 100\% by 20230
}
\seealso{
HPOP SDG scenarios
\code{\link{sdg_adult_obese}()},
\code{\link{sdg_alcohol}()},
\code{\link{sdg_child_obese}()},
\code{\link{sdg_child_viol}()},
\code{\link{sdg_devontrack}()},
\code{\link{sdg_fuel}()},
\code{\link{sdg_hpop_sanitation_rural}()},
\code{\link{sdg_hpop_sanitation_urban}()},
\code{\link{sdg_hpop_sanitation}()},
\code{\link{sdg_hpop_tobacco}()},
\code{\link{sdg_ipv}()},
\code{\link{sdg_overweight}()},
\code{\link{sdg_pm25}()},
\code{\link{sdg_road}()},
\code{\link{sdg_stunting}()},
\code{\link{sdg_suicide}()},
\code{\link{sdg_transfats}()},
\code{\link{sdg_wasting}()},
\code{\link{sdg_water_rural}()},
\code{\link{sdg_water_urban}()}
}
\concept{hpop_sdg}
