% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_hpop.R
\name{sdg_water_rural}
\alias{sdg_water_rural}
\title{Accelerate water_rural to SDG target}
\usage{
sdg_water_rural(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\value{
data frame with acceleration scenario binded to \code{df}. \code{scenario_col} is
set to \code{acceleration}
}
\description{
Accelerate water_rural by aiming at 95\% by 20230
}
\details{
Runs:
\itemize{
\item \code{scenario_fixed_target(df, target_value = 95,target_year = 2030, small_is_best = FALSE)}
}
}
