% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_uhc_summary_sheet.R
\name{style_asc_uhc_data_summary}
\alias{style_asc_uhc_data_summary}
\title{Style ASC Pillars summary sheet}
\usage{
style_asc_uhc_data_summary(
  wb,
  sheet_name,
  boxes_bounds,
  data_type,
  pillar = "fin_hardship",
  no_show = TRUE,
  no_show_cols,
  no_show_rows
)
}
\arguments{
\item{wb}{a \code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{boxes_bounds}{named list of bounds for data frame boxes to be written in sheet.}

\item{data_type}{named list with latest_reported and baseline_projection data
types. Passed to \code{style_data()}}

\item{pillar}{character Pillar identifying the pillar to style.}

\item{no_show}{Boolean with TRUE if data in \code{no_show_row} should be displayed}

\item{no_show_cols}{Boolean with TRUE if data in \code{no_show_row} should be displayed}

\item{no_show_rows}{integer identifying which row contains data no to show}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{style_asc_uhc_data_summary()} styles the UHC ASC box/section of the UHC
summary sheet. Used within \code{write_asc_uhc_data_summary()}.
}
