% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_hpop_summary_sheet.R
\name{style_billion_contrib_ind_hpop}
\alias{style_billion_contrib_ind_hpop}
\title{Style HPOP summary sheet contribution to billion}
\usage{
style_billion_contrib_ind_hpop(wb, sheet_name, bounds, data_type)
}
\arguments{
\item{wb}{\code{\link[openxlsx:createWorkbook]{openxlsx::createWorkbook()}} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{bounds}{named integer vector identifying integer sheet start and end reference column.
All parameters must be numerical. For converting Excel column references to integer references
see \code{\link[openxlsx:int2col]{openxlsx::int2col()}}.
Must have the following named elements:
\itemize{
\item \code{start_col}: start column
\item \code{end_col}: end column
\item \code{start_row}: start row
\item \code{end_row}: end row
}}

\item{data_type}{character vector of class(es) of the data to be styled.
Can be one of "numeric", "integer", "Date", "character" or "character_centered"}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
Used within \code{write_bilion_contrib_ind_hpop_summary()}
}
