% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{style_data}
\alias{style_data}
\title{Style data according to its type}
\usage{
style_data(df, wb, sheet_name, rows, cols)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{openxlsx workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{rows}{Rows to apply style to.}

\item{cols}{columns to apply style to.}
}
\description{
Style data according to its type
}
