% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{style_data_type}
\alias{style_data_type}
\title{Style data based on type of data}
\usage{
style_data_type(df, wb, iso, sheet_name, start_col, start_row, type_col)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{openxlsx workbook to be edited.}

\item{iso}{ISO3 codes of country to summarize.}

\item{sheet_name}{character name of the sheet to update}

\item{start_col}{numeric start row of the table in the excel sheet}

\item{start_row}{numeric start row of the table in the excel sheet}

\item{type_col}{Column name of column to store type.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_fldr}.
}
\description{
\code{export_country_summary_xls} Export a country-specific for all three
billions or for a specific billion.
}
\examples{
\dontrun{
data <- load_billion_data("hpop")
export_country_summary_xls(data, iso = "AFG")
}
}
