% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{style_data_type_timesseries}
\alias{style_data_type_timesseries}
\title{Style data based on type of data}
\usage{
style_data_type_timesseries(
  df,
  wb,
  sheet_name,
  start_col,
  start_row,
  ind,
  type_col,
  year
)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{openxlsx workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{start_col}{numeric start row of the table in the excel sheet}

\item{start_row}{numeric start row of the table in the excel sheet}

\item{ind}{Column name of column with indicator names.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{export_country_summary_xls} Export a country-specific for all three
billions or for a specific billion.
}
\examples{
\dontrun{
data <- load_billion_data("hpop")
export_country_summary_xls(data, iso = "AFG")
}
}
