% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_hep_summary_sheet.R
\name{style_header_hep_summary_sheet}
\alias{style_header_hep_summary_sheet}
\title{Style HEP summary worksheet header}
\usage{
style_header_hep_summary_sheet(wb, sheet_name, boxes_bounds)
}
\arguments{
\item{wb}{a \code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{boxes_bounds}{named list of bounds for data frame boxes to be written in sheet.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{style_header_hep_summary_sheet} styles the title and sub-title of the summary
worksheet header.
}
