% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_hpop_summary_sheet.R
\name{style_header_hpop_summary_sheet}
\alias{style_header_hpop_summary_sheet}
\title{Style worksheet header}
\usage{
style_header_hpop_summary_sheet(wb, sheet_name, start_row, start_col)
}
\arguments{
\item{wb}{\code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{start_row}{numeric start row of the table in the excel sheet}

\item{start_col}{numeric start row of the table in the excel sheet}
}
\description{
\code{style_header} styles the title and sub-title of the worksheet header.
}
