% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_uhc_summary_sheet.R
\name{style_header_uhc_summary_sheet}
\alias{style_header_uhc_summary_sheet}
\title{Style UHC summary worksheet header}
\usage{
style_header_uhc_summary_sheet(wb, sheet_name, bounds)
}
\arguments{
\item{wb}{\code{\link[openxlsx:createWorkbook]{openxlsx::createWorkbook()}} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{style_header_uhc_summary_sheet} styles the title and sub-title of the summary
worksheet header.
}
