% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_hep_summary_sheet.R
\name{style_hep_pillar}
\alias{style_hep_pillar}
\title{Style HEP Pillars summary sheet}
\usage{
style_hep_pillar(
  wb,
  sheet_name,
  boxes_bounds,
  data_type,
  pillar = c("prepare", "prevent", "detect_respond"),
  fade = FALSE,
  fade_row = 0
)
}
\arguments{
\item{wb}{a \code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{boxes_bounds}{named list of bounds for data frame boxes to be written in sheet.}

\item{data_type}{named list with latest_reported and baseline_projection data
types. Passed to \code{style_data()}}

\item{pillar}{character Pillar identifying the pillar to style. Must be one of
\itemize{
\item RMNCH
\item infectious
\item NCD
\item service
}}

\item{fade}{Boolean with TRUE if rows in \code{fade_row} should be faded}

\item{fade_row}{integer identifying which row contains data no to show}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{style_hep_pillar()} styles the HEP Pillars (Prevent, Prepare, Detect and Respond)
box/section of the HEP summary sheet. Used within \code{write_data_boxes_hep_summary()}.
}
