% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_hpop_datasheet.R
\name{style_hpop_main_data}
\alias{style_hpop_main_data}
\title{Style HPOP data sheet}
\usage{
style_hpop_main_data(
  df,
  wb,
  sheet_name,
  start_row,
  start_col,
  length_baseline_header,
  length_contrib_header
)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{openxlsx workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{start_row}{numeric start row of the table in the excel sheet}

\item{start_col}{numeric start row of the table in the excel sheet}

\item{length_baseline_header}{integer with length of the baseline sub-header}

\item{length_contrib_header}{integer with length of the billion's contributions
sub-header}
}
\description{
Applies styles to the data worksheet for country summary outputs
}
