% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_indicator_list.R
\name{style_indicator_list_sheet}
\alias{style_indicator_list_sheet}
\title{Styles indicator sheet}
\usage{
style_indicator_list_sheet(
  df,
  wb,
  sheet_name,
  start_row,
  end_row,
  start_col,
  end_col
)
}
\arguments{
\item{df}{data frame with the indicators to be styled}

\item{wb}{\code{\link[openxlsx:createWorkbook]{openxlsx::createWorkbook()}} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{start_row}{integer start row of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}

\item{end_row}{integer identifying end row.}

\item{start_col}{integer start col of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}

\item{end_col}{integer identifying end column.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
Used in \code{write_indicator_list_sheet()} to style the table.
}
