% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_notes.R
\name{style_notes}
\alias{style_notes}
\title{Styles data worksheet notes}
\usage{
style_notes(wb, sheet_name, bounds)
}
\arguments{
\item{wb}{\code{\link[openxlsx:createWorkbook]{openxlsx::createWorkbook()}} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{bounds}{named integer vector identifying integer sheet start and end reference column.
All parameters must be numerical. For converting Excel column references to integer references
see \code{\link[openxlsx:int2col]{openxlsx::int2col()}}.
Must have the following named elements:
\itemize{
\item \code{start_col}: start column
\item \code{end_col}: end column
\item \code{start_row}: start row
\item \code{end_row}: end row
}}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
Used within \code{write_indicators_hpop_summary()}
}
