% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_hpop_summary_sheet.R
\name{style_notes_data}
\alias{style_notes_data}
\title{Styles data worksheet notes}
\usage{
style_notes_data(df, wb, sheet_name, start_row, start_col, nrow_notes, end_col)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{\code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{start_row}{numeric start row of the table in the excel sheet}

\item{start_col}{numeric start row of the table in the excel sheet}

\item{nrow_notes}{integer with number of rows in the notes data frame being
styled}

\item{end_col}{integer indicating where the columns should stop being merged}
}
\description{
Applies styles to the data worksheet for country summary outputs
}
