% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_uhc_summary_sheet.R
\name{style_summary_box_uhc_summary}
\alias{style_summary_box_uhc_summary}
\title{Style UHC Summary box in summary sheet}
\usage{
style_summary_box_uhc_summary(wb, sheet_name, boxes_bounds)
}
\arguments{
\item{wb}{a \code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{boxes_bounds}{named list of bounds for data frame boxes to be written in sheet.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{style_summary_box_uhc_summary()} styles the HEP summary box/section of the
UHC summary sheet. Used within \code{write_summary_box_uhc_summary()}.
}
