% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_style.R
\name{style_timeseries}
\alias{style_timeseries}
\title{Styles time series sheet}
\usage{
style_timeseries(
  df,
  wb,
  billion = c("hep", "hpop", "uhc"),
  sheet_name,
  start_row,
  start_col,
  ind,
  year,
  type_col,
  df_wide,
  ind_df,
  this_iso3 = NULL
)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{a \code{openxlsx} workbook to be edited.}

\item{billion}{Billion to be used for billion styling when relevant: either
"hep", "hpop", or "uhc" when no billion to be applied.}

\item{sheet_name}{character name of the sheet to update}

\item{start_row}{integer start row of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}

\item{start_col}{integer start col of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}

\item{ind}{Column name of column with indicator names.}

\item{year}{Column name of column with years.}

\item{type_col}{Column name of column to store type.}

\item{df_wide}{wide version of \code{df} generated by \code{write_uhc_timeseries_sheet}}

\item{ind_df}{data frame containing information on indicators}

\item{this_iso3}{character iditenfying the country being styled.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{style_header_hpop_summary_sheet} styles the title and sub-title of the worksheet header.
}
