% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_country.R
\name{summarise_HPOP_cntry_data}
\alias{summarise_HPOP_cntry_data}
\title{Summarise Excel sheet 'data' for HPOP country summary}
\usage{
summarise_HPOP_cntry_data(
  df,
  iso3,
  start_year = 2018,
  end_year = 2019:2023,
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{iso3}{ISO3 codes of country to export.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2023.}

\item{...}{Additional arguments passed to \code{calculate_hpop_contributions} and
\code{calculate_hpop_billion}. Use if you need to provide additional parameters to
those functions.}
}
\value{
list of data frames to be used in the 'data' sheet of HPOP country summary
export
}
\description{
\code{summarise_HPOP_cntry_data} summarises the 'data' sgeet for the HPOP country summary.
}
