% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_country.R
\name{summarize_hpop_country_data}
\alias{summarize_hpop_country_data}
\title{Summarize Excel sheet 'data' for HPOP country summary}
\usage{
summarize_hpop_country_data(
  df,
  iso,
  year = "year",
  iso3 = "iso3",
  ind = "ind",
  value = "value",
  transform_glue = "transform_{value}",
  type = "type",
  source = "source",
  ind_ids = billion_ind_codes("hpop"),
  start_year = 2018,
  end_year = 2019:2023,
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{iso}{ISO3 codes of country to export.}

\item{year}{Column name of column with years.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{ind}{Column name of column with indicator names.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2023.}

\item{...}{Additional arguments passed to \code{calculate_hpop_contributions} and
\code{calculate_hpop_billion}. Use if you need to provide additional parameters to
those functions.}
}
\value{
list of data frames to be used in the 'data' sheet of HPOP country summary
export
}
\description{
\code{summarize_hpop_country_data} summarises the 'data' sgeet for the HPOP country summary.
}
