% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_alcohol}
\alias{transform_alcohol}
\title{Transform alcohol data}
\usage{
transform_alcohol(x)
}
\arguments{
\item{x}{Indicator value.}
}
\description{
Rescales alcohol data according to Billions methods report
}
