% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_glucose}
\alias{transform_glucose}
\title{Transform FPG data}
\usage{
transform_glucose(x)
}
\arguments{
\item{x}{Indicator value.}
}
\description{
Rescales FPG data from \verb{[7.4,5.1]} to \verb{[0,100]}
}
