% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hep.R
\name{transform_hep_data}
\alias{transform_hep_data}
\title{Transform Raw Indicator Values for HEP Billion}
\usage{
transform_hep_data(
  df,
  iso3 = "iso3",
  year = "year",
  ind = "ind",
  value = "value",
  type = "type",
  ind_ids = billion_ind_codes("hep"),
  cholera_latest_year = NULL,
  meningitis_latest_year = NULL,
  yellow_fever_latest_year = NULL,
  extrapolate_to = 2025
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{year}{Column name of column with year data.}

\item{ind}{Column name of column with indicator names.}

\item{value}{Column name of column with indicator values.}

\item{type}{Column name of column with type data.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{cholera_latest_year}{Latest year to calculate rolling average for cholera.
If \code{NULL}, uses the latest year with observations in the data.}

\item{meningitis_latest_year}{Latest year to calculate rolling average for meningitis
If \code{NULL}, uses the latest year with observations in the data.}

\item{yellow_fever_latest_year}{Latest year to calculate rolling average for meningitis
If \code{NULL}, uses the latest year with observations in the data.}

\item{extrapolate_to}{Year to extrapolate Prevent data to, defaults to 2025}
}
\value{
Data frame in long format.
}
\description{
\code{transform_hep_data()} applies transformations on HEP Billion indicators so
that transformed indicator values can be used within Billions calculations.
Details on the specific transformations applied can be found within the
Billions methods report.
}
\details{
Currently, this function only changes Prevent campaign data by calculating
the total sum of campaigns for each year for use in Prevent calculations.
For more details on the HEP Billion calculation process and how this function
ties in with the rest, see the vignette:

\href{../doc/hpop.html}{\code{vignette("hep", package = "billionaiRe")}}
}
