% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_hosp_beds}
\alias{transform_hosp_beds}
\title{Transform hospital beds data}
\usage{
transform_hosp_beds(x)
}
\arguments{
\item{x}{Indicator value.}
}
\description{
Rescales hospital beds according to Billions methods report
}
