% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_inversion}
\alias{transform_inversion}
\title{Reverse indicator and cap}
\usage{
transform_inversion(x)
}
\arguments{
\item{x}{Indicator value.}
}
\description{
Reverse indicator and capbetween 0 and 100
}
