% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_road_safety}
\alias{transform_road_safety}
\title{Transform road safety data}
\usage{
transform_road_safety(x, iso3)
}
\description{
Rescales road safety data using the SDI ratio according to the Billions methods report
}
