% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_suicide_rate}
\alias{transform_suicide_rate}
\title{Transform suicide rate data}
\usage{
transform_suicide_rate(x)
}
\arguments{
\item{x}{Indicator value.}
}
\description{
Rescales suicide rate data according to the Billions methods report
}
