% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_totl_pop}
\alias{transform_totl_pop}
\title{Combine data from two sexes into one}
\usage{
transform_totl_pop(x_mle, x_fmle, pop_male, pop_fmle)
}
\description{
Uses for if an indicator only has data produced for both sexes
}
