% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_transfats}
\alias{transform_transfats}
\title{Transform transfats policy data}
\usage{
transform_transfats(x)
}
\description{
Rescales transfats policy data according to the Billions methods report
}
