% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_uhc.R
\name{transform_uhc_data}
\alias{transform_uhc_data}
\title{Transform Raw Indicator Values for UHC Billion}
\usage{
transform_uhc_data(
  df,
  value_col = "value",
  transform_glue = "transform_{value_col}",
  ind_ids = billion_ind_codes("uhc"),
  recycle = FALSE,
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{value_col}{Column name of column with indicator values.}

\item{transform_glue}{Glue expression to be passed to \code{\link[glue:glue]{glue::glue()}}. Defaults to
\code{'transform_{value_col}'} which will create new column names by prefixing \code{transform_}
to the original name.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{recycle}{Boolean to indicate if data should be recycled}

\item{...}{additional parameters to to pass to \code{recycle_data}}
}
\value{
Data frame in long format.
}
\description{
\code{transform_uhc_data()} applies transformations on UHC Billion indicators so
that transformed indicator values can be used within Billions calculations.
Details on the specific transformations applied can be found within the
Billions methods report. Values in the transform column, if it already exists,
are replaced for UHC indicators that have data in the value column, otherwise
the column keeps its original data.
}
\details{
For more details on the UHC Billion calculation process and how this function
ties in with the rest, see the vignette:

\href{../doc/uhc.html}{\code{vignette("uhc", package = "billionaiRe")}}
}
\seealso{
Other uhc: 
\code{\link{calculate_uhc_billion}()},
\code{\link{calculate_uhc_contribution}()},
\code{\link{transform_uhc_single}()},
\code{\link{uhc_df}},
\code{\link{untransform_uhc_data}()},
\code{\link{untransform_uhc_single}()}
}
\concept{uhc}
