% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_uhc.R
\name{transform_uhc_single}
\alias{transform_uhc_single}
\title{Perform a transformation on a single column}
\usage{
transform_uhc_single(df, value_col, transform_value_col, ind_ids)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{value_col}{Column name of column with indicator values.}

\item{transform_value_col}{Column name of column(s) with transformed
indicator values}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
A single column data frame of transformed values.
}
\description{
This function is used within \code{\link[=transform_uhc_data]{transform_uhc_data()}} to generate transformed data
on a single column.
}
