% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_uhc_tobacco}
\alias{transform_uhc_tobacco}
\title{Transform UHC tobacco data}
\usage{
transform_uhc_tobacco(x)
}
\arguments{
\item{x}{Indicator value.}
}
\description{
Reverses the indicator and rescales from \verb{[29,100]} to \verb{[0,100]}
}
