% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{trim_clean_fuels}
\alias{trim_clean_fuels}
\title{Trim clean fuels data}
\usage{
trim_clean_fuels(x)
}
\arguments{
\item{x}{Indicator value.}
}
\description{
Trims clean fuels data to be between 5 and 95.
}
