% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{trim_transforms}
\alias{trim_transforms}
\title{Trim all transformed data}
\usage{
trim_transforms(x, min_val = 0, max_val = 100)
}
\arguments{
\item{x}{Indicator value.}

\item{min}{Minimum possible value for transform to take}

\item{max}{Maximum possible value for transform to take}
}
\description{
Trims transform data to ensure values are between 0 to 100
}
