% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scenarios.R
\name{trim_years}
\alias{trim_years}
\title{Trim values}
\usage{
trim_years(df, trim_years, start_year, end_year)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{trim_years}{logical to indicate if years before \code{start_year} and after
\code{end_year} should be removed}

\item{start_year}{Start year for scenario, defaults to 2018.}

\item{end_year}{End year for scenario, defaults to 2025}
}
\description{
Trim values
}
