% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_benchmarking.R
\name{benchmarking_anc4}
\alias{benchmarking_anc4}
\alias{benchmarking_art}
\alias{benchmarking_beds}
\alias{benchmarking_bp}
\alias{benchmarking_doctors}
\alias{benchmarking_nurses}
\alias{benchmarking_hwf}
\alias{benchmarking_dtp3}
\alias{benchmarking_fh}
\alias{benchmarking_fp}
\alias{benchmarking_fpg}
\alias{benchmarking_itn}
\alias{benchmarking_pneumo}
\alias{benchmarking_tb}
\alias{benchmarking_uhc_sanitation}
\alias{benchmarking_uhc_tobacco}
\alias{benchmarking_espar}
\title{Benchmarking scenarios}
\usage{
benchmarking_anc4(df, ...)

benchmarking_art(df, ...)

benchmarking_beds(df, ...)

benchmarking_bp(df, ...)

benchmarking_doctors(df, ...)

benchmarking_nurses(df, ...)

benchmarking_hwf(df, ...)

benchmarking_dtp3(df, ...)

benchmarking_fh(df, ...)

benchmarking_fp(df, ...)

benchmarking_fpg(df, start_year = 2018, ...)

benchmarking_itn(df, ...)

benchmarking_pneumo(df, ...)

benchmarking_tb(df, ...)

benchmarking_uhc_sanitation(df, ...)

benchmarking_uhc_tobacco(df, ...)

benchmarking_espar(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}
}
\value{
data frame with acceleration scenario binded to \code{df}. \code{scenario} is
set to \code{acceleration}
}
\description{
UHC benchmarking scenarios are generated by running 3 versions of
\code{scenario_top_n_iso3()}:
\itemize{
\item Top 20\% AROC globally
\item Top 20\% AROC in WHO regions
\item Top 20\% AROC in World Bank (WB) Income Groups (IC)
}

Some functions (see below) set additional parameters to guarantee comparability
with other scenarios
}
\section{Functions}{
\itemize{
\item \code{benchmarking_beds()}: Set a \code{upper_limit} of 10000

\item \code{benchmarking_doctors()}: Set a \code{upper_limit} of 10000

\item \code{benchmarking_hwf()}: Set \code{baseline_year} to 2012 and \code{aroc_end_year} to 2017

\item \code{benchmarking_fpg()}: Force \code{start_year} to 2018

}}
\concept{benchmarking}
