% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uhc_df}
\alias{uhc_df}
\title{UHC example data}
\format{
A data frame with 30 rows and 5 variables:
\describe{
\item{iso3}{Country ISO3 codes.}
\item{year}{Year.}
\item{ind}{UHC indicator code.}
\item{value}{Raw indicator value.}
\item{type}{Data type.}
}
}
\usage{
uhc_df
}
\description{
Fake UHC data used to test the Billions calculations code within the billionaiRe
package.
}
\details{
See the UHC vignette for its example use:

\href{../doc/uhc.html}{\code{vignette("uhc", package = "billionaiRe")}}
}
\seealso{
Other uhc: 
\code{\link{calculate_uhc_billion}()},
\code{\link{calculate_uhc_contribution}()},
\code{\link{transform_uhc_data}()},
\code{\link{transform_uhc_single}()},
\code{\link{untransform_uhc_data}()},
\code{\link{untransform_uhc_single}()}
}
\concept{uhc}
\keyword{datasets}
