% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{untransform_alcohol}
\alias{untransform_alcohol}
\title{Untransform alcohol data}
\usage{
untransform_alcohol(x)
}
\arguments{
\item{x}{Indicator value.}
}
\description{
Untransforms alcohol data according to Billions methods report
}
