% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{untransform_bp}
\alias{untransform_bp}
\title{Untransform blood pressure data}
\usage{
untransform_bp(x)
}
\arguments{
\item{x}{Indicator value.}
}
\description{
Reverses the transformation of \code{\link[=transform_bp]{transform_bp()}}
}
