% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{untransform_glucose}
\alias{untransform_glucose}
\title{Unransform FPG data}
\usage{
untransform_glucose(x)
}
\arguments{
\item{x}{Indicator value.}
}
\description{
Reverses transformation from \code{\link[=transform_glucose]{transform_glucose()}}
}
