% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{untransform_hosp_beds}
\alias{untransform_hosp_beds}
\title{Untransform hospital beds data}
\usage{
untransform_hosp_beds(x)
}
\arguments{
\item{x}{Indicator value.}
}
\description{
Reverses transformation from \code{\link[=transform_hosp_beds]{transform_hosp_beds()}}
}
