% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpop.R
\name{untransform_hpop_data}
\alias{untransform_hpop_data}
\title{Untransform Indicator Values for HPOP Billion}
\usage{
untransform_hpop_data(
  df,
  iso3 = "iso3",
  ind = "ind",
  transform_value = "transform_value",
  value = "value",
  ind_ids = billion_ind_codes("hpop")
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{ind}{Column name of column with indicator names.}

\item{transform_value}{Column name of column with transformed values to retrieve.}

\item{value}{Column name of column to place untransformed values. If the column already
exists, values are overwritten wherever \code{ind} and \code{transform_value} are available
to be untransformed for this Billion, but otherwise, the column retains its
other values.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
Data frame in long format.
}
\description{
\code{untransform_hpop_data()} reverses transformations on HPOP Billion indicators to
return raw indicator values. Details on the specific transformations applied
can be found within the Billions methods report.
}
\details{
For more details on the HPOP Billion calculation process and how this function
ties in with the rest, see the vignette:

\href{../doc/hpop.html}{\code{vignette("hpop", package = "billionaiRe")}}
}
