% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{untransform_hwf}
\alias{untransform_hwf}
\title{Untransform health workforce data}
\usage{
untransform_hwf(x)
}
\arguments{
\item{x}{Indicator value.}
}
\description{
Reverses transformation from \code{\link[=transform_hwf]{transform_hwf()}}
}
