% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{untransform_road_safety}
\alias{untransform_road_safety}
\title{Untransform road safety data}
\usage{
untransform_road_safety(x, iso3)
}
\arguments{
\item{x}{Indicator value.}

\item{iso3}{Country ISO3 code.}
}
\description{
Unscales road safety data using the SDI ratio according to the Billions methods report
}
