% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{untransform_suicide_rate}
\alias{untransform_suicide_rate}
\title{Untransform suicide rate date}
\usage{
untransform_suicide_rate(x)
}
\arguments{
\item{x}{Indicator value.}
}
\description{
Rescales transformed suicide rate date according to the Billions methods report
}
