% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{untransform_transfats}
\alias{untransform_transfats}
\title{Untransform transfats policy data}
\usage{
untransform_transfats(x)
}
\arguments{
\item{x}{Indicator value.}
}
\description{
Unscales transfats policy data according to the Billions methods report
}
