% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_uhc.R
\name{untransform_uhc_data}
\alias{untransform_uhc_data}
\title{Untransform Indicator Values for UHC Billion}
\usage{
untransform_uhc_data(
  df,
  transform_value_col = "transform_value",
  value_col = stringr::str_remove(transform_value_col, "transform_"),
  ind_ids = billion_ind_codes("uhc")
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{transform_value_col}{Column name(s) of column with transformed values to retrieve.}

\item{value_col}{Column name(s) of column to place untransformed values. Must be same
length as \code{transform_value_col}. If a column already exists, values are overwritten
wherever \code{ind} and \code{transform_value_col} are available to be untransformed for
this Billion, but otherwise, the column retains its
other values.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
Data frame in long format.
}
\description{
\code{untransform_uhc_data()} reverses transformations on UHC Billion indicators to
return raw indicator values. Details on the specific transformations applied
can be found within the Billions methods report.
}
\details{
For more details on the UHC Billion calculation process and how this function
ties in with the rest, see the vignette:

\href{../doc/uhc.html}{\code{vignette("uhc", package = "billionaiRe")}}
}
\seealso{
Other uhc: 
\code{\link{calculate_uhc_billion}()},
\code{\link{calculate_uhc_contribution}()},
\code{\link{transform_uhc_data}()},
\code{\link{transform_uhc_single}()},
\code{\link{uhc_df}},
\code{\link{untransform_uhc_single}()}
}
\concept{uhc}
