% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_uhc.R
\name{untransform_uhc_single}
\alias{untransform_uhc_single}
\title{Perform a transformation on a single column}
\usage{
untransform_uhc_single(df, ind, transform_value, value, ind_ids)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind}{Column name of column with indicator names.}

\item{transform_value}{Column name(s) of column with transformed values to retrieve.}

\item{value}{Column name(s) of column to place untransformed values. Must be same
length as \code{transform_value}. If a column already exists, values are overwritten
wherever \code{ind} and \code{transform_value} are available to be untransformed for
this Billion, but otherwise, the column retains its
other values.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
A single column data frame of transformed values.
}
\description{
This function is used within \code{\link[=untransform_uhc_data]{untransform_uhc_data()}} to generate transformed data
on a single column.
}
