% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_data.R
\name{upload_billion_data}
\alias{upload_billion_data}
\title{Upload Billions indicator data}
\usage{
upload_billion_data(
  df,
  data_type = c("wrangled_data", "projected_data", "final_data"),
  billion = c("hep", "hpop", "uhc"),
  ind_code,
  version = whdh::get_formatted_timestamp(),
  na_rm = TRUE,
  experiment = NULL,
  silent = TRUE
)
}
\arguments{
\item{df}{(data.frame) A data frame}

\item{data_type}{(string) The type of data to load.
\itemize{
\item \code{wrangled_data} (default): raw data that has been wrangled into a suitable
form for analysis.
\item \code{projected_data}:  data that has been fully projected to the target year but
has not yet been transformed or calculated upon.
\item \code{final_data}: the complete set of billions data with transformed values, contributions,
and all calculations available.
}}

\item{billion}{(string) One of "hep", "hpop", or "uhc". Ignored when
\code{data_type = "final_data"}.}

\item{ind_code}{(string) The name of the indicator to upload data
for. Ignored when \code{data_type = "final_data"}.}

\item{version}{A \code{yyyy-mm-ddTHH-MM-SS} formatted string. The default is the current
date time, as returned by \code{\link[whdh:get_formatted_timestamp]{whdh::get_formatted_timestamp()}}.}

\item{na_rm}{(logical) Specifies whether to remove rows where \code{value} is missing.
Defaults to \code{FALSE}.}

\item{experiment}{(string) Either \code{NULL} or a string ("unofficial" by default).
Identifies where the Bronze/Silver/Gold data layers to which data is uploaded
are located.
\itemize{
\item If \code{NULL}, the root folder for the data layers is the 3B folder (i.e., where
the "official" data is stored). For example, \verb{3B/Silver/...}.
\item If a string, the root folder for the data layers is a sub-folder within the
Sandbox layer of the 3B data lake (e.g., if \code{experiment = "my_exp"}, then
data is download from \verb{3B/Sandbox/my_exp/\{data_layer\}/...})
\item If an empty string, the root folder for the data layers is the Sandbox itself
(i.e., if \code{experiment = ""}, then data is download from \verb{3B/Sandbox/\{data_layer\}/...})
}}

\item{silent}{(logical) Specifies whether to show authentication messages and
a progress bar. Defaults to \code{TRUE}.}
}
\value{
The input data frame without any modifications. Note that a copy of the
input data frame is created for the upload and any modifications required by
the function all (such as setting \code{na_rm = TRUE}) are applied to only to this
copy. This means that the output data frame is identical to the input, even
if the uploaded data is modified in some way.
}
\description{
\code{upload_billion_data()} allows you to easily upload a data frame to the correct
location in the World Health Data Hub. By default, it appends the current
timestamp to the file name, but it can also accept a timestamp specified by the
user.
}
\details{
This function requires that the user to have the
\href{https://github.com/WorldHealthOrganization/whdh}{whdh} package installed
and setup to access the data. For quetions about getting the relevant permissions,
please contact \href{mailto:kanjim@who.int}{kanjim@who.int} or \href{mailto:messeillere@who.int}{messeillere@who.int}.
}
