% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_year.R
\name{valid_year}
\alias{valid_year}
\title{Check if year value is valid and controls for infinites}
\usage{
valid_year(x)
}
\arguments{
\item{x}{Year}
}
\description{
Check if year value is valid and controls for infinites
}
