% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_styles_modify.R
\name{validateColour}
\alias{validateColour}
\title{Validate colour inputed}
\usage{
validateColour(colour, errorMsg = "Invalid colour")
}
\arguments{
\item{colour}{colour as \code{colours()}}

\item{errorMsg}{error message to be passed.}
}
\description{
Validate colour inputed
}
\author{
Philipp Schauberger
}
