% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{vec2emptyDF}
\alias{vec2emptyDF}
\title{Create empty (no rows) data frame from a character vector}
\usage{
vec2emptyDF(vec)
}
\arguments{
\item{vec}{character vector}
}
\description{
Used in openxlsx-based function to go around limitation of writing only data
frames and thus in long. This allows to write in long.
}
