% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{warning_col_missing_values}
\alias{warning_col_missing_values}
\title{Warn user when any/all of the row are missing values for the the specified column}
\usage{
warning_col_missing_values(df, col_name, how)
}
\arguments{
\item{df}{Input data frame}

\item{col_name}{string specifying the name of column}

\item{how}{string specifying whether to check for any/all missing values}
}
\description{
Warn user when any/all of the row are missing values for the the specified column
}
