% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data.R
\name{wrangle_gho_data}
\alias{wrangle_gho_data}
\title{Put GHO data into billionaiRe format}
\usage{
wrangle_gho_data(df, source = NULL, type = NULL, ind = NULL)
}
\arguments{
\item{source}{Character string of source to be provided to the data frame if
the source is not already available in the GHO. If \code{NULL}, the source
column is generated from the GHO's \code{DataSourceDim} column.}

\item{type}{Character string of type to be provided to the data frame. If
\code{NULL}, the type column is just filled with \code{NA_character_}.}

\item{ind}{Character string of the indicator to be provided to the data frame. If
\code{NULL}, the indicator is determined by applying the \code{convert_ind_codes} function
on the \code{IndicatorCode} field of the GHO data.}
}
\value{
A data frame.
}
\description{

}
