% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data.R
\name{wrangle_rural_urban_gho_data}
\alias{wrangle_rural_urban_gho_data}
\title{Wrangle GHO data with TOTL/RUR/URB dimensions}
\usage{
wrangle_rural_urban_gho_data(
  df,
  source = NULL,
  type = NULL,
  ind = NULL,
  id_cols = c("SpatialDim", "TimeDim"),
  names_from = "Dim1",
  values_from = c("NumericValue", "High", "Low", "DataSourceDim", "Comments")
)
}
\arguments{
\item{df}{A data frame in GHO format, returned from \code{\link[ghost:gho_data]{ghost::gho_data()}}.}

\item{source}{Character string of source to be provided to the data frame.
If \code{NULL}, the source column is generated from the GHO's \code{DataSourceDim} column.
If not \code{NULL}, it overrides the source provided by the GHO.}

\item{type}{Character string of type to be provided to the data frame. If
\code{NULL}, the type column is filled with \code{NA_character_}.}

\item{ind}{Character string of the indicator name to be provided to the data frame.
This is a required argument and will raise an error if not provided.
If only a RUR or URB values is available, the indicator name has \verb{_rural} or \verb{_urban}
appended to it in the output data frame.}

\item{id_cols}{Character vector of the columns that are the same regardless of
the TOTL/RUR/URB dimension. Used as the argument of the same name in \code{pivot_wider}.}

\item{names_from, values_from}{A pair of character vectors used as the arguments
of the same name in \code{pivot_wider}.}
}
\value{
A data frame
}
\description{
Wrangle GHO data with TOTL/RUR/URB dimensions
}
\section{TODO}{

\itemize{
\item Convert to more generic \code{unspool_gho_dim} function which can
work with any other \code{DimType}, and not just TOTL/RUR/URB
\item Re-write to make it work better with \code{wrangle_gho_data} to avoid the significant amount
of redundant logic. This means that eventually, users may do
\code{gho_data(.) \%>\% unspool_gho_dim(.) \%>\% wrangle_gho_data(.)}
}
}

