% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data.R
\name{wrangle_unsd_data}
\alias{wrangle_unsd_data}
\title{Put UNSD data into billionaiRe format}
\usage{
wrangle_unsd_data(df, source = NULL, type = NULL)
}
\arguments{
\item{source}{Character string of source to be provided to the data frame if
the source is not already available in the UNSD database. If \code{NULL},
the source column is generated from the UNSD's \code{Source} column.}

\item{type}{Character string of type to be provided to the data frame. If
\code{NULL}, the type column is generated from the UNSD's \code{Nature} column.
"C" and "CA" are turned to "reported", while "E" and "M" are "estimated".}
}
\value{
A data frame.
}
\description{

}
