% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataframes_hpop_summary.R
\name{write_baseline_projection_hpop_summary}
\alias{write_baseline_projection_hpop_summary}
\title{Write the baseline/projection data frame to the data sheet}
\usage{
write_baseline_projection_hpop_summary(
  df,
  wb,
  sheet_name,
  ind_df,
  year,
  start_year,
  end_year,
  ind,
  value,
  transform_value,
  type_col,
  source_col,
  iso3,
  bounds
)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{\code{\link[openxlsx:createWorkbook]{openxlsx::createWorkbook()}} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{ind_df}{data frame containing the indicators in the correct order and format to be used.}

\item{year}{Column name of column with years.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2023.}

\item{ind}{Column name of column with indicator names.}

\item{value}{Column name of column with indicator values.}

\item{transform_value}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{type_col}{Column name of column to store type.}

\item{source_col}{Column name of column to store source information.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{bounds}{named integer vector identifying integer sheet start and end reference column.
All parameters must be numerical. For converting Excel column references to integer references
see \code{\link[openxlsx:int2col]{openxlsx::int2col()}}.
Must have the following named elements:
\itemize{
\item \code{start_col}: start column
\item \code{end_col}: end column
\item \code{start_row}: start row
\item \code{end_row}: end row
}}
}
\description{
Write the baseline/projection data frame to the data sheet
}
