% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataframes_hpop_summary.R
\name{write_bilion_contrib_ind_hpop_summary}
\alias{write_bilion_contrib_ind_hpop_summary}
\title{Write the contribution to billion per indicator data frame to the data sheet}
\usage{
write_bilion_contrib_ind_hpop_summary(
  df,
  wb,
  sheet_name,
  year,
  start_year,
  end_year,
  ind,
  contribution_pct,
  population,
  contribution,
  contribution_pct_total_pop,
  ind_df,
  bounds
)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{a \code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{year}{Column name of column with years.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{ind}{Column name of column with indicator names.}

\item{contribution_pct}{Column name of column(s) to store contribution (percent)
values. Must be the same length as \code{transform_value}.}

\item{population}{Column name of column with population figures.}

\item{contribution}{Column name of column(s) to store contribution (population)
values. Must be the same length as \code{transform_value}.}

\item{contribution_pct_total_pop}{Column name of column(s) to store contribution
(percent of total population of the country) values. Must be the same length
as \code{transform_value}.}

\item{ind_df}{data frame containing the indicators in the correct order and format to be used.}

\item{bounds}{named integer vector identifying integer sheet start and end reference column.
All parameters must be numerical. For converting Excel column references to integer references
see \code{\link[openxlsx:int2col]{openxlsx::int2col()}}.
Must have the following named elements:
\itemize{
\item \code{start_col}: start column
\item \code{end_col}: end column
\item \code{start_row}: start row
\item \code{end_row}: end row
}}
}
\description{
Used within \code{write_hpop_summary_sheet()}
}
