% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataframes_uhc_summary.R
\name{write_data_boxes_uhc_summary}
\alias{write_data_boxes_uhc_summary}
\title{Write and style RMNCH box in UHC summary data}
\usage{
write_data_boxes_uhc_summary(
  df,
  pillar = c("RMNCH", "infec_diseases", "ncd", "service_cap_access"),
  wb,
  sheet_name,
  value,
  transform_value,
  boxes_bounds,
  start_year,
  end_year,
  ind,
  ind_df,
  year,
  type_col,
  source_col,
  iso3
)
}
\arguments{
\item{df}{Data frame in long format filtered for a specific country, where 1 row corresponds
to a specific year, and indicator.}

\item{pillar}{character Pillar identifying the pillar to style. Must be one of
\itemize{
\item RMNCH
\item infectious
\item NCD
\item service
}}

\item{wb}{\code{\link[openxlsx:createWorkbook]{openxlsx::createWorkbook()}} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{value}{Column name of column with indicator values.}

\item{transform_value}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{boxes_bounds}{named list of bounds for data frame boxes to be written in sheet.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2023.}

\item{ind}{Column name of column with indicator names.}

\item{ind_df}{data frame containing the indicators in the correct order and format to be used.}

\item{year}{Column name of column with years.}

\item{type_col}{Column name of column to store type.}

\item{source_col}{Column name of column to store source information.}

\item{iso3}{Column name of column with country ISO3 codes.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{write_data_boxes_uhc_summary} writes and styles the RMNCH box in UHC summary worksheet.
Used in \code{write_uhc_summary_sheet()}.
}
