% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataframes_uhc_summary.R
\name{write_data_headers_uhc_summary}
\alias{write_data_headers_uhc_summary}
\title{Write and style headers for data in UHC summary sheets}
\usage{
write_data_headers_uhc_summary(
  wb,
  sheet_name,
  value,
  boxes_bounds,
  start_year,
  end_year
)
}
\description{
Used in \code{write_uhc_summary_sheet()}
}
