% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data_sheet.R
\name{write_data_sheet_HPOP}
\alias{write_data_sheet_HPOP}
\title{Write data sheet}
\usage{
write_data_sheet_HPOP(
  df,
  wb,
  sheet_name,
  start_year,
  end_year,
  value,
  year,
  iso3,
  iso,
  ind,
  population,
  scenario,
  ind_ids,
  transform_value,
  type_col,
  source_col,
  contribution
)
}
\description{
Write data sheet
}
