% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{write_empty_white_data}
\alias{write_empty_white_data}
\title{Write empty and white data on the maximum extend of bounds}
\usage{
write_empty_white_data(wb, sheet_name, bounds)
}
\arguments{
\item{wb}{a \code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{bounds}{a named list with the extent of the data to be written over.
Must contain the following named parameters:
\itemize{
\item start_row
\item end_row
\item start_col
\item end_col
}}
}
\value{
\code{openxslx} Workbook object
}
\description{
\code{write_empty_white_data} writes a matrix of the size of the maximum values
found in \code{bounds} (plus 4 column/rows as a margin)
}
