% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_hep_timeseries.R
\name{write_hep_timeseries_sheet}
\alias{write_hep_timeseries_sheet}
\title{Write HEP times series sheet}
\usage{
write_hep_timeseries_sheet(
  df,
  wb,
  sheet_name,
  start_row,
  start_col,
  transform_value,
  ind_df,
  ind,
  year,
  type_col,
  ind_ids,
  end_year,
  iso3
)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{a \code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{start_row}{integer start row of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}

\item{start_col}{integer start col of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}

\item{transform_value}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{ind_df}{data frame containing information on indicators}

\item{ind}{Column name of column with indicator names.}

\item{year}{Column name of column with years.}

\item{type_col}{Column name of column to store type.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{iso3}{Column name of column with country ISO3 codes.}
}
\description{
\code{export_hpop_country_summary_xls} Export a country-specific for HPOP billion.
}
