% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_hpop_datasheet.R
\name{write_hpop_billion_contrib}
\alias{write_hpop_billion_contrib}
\title{Write the HPOP billion contribution box}
\usage{
write_hpop_billion_contrib(
  df,
  wb,
  value,
  start_row,
  start_col,
  start_year,
  end_year,
  sheet_name
)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{\code{openxlsx} workbook to be edited.}

\item{value}{Column name of column with indicator values.}

\item{start_row}{numeric start row of the table in the excel sheet}

\item{start_col}{numeric start row of the table in the excel sheet}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2023.}

\item{sheet_name}{character name of the sheet to update}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
Write the HPOP billion contribution box
}
