% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_hpop_datasheet.R
\name{write_hpop_datasheet}
\alias{write_hpop_datasheet}
\title{Write data sheet}
\usage{
write_hpop_datasheet(
  df,
  wb,
  sheet_name,
  start_year,
  end_year,
  value,
  year,
  iso3,
  iso,
  ind,
  population,
  scenario,
  ind_ids,
  transform_value,
  type_col,
  source_col,
  contribution
)
}
\description{
Write data sheet
}
