% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_hpop_inter.R
\name{write_hpop_inter}
\alias{write_hpop_inter}
\title{Write Inter sheet for HPOP}
\usage{
write_hpop_inter(
  wb,
  sheet_name,
  data_sheet_name,
  ind_df,
  start_year,
  end_year,
  start_col,
  start_row,
  start_col_data = 1,
  start_row_data = 9,
  transform_value
)
}
\arguments{
\item{wb}{\code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{data_sheet_name}{character Name of the data sheet}

\item{ind_df}{data frame containing information on indicators}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2023.}

\item{start_col}{numeric start row of the table in the excel sheet}

\item{start_row}{numeric start row of the table in the excel sheet}

\item{start_col_data}{integer Start column of the data frame in the data sheet}

\item{start_row_data}{integer Start row of the data frame in the data sheet}

\item{transform_value}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{export_hpop_country_summary_xls} Export a country-specific for HPOP billion.
}
