% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_indicator_list.R
\name{write_indicator_list_sheet}
\alias{write_indicator_list_sheet}
\title{Write indicator list sheet}
\usage{
write_indicator_list_sheet(df, wb, sheet_name, start_row, start_col)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{\code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{start_row}{numeric start row of the table in the excel sheet}

\item{start_col}{numeric start row of the table in the excel sheet}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
Write indicator list sheet
}
