% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataframes_hpop_summary.R
\name{write_indicators_hpop_summary}
\alias{write_indicators_hpop_summary}
\title{Write indicator list for HPOP summary sheet}
\usage{
write_indicators_hpop_summary(ind_df, wb, sheet_name, bounds)
}
\arguments{
\item{ind_df}{data frame containing the indicators in the correct order and format to be used.}

\item{wb}{\code{\link[openxlsx:createWorkbook]{openxlsx::createWorkbook()}} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{bounds}{named integer vector identifying integer sheet start and end reference column.
All parameters must be numerical. For converting Excel column references to integer references
see \code{\link[openxlsx:int2col]{openxlsx::int2col()}}.
Must have the following named elements:
\itemize{
\item \code{start_col}: start column
\item \code{end_col}: end column
\item \code{start_row}: start row
\item \code{end_row}: end row
}}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{write_hpop_summary_sheet()} writes all the content and styling for the HPOP
summary sheet. Used within \code{export_hpop_country_summary_xls()}
}
