% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataframes_hpop_summary.R
\name{write_latest_reported_hpop_summary}
\alias{write_latest_reported_hpop_summary}
\title{Write latest reported data to HPOP summary sheet}
\usage{
write_latest_reported_hpop_summary(
  df,
  wb,
  sheet_name,
  ind_df,
  start_row,
  start_col,
  type_col,
  iso3,
  ind,
  year,
  value,
  transform_value,
  source_col,
  year_counts = c(2000, 2015),
  bounds
)
}
\arguments{
\item{df}{Data frame in long format filtered for a specific country, where 1 row corresponds
to a specific year, and indicator.}

\item{wb}{\code{\link[openxlsx:createWorkbook]{openxlsx::createWorkbook()}} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{ind_df}{data frame containing the indicators in the correct order and format to be used.}

\item{start_row}{integer start row of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}

\item{start_col}{integer start col of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}

\item{type_col}{Column name of column to store type.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{ind}{Column name of column with indicator names.}

\item{year}{Column name of column with years.}

\item{value}{Column name of column with indicator values.}

\item{transform_value}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{source_col}{Column name of column to store source information.}

\item{year_counts}{integer vector with the year(s) to count the number of values available for
a specific indicators. Defaults to 2000 and 2015.}

\item{bounds}{named integer vector identifying integer sheet start and end reference column.
All parameters must be numerical. For converting Excel column references to integer references
see \code{\link[openxlsx:int2col]{openxlsx::int2col()}}.
Must have the following named elements:
\itemize{
\item \code{start_col}: start column
\item \code{end_col}: end column
\item \code{start_row}: start row
\item \code{end_row}: end row
}}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{write_hpop_summary_sheet()} writes all the content and styling for the HPOP
summary sheet. Used within \code{export_hpop_country_summary_xls()}
}
