% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_main_data.R
\name{write_main_df}
\alias{write_main_df}
\title{Write the main data frame to the data sheet}
\usage{
write_main_df(
  df,
  wb,
  start_row,
  start_col,
  start_year,
  end_year,
  year,
  sheet_name,
  value,
  transform_value,
  type_col,
  source_col,
  contribution,
  contribution_pct,
  contribution_pct_pop_total,
  ind
)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{\code{openxlsx} workbook to be edited.}

\item{start_row}{numeric start row of the table in the excel sheet}

\item{start_col}{numeric start row of the table in the excel sheet}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2023.}

\item{year}{Column name of column with years.}

\item{sheet_name}{character name of the sheet to update}

\item{value}{Column name of column with indicator values.}

\item{transform_value}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{type_col}{Column name of column to store type.}

\item{source_col}{Column name of column to store source information.}

\item{contribution}{Column name of column(s) to store contribution (population)
values. Must be the same length as \code{transform_value}.}

\item{contribution_pct}{Column name of column(s) to store contribution (percent)
values. Must be the same length as \code{transform_value}.}

\item{contribution_pct_pop_total}{Column name of column(s) to store contribution
(percent of total population of the country) values. Must be the same length
as \code{transform_value}.}

\item{ind}{Column name of column with indicator names.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
Write the main data frame to the data sheet
}
