% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_notes.R
\name{write_notes}
\alias{write_notes}
\title{Write and style notes}
\usage{
write_notes(df, notes_title = "Notes:", wb, sheet_name, bounds)
}
\arguments{
\item{df}{data frame to be written}

\item{notes_title}{character string with the title (first line) of the notes.}

\item{wb}{\code{\link[openxlsx:createWorkbook]{openxlsx::createWorkbook()}} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{bounds}{named integer vector identifying integer sheet start and end reference column.
All parameters must be numerical. For converting Excel column references to integer references
see \code{\link[openxlsx:int2col]{openxlsx::int2col()}}.
Must have the following named elements:
\itemize{
\item \code{start_col}: start column
\item \code{end_col}: end column
\item \code{start_row}: start row
\item \code{end_row}: end row
}}
}
\description{
\code{write_notes()} writes notes/footnotes provided in \code{df} to the appropriate
sheet.
}
