% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_export.R
\name{write_notes_data}
\alias{write_notes_data}
\title{Write notes for data sheet}
\usage{
write_notes_data(data_table, wb, start_row, start_col, end_col, sheet_name)
}
\arguments{
\item{data_table}{data frame to be written}

\item{wb}{openxlsx workbook to be edited.}

\item{start_row}{numeric start row of the table in the excel sheet}

\item{start_col}{numeric start row of the table in the excel sheet}

\item{end_col}{integer indicating where the columns should stop being merged}

\item{sheet_name}{character name of the sheet to update}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_fldr}.
}
\description{
Write notes for data sheet
}
