% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_permanent_sheets.R
\name{write_permanent_sheets}
\alias{write_permanent_sheets}
\title{Write permanent sheets}
\usage{
write_permanent_sheets(billion, start_col, start_row)
}
\arguments{
\item{billion}{Billion indicator names to return, either "hep", "hpop", "uhc", or "all".}

\item{start_col}{integer start col of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}

\item{start_row}{integer start row of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}
}
\value{
A \code{\link[openxlsx:createWorkbook]{openxlsx::createWorkbook()}} object
}
\description{
\code{write_permanent_sheets()} writes and styles the Intro and Indicator List sheets
that are present in all cases taken into account by the different functions.
}
