% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataframes_uhc_summary.R
\name{write_RMNCH_uhc_summary}
\alias{write_RMNCH_uhc_summary}
\title{Write}
\usage{
write_RMNCH_uhc_summary(
  df,
  wb,
  sheet_name,
  value,
  transform_value,
  boxes_bounds,
  start_year,
  end_year,
  ind,
  ind_df,
  year,
  type_col,
  source_col,
  iso3
)
}
\description{
Write
}
