% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data_sheet.R
\name{write_sheet_header}
\alias{write_sheet_header}
\title{Write and style sheet header}
\usage{
write_sheet_header(wb, sheet_name, billion_long, iso, start_col, start_row)
}
\arguments{
\item{wb}{openxlsx workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{billion_long}{character Long name of the billion to be written}

\item{iso}{ISO3 code of country to summarize.}

\item{start_col}{numeric start row of the table in the excel sheet}

\item{start_row}{numeric start row of the table in the excel sheet}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{export_hpop_country_summary_xls} Export a country-specific for HPOP billion.
}
