% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_hep_summary_sheet.R
\name{write_sheet_header_hep_summary}
\alias{write_sheet_header_hep_summary}
\title{Write and style HEP summary sheet header}
\usage{
write_sheet_header_hep_summary(
  wb,
  sheet_name,
  iso,
  end_year,
  value,
  boxes_bounds
)
}
\arguments{
\item{wb}{a \code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{iso}{ISO3 code of country to summarize.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{value}{Column name of column with indicator values.}

\item{boxes_bounds}{named list of bounds for data frame boxes to be written in sheet.}
}
\description{
Write and style HEP summary sheet header
}
