% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_uhc_summary_sheet.R
\name{write_sheet_header_uhc_summary}
\alias{write_sheet_header_uhc_summary}
\title{Write and style UHC summary sheet header}
\usage{
write_sheet_header_uhc_summary(
  wb,
  sheet_name,
  iso,
  end_year,
  value,
  boxes_bounds
)
}
\arguments{
\item{wb}{\code{\link[openxlsx:createWorkbook]{openxlsx::createWorkbook()}} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{iso}{ISO3 code of country to summarize.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2023.}

\item{value}{Column name of column with indicator values.}

\item{boxes_bounds}{named list of bounds for data frame boxes to be written in sheet.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{export_hpop_country_summary_xls} Export a country-specific for HPOP billion.
}
