% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataframes_uhc_summary.R
\name{write_summary_box_uhc_summary}
\alias{write_summary_box_uhc_summary}
\title{Write and style summary box in UHC summary sheet}
\usage{
write_summary_box_uhc_summary(
  wb,
  sheet_name,
  iso,
  start_year,
  end_year,
  boxes_bounds
)
}
\arguments{
\item{wb}{a \code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{iso}{ISO3 code of country to summarize.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{boxes_bounds}{named list of bounds for data frame boxes to be written in sheet.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{write_summary_box_uhc_summary} writes and styles the summary box in UHC
summary worksheet. Used in \code{write_uhc_summary_sheet()}.
}
