% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_export.R
\name{write_timeseries_sheet}
\alias{write_timeseries_sheet}
\title{Write times series sheet}
\usage{
write_timeseries_sheet(
  df,
  wb,
  sheet_name,
  start_row,
  start_col,
  transform_value,
  df_ind,
  ind,
  year,
  type_col
)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{openxlsx workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{start_row}{numeric start row of the table in the excel sheet}

\item{start_col}{numeric start row of the table in the excel sheet}

\item{transform_value}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{df_ind}{data frame containing information on indicators}

\item{ind}{Column name of column with indicator names.}

\item{year}{Column name of column with years.}

\item{type_col}{Column name of column to store type.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_fldr}.
}
\description{
\code{export_hpop_country_summary_xls} Export a country-specific for HPOP billion.
}
