% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_uhc_summary_sheet.R
\name{write_uhc_summary_sheet}
\alias{write_uhc_summary_sheet}
\title{Write UHC summary sheet}
\usage{
write_uhc_summary_sheet(
  df,
  wb,
  sheet_name,
  iso,
  start_year = 2018,
  end_year = 2019:2023,
  value = "value",
  transform_value = "transform_value",
  year = "year",
  iso3 = "iso3",
  ind = "ind",
  population = "population",
  type_col = "type",
  source_col = "source",
  ind_df,
  ind_ids
)
}
\arguments{
\item{df}{Data frame in long format filtered for a specific country, where 1 row corresponds
to a specific year, and indicator.}

\item{iso}{ISO3 code of country to summarize.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2023.}

\item{value}{Column name of column with indicator values.}

\item{transform_value}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{year}{Column name of column with years.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{ind}{Column name of column with indicator names.}

\item{population}{Column name of column with population figures.}

\item{type_col}{Column name of column to store type.}

\item{source_col}{Column name of column to store source information.}

\item{ind_df}{data frame containing the indicators in the correct order and format to be used.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{write_uhc_summary_sheet()} writes all the content and styling for the UHC
summary sheet. Used within \code{export_uhc_country_summary_xls()}
}
