% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_uhc_timeseries.R
\name{write_uhc_timeseries_sheet}
\alias{write_uhc_timeseries_sheet}
\title{Write times series sheet}
\usage{
write_uhc_timeseries_sheet(
  df,
  wb,
  sheet_name,
  start_row,
  start_col,
  value,
  ind_df,
  ind,
  year,
  type_col
)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{\code{\link[openxlsx:createWorkbook]{openxlsx::createWorkbook()}} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{start_row}{integer start row of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}

\item{start_col}{integer start col of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}

\item{value}{Column name of column with indicator values.}

\item{ind_df}{data frame containing information on indicators}

\item{ind}{Column name of column with indicator names.}

\item{year}{Column name of column with years.}

\item{type_col}{Column name of column to store type.}
}
\description{
\code{export_hpop_country_summary_xls} Export a country-specific for HPOP billion.
}
