% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_wrangling.R
\name{xmart_col_types}
\alias{xmart_col_types}
\title{Get the column types for xMart columns}
\usage{
xmart_col_types(data_type = c("wrangled_data", "projected_data", "final_data"))
}
\arguments{
\item{data_type}{(string): the type of data
\itemize{
\item \code{wrangled_data} (default): raw data that has been wrangled into a suitable
form for analysis.
\item \code{projected_data}:  data that has been fully projected to the target year but
has not yet been transformed or calculated upon.
\item \code{final_data}: the complete set of billions data with transformed values,
contributions, and all calculations available.
}}
}
\value{
a named vector with column specifications
}
\description{
A helper function for specifying the expected column types for the Triple
Billions data.
}
