% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_uhc.R
\name{accelerate_anc4}
\alias{accelerate_anc4}
\title{Accelerate anc4}
\usage{
accelerate_anc4(
  df,
  ind_ids = billion_ind_codes("uhc"),
  scenario_col = "scenario",
  default_scenario = "default",
  bau_scenario = "historical",
  scenario_name = "acceleration",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{default_scenario}{name of the default scenario.}

\item{bau_scenario}{name of scenario to be used for business as usual.
Default is \code{historical}.}

\item{scenario_name}{name of scenario}

\item{...}{additional parameters to be passed to scenario function}
}
\value{
data frame with acceleration scenario binded to \code{df}. \code{scenario_col} is
set to \code{acceleration}
}
\description{
Accelerate anc4 by first dividing countries into those with reported data and
those without.
\itemize{
\item For countries without reported data, the acceleration scenario_col is the same
as business as usual.
\item For countries with reported data, scenarios with both a \strong{fixed target of 95\%
by 2030} and a \strong{applying the AROC of the top 10 performing countries with at
least 4 reported/estimated values} are tried, with the easiest to achieve of
the two selected. The selected scenario is then compared against the business
as usual scenario for reported data, and the best of the two chosen as the
acceleration scenario.
}
}
\seealso{
UHC acceleration scenarios
\code{\link{accelerate_art}()},
\code{\link{accelerate_beds}()},
\code{\link{accelerate_bp}()},
\code{\link{accelerate_doctors}()},
\code{\link{accelerate_dtp3}()},
\code{\link{accelerate_fh}()},
\code{\link{accelerate_fpg}()},
\code{\link{accelerate_fp}()},
\code{\link{accelerate_hwf}()},
\code{\link{accelerate_itn}()},
\code{\link{accelerate_nurses}()},
\code{\link{accelerate_pneumo}()},
\code{\link{accelerate_tb}()},
\code{\link{accelerate_uhc_sanitation}()},
\code{\link{accelerate_uhc_tobacco}()}
}
\concept{uhc_acceleration}
