% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_uhc.R
\name{accelerate_bp}
\alias{accelerate_bp}
\title{Accelerate bp}
\usage{
accelerate_bp(
  df,
  ind_ids = billion_ind_codes("uhc"),
  value_col = "value",
  scenario_col = "scenario",
  start_year = 2018,
  end_year = 2025,
  default_scenario = "default",
  bau_scenario = "historical",
  scenario_name = "acceleration",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{value_col}{Name of the column containing indicator value in \code{df}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{start_year}{Year from which the acceleration scenario begins, inclusive.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{default_scenario}{name of the default scenario.}

\item{bau_scenario}{name of scenario to be used for business as usual.
Default is \code{historical}.}

\item{scenario_name}{name of scenario}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate bp by taking the best of aiming at reaching 80\% by 2030 and
business as usual.
}
\seealso{
UHC acceleration scenarios
\code{\link{accelerate_anc4}()},
\code{\link{accelerate_art}()},
\code{\link{accelerate_beds}()},
\code{\link{accelerate_doctors}()},
\code{\link{accelerate_dtp3}()},
\code{\link{accelerate_fh}()},
\code{\link{accelerate_fpg}()},
\code{\link{accelerate_fp}()},
\code{\link{accelerate_hwf}()},
\code{\link{accelerate_itn}()},
\code{\link{accelerate_nurses}()},
\code{\link{accelerate_pneumo}()},
\code{\link{accelerate_tb}()},
\code{\link{accelerate_uhc_sanitation}()},
\code{\link{accelerate_uhc_tobacco}()}
}
\concept{uhc_acceleration}
