% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_hpop.R
\name{accelerate_child_obese}
\alias{accelerate_child_obese}
\title{Accelerate child_obese}
\usage{
accelerate_child_obese(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate child_obese by halting upwards trend in the data to the 2010 value.
}
