% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_hpop.R
\name{accelerate_hpop_sanitation_urban}
\alias{accelerate_hpop_sanitation_urban}
\title{Accelerate hpop_sanitation_urban}
\usage{
accelerate_hpop_sanitation_urban(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\value{
data frame with acceleration scenario binded to \code{df}. \code{scenario} is
set to \code{acceleration}
}
\description{
Accelerate hpop_sanitation_urban by aiming at best value in quintile.
}
\details{
Runs:
\itemize{
\item \verb{scenario_quantile(df, n = 5, quantile_year = 2017, baseline_quantile_year = 2018, upper_limit = 99, lower_limit = 0 small_is_best = FALSE,...)}
}
}
