% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_uhc.R
\name{accelerate_hwf}
\alias{accelerate_hwf}
\title{Accelerate hwf}
\usage{
accelerate_hwf(
  df,
  ind_ids = billion_ind_codes("uhc"),
  end_year = 2025,
  scenario = "scenario",
  ind = "ind",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{scenario}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{ind}{Column name of column with indicator names.}

\item{...}{additional parameters to be passed to scenario function}
}
\value{
data frame with acceleration scenario binded to \code{df}. \code{scenario} is
set to \code{acceleration}
}
\description{
Accelerate hwf by first dividing countries into two groups:
\itemize{
\item For countries with a 2018 value greater than or equal to the 2018 global median,
business as usual is returned.
\item For countries with a 2018 value less than the 2018 global median, a \strong{linear change
of 4.54 per year from 2018 to 2025} is returned.
}
}
