% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_hpop.R
\name{accelerate_ipv}
\alias{accelerate_ipv}
\title{Accelerate ipv}
\usage{
accelerate_ipv(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate ipv by targeting 0 by 2030.
}
\details{
Runs:
\itemize{
\item \code{scenario_fixed_target(df, target_value = 0, target_year = 2030, small_is_best = TRUE,...)},
}
}
\seealso{
HPOP acceleration scenarios
\code{\link{accelerate_adult_obese}()},
\code{\link{accelerate_alcohol}()},
\code{\link{accelerate_child_obese}()},
\code{\link{accelerate_child_viol}()},
\code{\link{accelerate_devontrack}()},
\code{\link{accelerate_fuel}()},
\code{\link{accelerate_hpop_sanitation_rural}()},
\code{\link{accelerate_hpop_sanitation_urban}()},
\code{\link{accelerate_hpop_sanitation}()},
\code{\link{accelerate_hpop_tobacco}()},
\code{\link{accelerate_overweight}()},
\code{\link{accelerate_pm25}()},
\code{\link{accelerate_road}()},
\code{\link{accelerate_stunting}()},
\code{\link{accelerate_suicide}()},
\code{\link{accelerate_transfats}()},
\code{\link{accelerate_wasting}()},
\code{\link{accelerate_water_rural}()},
\code{\link{accelerate_water_urban}()},
\code{\link{accelerate_water}()}
}
\concept{hpop_acceleration}
