% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_hpop.R
\name{accelerate_stunting}
\alias{accelerate_stunting}
\title{Accelerate stunting}
\usage{
accelerate_stunting(
  df,
  ind_ids = billion_ind_codes("hpop"),
  scenario_col = "scenario",
  default_scenario = "default",
  bau_scenario = "historical",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{default_scenario}{name of the default scenario.}

\item{bau_scenario}{name of scenario to be used for business as usual.
Default is \code{historical}.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate stunting by picking the best results between business as usual,
halt downwards trend, and AROC of -50\% change between 2012 and 2030.
}
\details{
Runs:
\itemize{
\item \code{scenario_bau(df, small_is_best = TRUE,...)},
\item \code{scenario_aroc(df, aroc_type = "percent_change", percent_change = -50, baseline_year = 2012, target_year = 2030, small_is_best = TRUE, ...)}
\item \code{scenario_halt_rise(df, small_is_best = TRUE,...)}
}

Then picks the best result between the three scenarios.
}
