% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_target_uhc.R
\name{accelerate_target_anc4}
\alias{accelerate_target_anc4}
\title{Accelerate anc4 to 95\% by 2030}
\usage{
accelerate_target_anc4(
  df,
  scenario_col = "scenario",
  default_scenario = "default",
  bau_scenario = "historical",
  scenario_name = "acceleration_target",
  ind_ids = billion_ind_codes("uhc"),
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{default_scenario}{name of the default scenario.}

\item{bau_scenario}{name of scenario to be used for business as usual.
Default is \code{historical}.}

\item{scenario_name}{name of scenario}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{...}{additional parameters to be passed to scenario function}
}
\value{
data frame with acceleration_target scenario binded to \code{df}. \code{scenario_name} is
set to \code{acceleration_target}
}
\description{
Accelerate \code{anc4} by aiming at \strong{fixed target of 95\% by 2030} for countries
with at least 2 reported values. For others, it is business as usual.
}
\seealso{
Target scenario
\code{\link{accelerate_target_art}()},
\code{\link{accelerate_target_beds}()},
\code{\link{accelerate_target_bp}()},
\code{\link{accelerate_target_doctors}()},
\code{\link{accelerate_target_dtp3}()},
\code{\link{accelerate_target_fh}()},
\code{\link{accelerate_target_fpg}()},
\code{\link{accelerate_target_fp}()},
\code{\link{accelerate_target_hwf}()},
\code{\link{accelerate_target_itn}()},
\code{\link{accelerate_target_nurses}()},
\code{\link{accelerate_target_pneumo}()},
\code{\link{accelerate_target_tb}()},
\code{\link{accelerate_target_uhc_sanitation}()},
\code{\link{accelerate_target_uhc_tobacco}()}
}
\concept{accelerate_target}
