% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_uhc_target.R
\name{accelerate_target_beds}
\alias{accelerate_target_beds}
\title{Accelerate beds}
\usage{
accelerate_target_beds(
  df,
  ind_ids = billion_ind_codes("uhc"),
  scenario_col = "scenario",
  value_col = "value",
  start_year = 2018,
  default_scenario = "default",
  bau_scenario = "historical",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{value_col}{Name of the column containing indicator value in \code{df}.}

\item{start_year}{Year from which the acceleration scenario begins, inclusive.}

\item{default_scenario}{name of the default scenario.}

\item{bau_scenario}{name of scenario to be used for business as usual.
Default is \code{historical}.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate beds by first dividing countries into two groups:
\itemize{
\item For countries with 18 or more beds for all years after 2018, business
as usual is returned.
\item For countries which have less than 18 beds for any of the years after 2018 (inclusive),
the best of business as usual and a \strong{linear change of 0.36 per year up to 2025},
with an upper limit of 18, is returned.
}
}
\details{
Runs:
\itemize{
\item \code{scenario_halt_rise(df, baseline_year = 2018, small_is_best = TRUE,...)},
\item \code{scenario_bau(df, small_is_best = TRUE,...)}, and
\item \code{scenario_percent_baseline(df, percent_change = -10, baseline_year = 2010, small_is_best = TRUE, ...)}
}

Then picks the best result between the three scenarios.
}
