% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_target_uhc.R
\name{accelerate_target_bp}
\alias{accelerate_target_bp}
\title{Accelerate bp}
\usage{
accelerate_target_bp(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate bp by aiming at reaching 80\% by 2030.
}
