% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_uhc_target.R
\name{accelerate_target_doctors}
\alias{accelerate_target_doctors}
\title{Accelerate doctors}
\usage{
accelerate_target_doctors(
  df,
  ind_ids = billion_ind_codes("uhc"),
  scenario_col = "scenario",
  bau_scenario = "historical",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{bau_scenario}{name of scenario to be used for business as usual.
Default is \code{historical}.}

\item{...}{additional parameters to be passed to scenario function}
}
\value{
data frame with acceleration scenario binded to \code{df}. \code{scenario_col} is
set to \code{acceleration}
}
\description{
Accelerate doctors using the business as usual scenario.
}
\details{
Runs:
\itemize{
\item \code{scenario_halt_rise(df, baseline_year = 2018, small_is_best = TRUE,...)},
\item \code{scenario_bau(df, small_is_best = TRUE,...)}, and
\item \code{scenario_percent_baseline(df, percent_change = -10, baseline_year = 2010, small_is_best = TRUE, ...)}
}

Then picks the best result between the three scenarios.
}
