% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_target_uhc.R
\name{accelerate_target_fh}
\alias{accelerate_target_fh}
\title{Accelerate fh}
\usage{
accelerate_target_fh(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate fh by taking the best of business as usual and halting upward trends
in the data to the 2018 value.
}
