% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_target_uhc.R
\name{accelerate_target_fp}
\alias{accelerate_target_fp}
\title{Accelerate fp}
\usage{
accelerate_target_fp(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate fp by returning to business as usual, as there are no globally agreed target.
}
\seealso{
Target scenario
\code{\link{accelerate_target_anc4}()},
\code{\link{accelerate_target_art}()},
\code{\link{accelerate_target_beds}()},
\code{\link{accelerate_target_bp}()},
\code{\link{accelerate_target_doctors}()},
\code{\link{accelerate_target_dtp3}()},
\code{\link{accelerate_target_fh}()},
\code{\link{accelerate_target_fpg}()},
\code{\link{accelerate_target_hwf}()},
\code{\link{accelerate_target_itn}()},
\code{\link{accelerate_target_nurses}()},
\code{\link{accelerate_target_pneumo}()},
\code{\link{accelerate_target_tb}()},
\code{\link{accelerate_target_uhc_sanitation}()},
\code{\link{accelerate_target_uhc_tobacco}()}
}
\concept{accelerate_target}
