% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_wrangling.R
\name{add_missing_xmart_rows}
\alias{add_missing_xmart_rows}
\title{Add missing rows for xMart upload}
\usage{
add_missing_xmart_rows(df, billion, ind_code, projected)
}
\arguments{
\item{df}{data frame}

\item{billion}{the Billion that the indicator belongs to}

\item{ind_code}{the GPW13 code for the indicator(s)}

\item{projected}{boolean value indicating where the indicator has already been
projected by the technical programme}
}
\value{
a data frame
}
\description{
Compares the rows in a given data frame to the existing xMart data. If the data frame
is missing any combinations of \code{c(iso3, ind, year)} found in xMart, it appends
these rows to the data frame,with NA values for \verb{value, upper, lower, source, type}
and \code{other_detail}.
}
