% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_fileext}
\alias{assert_fileext}
\title{Assert that a file's extension is one of a few options}
\usage{
assert_fileext(file_name, valid_exts)
}
\arguments{
\item{file_name}{A string. The name of the file}

\item{valid_exts}{A character vector. The list of valid extensions.}
}
\description{
Assert that a file's extension is one of a few options
}
