% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_same_length}
\alias{assert_same_length}
\title{Assert that two or more vectors are the same length}
\usage{
assert_same_length(..., recycle = FALSE, remove_null = FALSE)
}
\arguments{
\item{...}{Two or more vectors that should be the same length.}

\item{recycle}{(logical) Whether vectors of length one can be recycled to match the length
of the other vectors.}

\item{remove_null}{(logical) whether NULL values should be removed from the inputs before
comparison}
}
\description{
Assert that two or more vectors are the same length
}
