% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scenarios.R
\name{calculate_aarc}
\alias{calculate_aarc}
\title{Calcualte Average annual rate of change}
\usage{
calculate_aarc(baseline_year, baseline_value, end_year, end_value)
}
\arguments{
\item{baseline_year}{year where the AARC starts}

\item{baseline_value}{value at start_year}

\item{end_year}{year where the AARC ends}

\item{end_value}{value at end_year}
}
\value{
numeric with AARC
}
\description{
Calculates annual average rate of change (AARC or AROC) \emph{with} the compounded
formula:
((\code{end_value} / \code{baseline_value}))^(1/(\code{end_year} - \code{baseline_year})) - 1
}
\details{
To forecast change the baseline value:
\code{baseline_value} * ((1 + \code{aroc})^(\code{current_year} - \code{start_year}))
}
\seealso{
General scenario functions
\code{\link{add_scenario}()},
\code{\link{calculate_aroc}()},
\code{\link{exec_scenario}()},
\code{\link{fill_cols_scenario}()},
\code{\link{flat_extrapolation}()},
\code{\link{get_aarr}()},
\code{\link{get_baseline_value}()},
\code{\link{get_baseline_year}()},
\code{\link{get_last_value}()},
\code{\link{get_last_year_scenario}()},
\code{\link{get_latest_aarc}()},
\code{\link{get_percent_change_aarc}()},
\code{\link{get_target_aarc}()},
\code{\link{remove_unwanted_scenarios}()},
\code{\link{trim_values}()},
\code{\link{trim_years}()}
}
\concept{scenarios}
\keyword{internal}
