% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hep_components.R
\name{calculate_hep_components}
\alias{calculate_hep_components}
\title{Calculate HEP component indicators}
\usage{
calculate_hep_components(
  df,
  iso3 = "iso3",
  year = "year",
  ind = "ind",
  scenario = NULL,
  transform_value = "transform_value",
  type_col = "type",
  source_col = "source",
  source = sprintf("WHO DDI, \%s", format(Sys.Date(), "\%B \%Y")),
  level = stringr::str_replace(transform_value, "transform_value", "level"),
  hepi_start_year = 2018,
  ind_ids = billion_ind_codes("hep")
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{year}{Column name of column with year data.}

\item{ind}{Column name of column with indicator names.}

\item{scenario}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{transform_value}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{type_col}{Column name of column with type data.}

\item{source_col}{Column name of column to store source information.}

\item{source}{Source to use if no unique source available for the calculation.}

\item{level}{Column name(s) to create to hold levels data. Should be same length
as \code{transform_value}.}

\item{hepi_start_year}{First year to calculate HEP index for.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
Data frame in long format.
}
\description{
Currently, the Prepare and DNR indicators are already calculated in the input
data, so only Prevent is calculated in this function. It takes the numerator
and denominator data from the inputs and calculates the vaccination coverage
per year and country for each relevant pathogen, as well as the overall Prevent
score. For DNR and its components, the level is calculated in this function.
}
