% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpop.R
\name{calculate_hpop_billion}
\alias{calculate_hpop_billion}
\title{Calculate HPOP Billion}
\usage{
calculate_hpop_billion(
  df,
  iso3 = "iso3",
  ind = "ind",
  change = "change",
  ind_ids = billion_ind_codes("hpop")
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{ind}{Column name of column with indicator names.}

\item{change}{Column name of column with indicator-level change between base
year and end year.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
Data frame in long format.
}
\description{
\code{calculate_hpop_billion()} calculates country-level HPOP Billion based on
indicator level changes
}
\details{
For more details on the HPOP Billion calculation process and how this function
ties in with the rest, see the vignette:

\href{../doc/hpop.html}{\code{vignette("hpop", package = "billionaiRe")}}
}
