% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhc.R
\name{calculate_uhc_billion}
\alias{calculate_uhc_billion}
\title{Calculate UHC Billion}
\usage{
calculate_uhc_billion(
  df,
  year = "year",
  iso3 = "iso3",
  ind = "ind",
  billion_group = "billion_group",
  transform_value = "transform_value",
  ind_ids = billion_ind_codes("uhc")
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{year}{Column name of column with years.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{ind}{Column name of column with indicator names.}

\item{billion_group}{Column name of column indicating UHC Billion group that
each indicator belongs to.}

\item{transform_value}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
Data frame in long format.
}
\description{
\code{calculate_uhc_billion()} calculates country-level UHC Billion based on
indicator level data. Calculates it for each country-year combination in the provided data.
}
\details{
For more details on the UHC Billion calculation process and how this function
ties in with the rest, see the vignette:

\href{../doc/uhc.html}{\code{vignette("uhc", package = "billionaiRe")}}
}
