% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_uhc_billion.R
\name{calculate_uhc_billion_single}
\alias{calculate_uhc_billion_single}
\title{Calculate UHC Billion for one set of columns}
\usage{
calculate_uhc_billion_single(
  df,
  transform_value_col,
  value_col,
  scenario_col,
  source,
  projected_year,
  ind_ids
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{transform_value_col}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{value_col}{Column name of column with indicator values.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{source}{Source to provide for calculated average service coverage and
single measure.}

\item{projected_year}{Year that calculated data will be regarded as projected.
Years prior reported as estimated.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\description{
Calculate UHC Billion for one set of columns
}
