% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhc.R
\name{calculate_uhc_contribution}
\alias{calculate_uhc_contribution}
\title{Calculate UHC Billion}
\usage{
calculate_uhc_contribution(
  df,
  year = "year",
  iso3 = "iso3",
  ind = "ind",
  start_year = 2018,
  end_year = 2025,
  pop_year = 2025
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{year}{Column name of column with years.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{ind}{Column name of column with indicator names.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year for contribution calculation, defaults to 2025.}

\item{pop_year}{Year used to pull in HPOP populations, defaults to 2025.}
}
\value{
Data frame in long format.
}
\description{
\code{calculate_uhc_billion()} calculates country-level UHC Billion based on
indicator level data. Calculates it for each country-year combination in the provided data.
}
\details{
For more details on the UHC Billion calculation process and how this function
ties in with the rest, see the vignette:

\href{../doc/uhc.html}{\code{vignette("uhc", package = "billionaiRe")}}
}
