% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-summary.R
\name{count_since}
\alias{count_since}
\title{Count number of data points since specified date}
\usage{
count_since(df, year)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{year}{numeric}
}
\value{
data frame
}
\description{
\code{count_since} takes a \code{billionaire::load_billion_data()} dataframe and
count the number of data points exists for each indicator in \code{df} since the
specified \code{year}.
}
