% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_styles.R
\name{excel_styles}
\alias{excel_styles}
\title{Stores styles to be used in excel outputs}
\usage{
excel_styles(font = "Calibri")
}
\arguments{
\item{font}{character name with the name of the font to use.}
}
\description{
Stores styles to be used in excel outputs
}
