% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_country_summary.R
\name{export_all_country_summary_xls}
\alias{export_all_country_summary_xls}
\title{Export country summary to Excel for all billions.
\code{export_all_country_summary_xls} Export a country-specific for all billions.}
\usage{
export_all_country_summary_xls(
  df,
  iso,
  iso3 = "iso3",
  start_year = 2018,
  end_year = 2019:2025,
  output_folder = "outputs"
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{iso}{ISO3 code of country to summarize.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{output_folder}{Folder path to where the Excel files should be written}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
Export country summary to Excel for all billions.
\code{export_all_country_summary_xls} Export a country-specific for all billions.
}
