% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_country_summary.R
\name{export_hpop_country_summary_xls}
\alias{export_hpop_country_summary_xls}
\title{Export country summary to Excel for HPOP billion}
\usage{
export_hpop_country_summary_xls(
  list_df,
  iso,
  iso3 = "iso3",
  year = "year",
  ind = "ind",
  value = "value",
  transform_value = "transform_value",
  type_col = "type",
  source_col = "source",
  start_year = 2018,
  end_year = 2019:2023,
  sheet_prefix = "HPOP",
  output_fldr = "outputs",
  xls_template = "data-raw/CountrySummary_template.xlsx"
)
}
\arguments{
\item{list_df}{list object with dataframes to be used by function. Typically
comes from \code{summarize_hpop_country_data}. Names of the data frames should follow
the}

\item{iso}{ISO3 codes of country to summarize.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{year}{Column name of column with years.}

\item{ind}{Column name of column with indicator names.}

\item{value}{Column name of column with indicator values.}

\item{transform_value}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{type_col}{Column name of column to store type.}

\item{source_col}{Column name of column to store source information.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2023.}

\item{sheet_prefix}{Character prefix to add in front of export sheets}

\item{output_fldr}{Folder path to where the Excel files should be written}

\item{xls_template}{Path to the Excel template to be used.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_fldr}.
}
\description{
\code{export_hpop_country_summary_xls} Export a country-specific for HPOP billion.
}
