% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_country_summary.R
\name{export_hpop_country_summary_xls}
\alias{export_hpop_country_summary_xls}
\title{Export country summary to Excel for HPOP billion}
\usage{
export_hpop_country_summary_xls(
  df,
  iso,
  start_year = 2018,
  end_year = 2019:2023,
  output_fldr = "outputs",
  xls_template = "data-raw/CountrySummary_template.xlsx",
  ...
)
}
\arguments{
\item{output_fldr}{Folder path to where the Excel files should be written}

\item{xls_template}{Path to the Excel template to be used.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_fldr}.
}
\description{
\code{export_hpop_country_summary_xls} Export a country-specific for HPOP billion.
}
