% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate_campaign_data.R
\name{extrapolate_campaign_data}
\alias{extrapolate_campaign_data}
\title{Extrapolate campaign data for the transform value, type, and source}
\usage{
extrapolate_campaign_data(
  pathogen,
  pathogen_year,
  df,
  ind,
  year,
  transform_value,
  source
)
}
\arguments{
\item{pathogen}{Indicator IDs of the pathogens}

\item{pathogen_year}{Latest year of data to use for flat extrapolation (extrapolating after that year)}

\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind}{Column name of column with indicator names.}

\item{year}{Column name of column with year data.}

\item{transform_value}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{source}{Source to use if no unique source available for the calculation.}
}
\description{
Works across multiple \code{transform_value} columns.
}
