% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scenarios.R
\name{fill_cols_scenario}
\alias{fill_cols_scenario}
\title{Infills scenarios with \code{values} when \code{cols} are missing values.}
\usage{
fill_cols_scenario(
  df,
  scenario_col = "scenario",
  cols = list("type", "source", "use_dash", "use_calc"),
  values = list("projected",
    glue::glue("WHO DDI interim infilling and projections,\\n                                                     {format(Sys.Date(), '\%B \%Y')}"),
    TRUE, TRUE)
)
}
\arguments{
\item{df}{(data.frame) containing the data}

\item{scenario_col}{Column name of column with scenario identifiers.}

\item{cols}{(list) of columns to be infilled}

\item{values}{(list) of values to infill with.}
}
\description{
Infills scenarios with \code{values} when \code{cols} are missing values.
}
