% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{get_df_one_scenario}
\alias{get_df_one_scenario}
\title{Get data frame with just one scenario}
\usage{
get_df_one_scenario(df, scenario)
}
\description{
Get just one scenario in the data frame. Is used to avoid issues with when
multiple scenarios are present in export functions, as scenarios are not implemented
at the moment.
}
\details{
@param df data.frame to be treated
@param scenario character string identifying the scenario column in \code{df}

@return A data frame with one scenario
}
