% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{get_ind_billion}
\alias{get_ind_billion}
\title{Denotes which billion a given indicator belongs to}
\usage{
get_ind_billion(ind_code)
}
\arguments{
\item{ind_code}{(character) string identifying the indicator to identify.}
}
\description{
For example, \code{get_ind_billion("alcohol")}
returns "hpop". Note that espar currently returns only "hep".
}
