% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scenarios.R
\name{get_last_value}
\alias{get_last_value}
\title{Get last value for the filter}
\usage{
get_last_value(
  df,
  type_filter = c("reported", "estimated", "projected", "estimated")
)
}
\arguments{
\item{df}{(data.frame) containing the data. Needs to have at least the \code{year} and \code{type} columns.}

\item{type_filter}{(character vector) vector of types to be filtered for.}
}
\description{
Get last value for the filter
}
\seealso{
General scenario functions
\code{\link{add_scenario}()},
\code{\link{calculate_aarc}()},
\code{\link{calculate_aroc}()},
\code{\link{exec_scenario}()},
\code{\link{fill_cols_scenario}()},
\code{\link{flat_extrapolation}()},
\code{\link{get_aarr}()},
\code{\link{get_baseline_value}()},
\code{\link{get_baseline_year}()},
\code{\link{get_last_year_scenario}()},
\code{\link{get_latest_aarc}()},
\code{\link{get_percent_change_aarc}()},
\code{\link{get_target_aarc}()},
\code{\link{remove_unwanted_scenarios}()},
\code{\link{trim_values}()},
\code{\link{trim_years}()}
}
\concept{scenarios}
\keyword{internal}
