% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scenarios.R
\name{get_last_year_scenario}
\alias{get_last_year_scenario}
\title{Get last value for the filter in default scenario}
\usage{
get_last_year_scenario(
  df,
  indicator,
  start_year = 2018,
  scenario = "default",
  scenario_col = "scenario",
  type_filter = c("reported", "estimated")
)
}
\arguments{
\item{df}{(data.frame) containing the data. Needs to have at least the \code{year}, \code{scenario_col} and \code{type} columns.}

\item{indicator}{(character) identifying the indicator for which to get the last year of \code{type_filter} data.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{scenario}{name of scenario to recycle for.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{type_filter}{(character vector) vector of types to be filtered for.}
}
\description{
Accelerate alcohol by picking the best results between business as usual,
halt downwards trend from 2018, and -10 percent from 2010.
}
\details{
Runs:
\itemize{
\item \code{scenario_halt_rise(df, baseline_year = 2018, small_is_best = TRUE,...)},
\item \code{scenario_bau(df, small_is_best = TRUE,...)}, and
\item \code{scenario_percent_baseline(df, percent_change = -10, baseline_year = 2010, small_is_best = TRUE, ...)}
}

Then picks the best result between the three scenarios.
}
