% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scenarios.R
\name{get_latest_aarc}
\alias{get_latest_aarc}
\title{Get the latest AARC for data frame}
\usage{
get_latest_aarc(df, baseline_year, value_col = "value")
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{baseline_year}{Year from which the scenario is measured.
Defaults to \code{start_year}}

\item{value_col}{Column name of column with indicator values.}
}
\value{
dataframe with AROC (in \code{aroc} column) for every combination of
\code{iso3} and \code{ind}
}
\description{
Gets the annual average rate of change (AARC or AROC) between the
baseline year and the year immediately preceding for every combination of
\code{iso3} and \code{ind} present in \code{df}.
}
