% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hep.R
\name{get_latest_year}
\alias{get_latest_year}
\title{Get latest years for pathogens}
\usage{
get_latest_year(
  df,
  ind,
  year,
  ind_ids,
  cholera_latest_year,
  meningitis_latest_year,
  yellow_fever_latest_year
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind}{Column name of column with indicator names.}

\item{year}{Column name of column with year data.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{cholera_latest_year}{Latest year to calculate rolling sum for cholera campaign
data. From there, data is flat extrapolated. If \code{NULL}, uses the latest
year with observations in the data.}

\item{meningitis_latest_year}{Latest year to calculate rolling sum for meningitis campaign
data. From there, data is flat extrapolated. If \code{NULL}, uses the latest
year with observations in the data.}

\item{yellow_fever_latest_year}{Latest year to calculate rolling sum for yellow
fever campaign data. From there, data is flat extrapolated. If \code{NULL},
uses the latest year with observations in the data.}
}
\description{
Since campaign data is not projected, we only want to do rolling sums on the
variables until the latest year that data is available. A user can either
provide a latest year for each pathogen, or it will be determined by the available
data. This function processes that for all three pathogens to get the latest
available year.
}
