% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scenarios.R
\name{get_percent_change_aarc}
\alias{get_percent_change_aarc}
\title{Get AARC for data frame based on a percent change to baseline}
\usage{
get_percent_change_aarc(
  df,
  percent_change,
  baseline_year,
  target_year,
  value_col = "value"
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{percent_change}{Numeric with the percent_change. Should be
expressed a as percentage point and not a fraction of 100 (e.g. 6\% increase = 6,
and not 0.06).}

\item{baseline_year}{Year from which the scenario is measured.
Defaults to \code{start_year}}

\item{target_year}{Year by which the scenario should eventually be
achieved. Defaults to \code{end_year}}

\item{value_col}{Column name of column with indicator values.}
}
\value{
dataframe with AROC (in \code{aroc} column) for every combination of
\code{iso3} and \code{ind}
}
\description{
Gets the annual average rate of change (AARC or AROC) between the
baseline year and the target year given a percent change between the two
years for every combination of \code{iso3} and \code{ind} present in \code{df}.
}
\seealso{
General scenario functions
\code{\link{add_scenario}()},
\code{\link{calculate_aarc}()},
\code{\link{calculate_aroc}()},
\code{\link{exec_scenario}()},
\code{\link{fill_cols_scenario}()},
\code{\link{flat_extrapolation}()},
\code{\link{get_aarr}()},
\code{\link{get_baseline_value}()},
\code{\link{get_baseline_year}()},
\code{\link{get_last_value}()},
\code{\link{get_last_year_scenario}()},
\code{\link{get_latest_aarc}()},
\code{\link{get_target_aarc}()},
\code{\link{remove_unwanted_scenarios}()},
\code{\link{trim_values}()},
\code{\link{trim_years}()}
}
\concept{scenarios}
\keyword{internal}
