% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{get_sdi_ratio}
\alias{get_sdi_ratio}
\title{Get SDI ratio data}
\usage{
get_sdi_ratio(iso3)
}
\arguments{
\item{iso3}{Country ISO3 codes.}
}
\value{
Numeric vector of SDI ratios.
}
\description{
Matches country ISO3 codes to data in the \code{sdi_ratio} data frame.
}
