% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_whdh.R
\name{get_whdh_path}
\alias{get_whdh_path}
\title{Generate file paths for the Triple Billions WHDH data lake}
\usage{
get_whdh_path(
  operation = c("download", "upload"),
  billion = c("hep", "hpop", "uhc"),
  ind_code,
  data_type = c("ingestion_data", "wrangled_data", "projected_data", "final_data"),
  file_name = NULL
)
}
\arguments{
\item{operation}{Either "download" or "upload".}

\item{billion}{The billion to which the indicator belongs.}

\item{ind_code}{The GPW13 code for the indicator.}

\item{data_type}{The type of file/data asset. Cane be one of "ingestion_data",
"gho_backup", "unproj_data", "proj_data", "calc_data" and "full_data"}

\item{file_name}{The name of the file to download. NULL by default}
}
\value{
A string with the properly formatted file path, abiding by the standardised
directory structure and naming conventions for the 3B WHDH data lake.
}
\description{
\code{get_whdh_path()} simplifies the process of generating accurate file paths for
downloading/uploading files from the Triple Billions WHDH data lake.
}
\details{
Using this function where working with the data lake is highly recommended as
it ensures file paths abide by the established standard and directory structure
for the data lake.
}
