% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicator_df}
\alias{indicator_df}
\title{Dataset of indicators used within the Billions calculations.}
\format{
A data frame with 111 rows and 22 variables:
\describe{
\item{dashboard_id}{Dashboard ID used within the GPW13 xMart4 instance}
\item{ind}{Code used in the analysis scripts within the billionaiRe package}
\item{gho_code}{GHO storage code for indicator}
\item{ind_type}{"proj_data" if the indicator is projected by a technical programme,
and "unproj_data" if projected by DDI}
\item{uhc}{Logical, is a UHC Billions indicator}
\item{hpop}{Logical, is an HPOP Billions indicator}
\item{hep}{Logical, is a HEP Billions indicator}
\item{covariate}{Logical, is the indicator just a covariate for the Billions}
\item{calculated}{Logical, is the indicator calculated within the package and thus not an input}
\item{sdg}{Indicator code from SDG and attributed if not in SDG}
\item{short_name}{Short name of the indicator}
\item{unit_raw}{Unit of raw values for indicator}
\item{unit_transformed}{Unit of raw values for indicator}
\item{medium_name}{Medium length name of indicator (typically used for tables)}
\item{transformed_name}{Name of indicator after transformation}
\item{order}{Order of the indicators}
\item{pillar}{Pillar in which the indicators sits for UHC and HEP (e.g. Prevent, Non communicable diseases (NCDs), etc.)}
\item{small_is_best}{Logical, is a lower value is a better public health outcome.}
\item{gho_query}{Query to be passed to ghost to retrieve the data}
\item{data_source}{Name of the source of the data for a specific indicator.}
\item{acceleration_scenario}{Definition of the acceleration scenario.}
\item{sdg_scenario}{Definition of the acceleration scenario.}
}
}
\usage{
indicator_df
}
\description{
A dataset containing dashboard and analysis indicator codes and their uses within the GPW13
Billions.
}
\seealso{
Functions to access metadata
\code{\link{billion_ind_codes}()},
\code{\link{convert_ind_codes}()},
\code{\link{get_country_shares}()},
\code{\link{get_ind_billion}()},
\code{\link{get_ind_metadata}()},
\code{\link{get_sdi_ratio}()}
}
\concept{metadata}
\keyword{datasets}
