% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicator_df}
\alias{indicator_df}
\title{Dataset of indicators used within the Billions calculations.}
\format{
A data frame with 101 rows and 8 variables:
\describe{
\item{dashboard_id}{Dashboard ID used within the GPW13 xMart4 instance}
\item{analysis_code}{Code used in the analysis scripts within the billionaiRe package}
\item{gho_code}{GHO storage code for indicator}
\item{uhc}{Logical, is a UHC Billions indicator}
\item{hpop}{Logical, is an HPOP Billions indicator}
\item{hep}{Logical, is a HEP Billions indicator}
\item{covariate}{Logical, is the indicator just a covariate for the Billions}
\item{calculated}{Logical, is the indicator calculated within the package and thus not an input}
}
}
\usage{
indicator_df
}
\description{
A dataset containing dashboard and analysis indicator codes and their uses within the GPW13
Billions.
}
\keyword{datasets}
