% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicator_order}
\alias{indicator_order}
\title{Indicator order}
\format{
A data frame with  rows and 8
variables:
\describe{
\item{sdg}{Indicator code from SDG and attributed if not in SDG}
\item{short_name}{Short name of the indicator}
\item{unit_raw}{Unit of raw values for indicator}
\item{unit_transformed}{Unit of raw values for indicator}
\item{medium_name}{Medium length name of indicator (typically used for tables)}
\item{transformed_name}{Name of indicator after transformation}
\item{ind}{Code of the indicator in billionaiRe}
\item{order}{Order of the indicators }
}
}
\usage{
indicator_order
}
\description{
A dataset containing analysis indicators from \code{indicator_df}
combined with ordering number, and expended indicator definition to create
good looking export files.
}
\keyword{datasets}
