% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_billion_data}
\alias{load_billion_data}
\title{Load Raw Billions Indicator Data}
\usage{
load_billion_data(
  billion = c("hep", "hpop", "uhc", "all"),
  mart_table = c("full_ind", "unproj_data", "proj_data"),
  date_filter = "latest",
  na_rm = TRUE,
  format = c("none", "csv", "streaming"),
  ...
)
}
\arguments{
\item{billion}{Billions data to load, one of "hep" (default), "hpop", "uhc", or "all".}

\item{mart_table}{Name of xMart4 table in the GPW13 mart to load data from. Currently,
pulls from the full input data to the Billions by default, but can also pull
from the table containing data needing DDI infilling and projections or
those already infilled and projected by technical programmes.}

\item{date_filter}{One of \code{NULL}, "latest", or a single date string. The date
string needs to be in ISO6801 format, such as "1989-4-4" or "1988-06-21".
If a date is provided, will only take values loaded on or prior to that
date into the xMart4 database.}

\item{na_rm}{Logical value, specifying whether to filter the data to only rows
where \code{value} is not missing. Defaults to \code{TRUE}.}

\item{format}{Specification of the output format to be returned by the xMart API.
Defaults to \code{"none"}, but consider switching to \code{"csv"} if you are
loading an extremely large table. Passed to \code{\link[xmart4:xmart4_table]{xmart4::xmart4_table()}}. See the
\href{https://portal-uat.who.int/xmart4/docs/xmart_api/use_API.html}{xMart4 API documentation}
for details on the three options.}

\item{...}{Additional arguments passed to \code{xmart4::xmart4_table()}. Use if
you need to provide additional token specifications for Azure authentication.}
}
\value{
A data frame.
}
\description{
\code{load_billion_data()} provides access to the raw inputs to the Billions stored
within the World Health Organization's xMart4 database. By default, it filters
the loaded data to take the most recently uploaded data for each indicator,
country, and year. The user can also specify to take the latest data before a
certain time period, or to extract all data from the database.
}
\details{
This function requires that the user have the \href{https://github.com/caldwellst/xmart4}{xmart4} package
installed and setup to access the GPW13 mart. Details are available on the GitHub
landing page linked above.
}
