% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_billion_data}
\alias{load_billion_data}
\title{Load Raw Billions Indicator Data}
\usage{
load_billion_data(
  billion = c("hep", "hpop", "uhc", "all"),
  date_filter = "latest"
)
}
\arguments{
\item{billion}{Billions data to load, one of "hep" (default), "hpop", "uhc", or "all".}

\item{date_filter}{One of \code{NULL}, "latest", or a single date string. The date
string needs to be in ISO6801 format, such as "1989-4-4" or "1988-06-21".}
}
\value{
A data frame.
}
\description{
\code{load_billion_data()} provides access to the raw inputs to the Billions stored
within the World Health Organization's xMart4 database. By default, it filters
the loaded data to take the most recently uploaded data for each indicator,
country, and year. The user can also specify to take the latest data before a
certain time period, or to extract all data from the database.
}
\details{
This function requires that the user have the \href{https://github.com/caldwellst/xmart4}{xmart4} package
installed and setup to access the GPW13 mart. Details are available on the GitHub
landing page linked above. The date filter
}
