% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_billion_data_xmart}
\alias{load_billion_data_xmart}
\title{Load Billions indicator data from xMart}
\usage{
load_billion_data_xmart(
  data_type = c("wrangled_data", "projected_data", "final_data"),
  billion = c("all", "hep", "hpop", "uhc"),
  ind_codes = "all",
  date_filter = "latest",
  na_rm = TRUE,
  sandbox = FALSE,
  silent = TRUE
)
}
\arguments{
\item{data_type}{(string) The type of data to load.
\itemize{
\item \code{wrangled_data} (default): raw data that has been wrangled into a suitable
form for analysis.
\item \code{projected_data}:  data that has been fully projected to the target year but
has not yet been transformed or calculated upon.
\item \code{final_data}: the complete set of billions data with transformed values, contributions,
and all calculations available.
}}

\item{billion}{(string) One of "all" (default), "hep", "hpop", or "uhc". If "all",
downloads all indicators for all three billions.}

\item{ind_codes}{(character vector) The name of the indicator (or indicators) to load data for.
If \code{all}, downloads data for all indicators for a given billion. Ignored if
billion = "all".}

\item{date_filter}{(string) Either "latest", or a single date string. The date
string needs to be in ISO6801 format, such as "1989-4-4" or "1988-06-21".
If a date is provided, the returned data frame is a snapshot of the data as
stored on that date (if no updates were made on the given date, then the most
recent update before that date is used).}

\item{na_rm}{(logical) Specifies whether to filter the data to only rows
where \code{value} is not missing. Defaults to \code{TRUE}.}

\item{sandbox}{(logical) Whether or not to use the sandbox folders in the data lake.}

\item{silent}{(logical) Specifies whether to show authentication messages and
a download progress bar. Defaults to \code{TRUE}.}
}
\value{
A data frame.
}
\description{
Load Billions indicator data from xMart
}
