% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hep.R
\name{prevent_calculations}
\alias{prevent_calculations}
\title{Calculate prevent indicators}
\usage{
prevent_calculations(df, iso3, year, ind, type, transform_value, ind_ids)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{year}{Column name of column with year data.}

\item{ind}{Column name of column with indicator names.}

\item{type}{Column name of column with type data.}

\item{transform_value}{Column name of column with transformed values.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\description{
This function calculates the overall vaccine coverage scores for each pathogen,
including the overall prevent indicator. It is used within \code{calculate_hep_components()}
and primarily relies on using \code{purrr::pmap_dfr()} to apply the \code{pathogen_calc()}
function for each prevent component and overall score.
}
