% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scenarios.R
\name{remove_unwanted_scenarios}
\alias{remove_unwanted_scenarios}
\title{Remove unwanted scenarios from df}
\usage{
remove_unwanted_scenarios(df, scenario_col = "scenario", unwanted_scenarios)
}
\arguments{
\item{df}{(data.frame) containing the data. Needs to have at least the \code{scenario_col} column}

\item{scenario_col}{Column name of column with scenario identifiers.}

\item{unwanted_scenarios}{character vector of scenarios to be remove}
}
\description{
Remove unwanted scenarios from df
}
\seealso{
General scenario functions
\code{\link{add_scenario}()},
\code{\link{calculate_aarc}()},
\code{\link{calculate_aroc}()},
\code{\link{exec_scenario}()},
\code{\link{fill_cols_scenario}()},
\code{\link{flat_extrapolation}()},
\code{\link{get_aarr}()},
\code{\link{get_baseline_value}()},
\code{\link{get_baseline_year}()},
\code{\link{get_last_value}()},
\code{\link{get_last_year_scenario}()},
\code{\link{get_latest_aarc}()},
\code{\link{get_percent_change_aarc}()},
\code{\link{get_target_aarc}()},
\code{\link{trim_values}()},
\code{\link{trim_years}()}
}
\concept{scenarios}
\keyword{internal}
