% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scenarios.R
\name{remove_unwanted_scenarios}
\alias{remove_unwanted_scenarios}
\title{Remove unwanted scenarios from df}
\usage{
remove_unwanted_scenarios(df, scenario_col = "scenario", unwanted_scenarios)
}
\arguments{
\item{df}{(data.frame) containing the data. Needs to have at least the \code{scenario_col} column}

\item{scenario_col}{Column name of column with scenario identifiers.}

\item{unwanted_scenarios}{character vector of scenarios to be remove}
}
\description{
Remove unwanted scenarios from df
}
