% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{scenario_bau}
\alias{scenario_bau}
\title{Business as usual scenario}
\usage{
scenario_bau(df, value = "value", baseline = 2018, sname = "scen_bau")
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{value}{Column name of column with indicator values.}

\item{baseline}{Year which is to be used as baseline for scenario (discuss Seth)}

\item{sname}{Scenario name. Defaults to \code{'scen_bau'} (discuss with Seth)}
}
\value{
Dataframe with additional scenario column
}
\description{
Scenario that uses existing estimates/projections for years
after the baseline year
}
\examples{
hpop_df \%>\% scenario_bau()
}
