% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R, R/scenarios_temp.R
\name{scenario_bau}
\alias{scenario_bau}
\title{Business as usual scenario}
\usage{
scenario_bau(df, value = "value", baseline = 2018, sname = "scen_bau")

scenario_bau(df, value = "value", baseline = 2018, sname = "scen_bau")

scenario_bau(df, value = "value", baseline = 2018, sname = "scen_bau")

scenario_bau(df, value = "value", baseline = 2018, sname = "scen_bau")
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{value}{Column name of column with indicator values.}

\item{baseline}{Year which is to be used as start of output scenario (discuss Seth)}

\item{sname}{Scenario name. Will be automatically constructed if not set (discuss with Seth)}

\item{percent_decrease}{The percentage decrease that is to be acheived from value
in baseline year by target_year}

\item{old_baseline}{Year from which the percent decrease is measured}

\item{target_year}{Year by which the percent of decrease should eventually be acheived}
}
\value{
Dataframe with additional scenario column

Dataframe with additional scenario column

Dataframe with additional scenario column

Dataframe with additional scenario column
}
\description{
Scenario that uses existing estimates/projections for years
after the baseline year

The scenario returned is often for a different period to the original scenario
For example: if the original scenario is to reduce the 2012 value by 30\% by 2030
then the scenario returned will be a portion of the straight line
drawn from the baseline year value (eg 2018)  to the 2030 target (but it will
only go out as far as the final year in the input data frame)

Scenario that uses existing estimates/projections for years
after the baseline year

The scenario returned is often for a different period to the original scenario
For example: if the original scenario is to reduce the 2012 value by 30\% by 2030
then the scenario returned will be a portion of the straight line
drawn from the baseline year value (eg 2018)  to the 2030 target (but it will
only go out as far as the final year in the input data frame)
}
\examples{
hpop_df \%>\% scenario_bau()
hpop_df \%>\% scenario_precent_decrease(percent_decrease = 25)
hpop_df \%>\% scenario_bau()
hpop_df \%>\% scenario_precent_decrease(percent_decrease = 25)
}
