% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios_target.R
\name{scenario_fixed_target}
\alias{scenario_fixed_target}
\title{Scenario to reach a fixed target from a baseline year by a target year}
\usage{
scenario_fixed_target(
  df,
  target_value,
  value = "value",
  ind = "ind",
  iso3 = "iso3",
  year = "year",
  scenario = "scenario",
  start_year = 2018,
  end_year = 2025,
  baseline_year = start_year,
  target_year = end_year,
  scenario_name = glue::glue("{target_value}_{target_year}"),
  small_is_best = FALSE,
  trim = TRUE,
  keep_better_values = TRUE,
  upper_limit = 100,
  lower_limit = 0,
  trim_years = TRUE
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{target_value}{value to be achieved by scenario by \code{target_year}}

\item{value}{Column name of column with indicator values.}

\item{ind}{Column name of column with indicator names.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{year}{Column name of column with years.}

\item{scenario}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{start_year}{Start year for scenario, defaults to 2018.}

\item{end_year}{End year for scenario, defaults to 2025}

\item{baseline_year}{Year from which the scenario is measured.
Defaults to \code{start_year}}

\item{target_year}{Year by which the scenario should eventually be
achieved. Defaults to \code{end_year}}

\item{scenario_name}{Name of the scenario. Defaults to scenario_{percent_change}_{baseline_year}}

\item{small_is_best}{Logical to identify if a lower value is better than a higher
one (e.g. lower obesity in a positive public health outcome, so obesity rate
should have small_is_best = TRUE).}

\item{trim}{logical to indicate if the data should be trimmed between
\code{upper_limit} and \code{lower_limit}.}

\item{keep_better_values}{logical to indicate if "better" values should be
kept from \code{value} if they are present. Follows the direction set in
\code{small_is_best}.  For instance, if small_is_best is TRUE, then \code{value} lower
than \code{col} will be kept.}

\item{upper_limit}{limit at which the indicator should be caped.
Can take any of "guess", or any numeric. \code{guess} (default) will take 100 as
the limit if \code{percent_change} is positive, and 0 if negative.}

\item{lower_limit}{limit at which the indicator should be caped.
Can take any of "guess", or 0 to 100. \code{guess} (default) will take 0 as the
limit if \code{percent_change} is positive, and 100 if negative.}

\item{trim_years}{logical to indicate if years before \code{start_year} and after
\code{end_year} should be removed}
}
\value{
Dataframe with scenario rows
}
\description{
This scenario allows to reach a target value from a baseline year by target
year It provides values for scenarios stated as "Reach XX\% in INDICATOR by
YEAR" or "Eliminate XX by YEAR".
}
\details{
The returned scenario is a portion of the straight line drawn from the
\code{baseline_year} value to the \code{target_year}. Only values for years between
\code{start_year} and \code{end_year} will be returned.

If \code{value} has values that are higher or lower than the scenario values, then
only \code{value} will be kept, depending on \code{small_is_best}. For instance,
if the scenario value is 80 and the value 75 and small_is_best is TRUE, then
80 will be kept.
}
