% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R, R/scenarios_temp.R
\name{scenario_percent_baseline}
\alias{scenario_percent_baseline}
\title{Scenario to reduce by a fixed percentage from a baseline value}
\usage{
scenario_percent_baseline(
  df,
  percent_decrease,
  value = "value",
  ind = "ind",
  iso3 = "iso3",
  year = "year",
  baseline = 2018,
  old_baseline = baseline,
  target_year = 2025,
  sname = NULL
)

scenario_percent_baseline(
  df,
  percent_decrease,
  value = "value",
  ind = "ind",
  iso3 = "iso3",
  year = "year",
  baseline = 2018,
  old_baseline = baseline,
  target_year = 2025,
  sname = NULL
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{percent_decrease}{The percentage decrease that is to be acheived from value
in baseline year by target_year}

\item{value}{Column name of column with indicator values.}

\item{baseline}{Year which is to be used as start of output scenario (discuss Seth)}

\item{old_baseline}{Year from which the percent decrease is measured}

\item{target_year}{Year by which the percent of decrease should eventually be acheived}

\item{sname}{Scenario name. Will be automatically constructed if not set (discuss with Seth)}
}
\value{
Dataframe with additional scenario column

Dataframe with additional scenario column
}
\description{
The scenario returned is often for a different period to the original scenario
For example: if the original scenario is to reduce the 2012 value by 30\% by 2030
then the scenario returned will be a portion of the straight line
drawn from the baseline year value (eg 2018)  to the 2030 target (but it will
only go out as far as the final year in the input data frame)

The scenario returned is often for a different period to the original scenario
For example: if the original scenario is to reduce the 2012 value by 30\% by 2030
then the scenario returned will be a portion of the straight line
drawn from the baseline year value (eg 2018)  to the 2030 target (but it will
only go out as far as the final year in the input data frame)
}
