% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios_return_previous_trajectory.R
\name{scenario_return_previous_trajectory}
\alias{scenario_return_previous_trajectory}
\title{Scenario return to previpus trajectory}
\usage{
scenario_return_previous_trajectory(
  df,
  dip_year = 2020,
  recovery_year = 2022,
  start_year = 2018,
  end_year = 2025,
  value_col = "value",
  scenario_col = "scenario",
  scenario_name = "return_previous_trajectory",
  scenario_shock = "covid_shock",
  scenario_previous_trajectory = "pre_covid_trajectory",
  trim = TRUE,
  small_is_best = FALSE,
  keep_better_values = FALSE,
  upper_limit = 100,
  lower_limit = 0,
  trim_years = TRUE,
  source = sprintf("WHO DDI, \%s", format(Sys.Date(), "\%B \%Y"))
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{dip_year}{(integer) year where the dip appends}

\item{recovery_year}{(integer) year from which the AROC will be applied}

\item{start_year}{Start year for scenario, defaults to 2018.}

\item{end_year}{End year for scenario, defaults to 2025}

\item{value_col}{Column name of column with indicator values. This column will be
used to return the results.}

\item{scenario_col}{(character) name of the column with the scenarios.}

\item{scenario_name}{name of scenario}

\item{scenario_shock}{(character) name of the scenario with the shock}

\item{scenario_previous_trajectory}{(character) name of the scenario with the previous trajectories.}

\item{trim}{logical to indicate if the data should be trimmed between
\code{upper_limit} and \code{lower_limit}.}

\item{small_is_best}{Logical to identify if a lower value is better than a higher
one (e.g. lower obesity in a positive public health outcome, so obesity rate
should have small_is_best = TRUE).}

\item{keep_better_values}{logical to indicate if "better" values should be
kept from \code{value_col} if they are present. Follows the direction set in
\code{small_is_best}.  For instance, if small_is_best is TRUE, then \code{value_col} lower
than \code{col} will be kept.}

\item{upper_limit}{upper limit at which the indicator should be caped.}

\item{lower_limit}{lower_limit limit at which the indicator should be caped.}

\item{trim_years}{logical to indicate if years before \code{start_year} and after
\code{end_year} should be removed}

\item{source}{Source to use if no unique source available for the calculation.}
}
\value{
a data frame with scenario values in \code{value_col} with a \code{scenario} column.
}
\description{
This scenario returns to the \code{scenario_previous_trajectory} after the last
value of \code{scenario_shock}.
}
\details{
Two types of AROC are supported:
\itemize{
\item \code{lastest_year}: the AROC is calculated the values between the \code{start_year} and the last \code{reported} or
\code{estimated} value before \code{dip_year} is applied to the last reported value to
\code{recovery_year} onward.
\item \code{average_years_in_range}: all AROC between the \code{aroc_start_year} and the last \code{reported} or
\code{estimated} value before \code{dip_year} are calculated. The average AROC for that period is then
applied to the last reported value to \code{recovery_year} onward.
}

If there are missing values between \code{dip_year} and
\code{recovery_year}, the last value from \code{dip_year} is carried forward. This
applies only to countries where the indicator value for \code{dip_year} is
\code{reported} or \code{estimated}. Otherwise, the value is carried with
\code{scenario_bau}.

If \code{progressive_recovery} is TRUE, then the recovery is
spread between the years between \code{recovery_year} and \code{end_year}. For
instance, if \code{recovery_year} is 2021 and \code{end_year} 2025, then 2021 will have 0\%
of AROC, 2022 25\%, 2023 50\%, 2024 75\%, and 2025 100\%.
}
