% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_uhc.R
\name{sdg_art}
\alias{sdg_art}
\title{Accelerate art to SDG target}
\usage{
sdg_art(df, scenario_col = "scenario", ind_ids = billion_ind_codes("uhc"), ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{...}{additional parameters to be passed to scenario function}
}
\value{
data frame with acceleration scenario binded to \code{df}. \code{scenario_col} is
set to \code{acceleration}
}
\description{
Accelerate art by first dividing countries into those with reported data and
those without.
\itemize{
\item For countries without reported data, business as usual is returned.
\item For countries with reported data, the best of business as usual and \strong{fixed
target of 90.25\% by 2025} is chosen.
}
}
