% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_uhc.R
\name{sdg_beds}
\alias{sdg_beds}
\title{Accelerate beds to SDG target}
\usage{
sdg_beds(
  df,
  ind_ids = billion_ind_codes("uhc"),
  scenario = "scenario",
  ind = "ind",
  year = "year",
  value = "value",
  iso3 = "iso3",
  start_year = 2018,
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{ind}{Column name of column with indicator names.}

\item{year}{Column name of column with years.}

\item{value}{Column name of column with indicator values.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{...}{additional parameters to be passed to scenario function}
}
\value{
data frame with acceleration scenario binded to \code{df}. \code{scenario} is
set to \code{acceleration}
}
\description{
Accelerate beds by first dividing countries into two groups:
\itemize{
\item For countries with 18 or more beds for all years after 2018, business
as usual is returned.
\item For countries which have less than 18 beds for any of the years after 2018 (inclusive),
the best of business as usual and a \strong{linear change of 0.36 per year up to 2025},
with an upper limit of 18, is returned.
}
}
