% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_hpop.R
\name{sdg_hpop_sanitation}
\alias{sdg_hpop_sanitation}
\title{Accelerate hpop_sanitation to SDG target}
\usage{
sdg_hpop_sanitation(
  df,
  ind_ids = billion_ind_codes("hpop"),
  end_year = 2025,
  scenario_col = "scenario",
  scenario_name = "sdg",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{scenario_name}{name of scenario}

\item{...}{additional parameters to be passed to scenario function}
}
\value{
data frame with acceleration scenario binded to \code{df}. \code{scenario_col} is
set to \code{acceleration}
}
\description{
Accelerate hpop_sanitation by aiming at 100\% by 20230
}
\details{
Runs:
\itemize{
\item \code{scenario_fixed_target(df, target_value = 95,target_year = 2030, small_is_best = FALSE)}
}
}
