% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_hpop.R
\name{sdg_hpop_sanitation_urban}
\alias{sdg_hpop_sanitation_urban}
\title{Accelerate hpop_sanitation_urban to SDG target}
\usage{
sdg_hpop_sanitation_urban(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate hpop_sanitation_urban by aiming at 95\% by 20230
}
