% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_hpop.R
\name{sdg_ipv}
\alias{sdg_ipv}
\title{Accelerate ipv to SDG target}
\usage{
sdg_ipv(df, scenario_col = "scenario", ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{...}{additional parameters to be passed to scenario function}
}
\value{
data frame with acceleration scenario binded to \code{df}. \code{scenario_col} is
set to \code{acceleration}
}
\description{
Accelerate ipv by targeting 0 by 2030.
}
\details{
Runs:
\itemize{
\item \code{scenario_fixed_target(df, target_value = 0, target_year = 2030, small_is_best = TRUE,...)},
}
}
