% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_uhc.R
\name{sdg_nurses}
\alias{sdg_nurses}
\title{Accelerate nurses to SDG target}
\usage{
sdg_nurses(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate nurses using the business as usual scenario.
}
