% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_hpop.R
\name{sdg_overweight}
\alias{sdg_overweight}
\title{Accelerate overweight to SDG target}
\usage{
sdg_overweight(
  df,
  ind_ids = billion_ind_codes("hpop"),
  scenario_col = "scenario",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{...}{additional parameters to be passed to scenario function}
}
\value{
data frame with acceleration scenario binded to \code{df}. \code{scenario_col} is
set to \code{acceleration}
}
\description{
Accelerate overweight by picking the best value between business as usual and
AROC of 3 by 2030.
}
\details{
Runs:
\itemize{
\item \code{scenario_bau(df, small_is_best = TRUE,...)}
\item \code{scenario_aroc(df, aroc_type = "target", target_value = 3, target_year = 2030, small_is_best = TRUE,...)}
}

Then picks the best result between the two scenarios.
}
