% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_uhc.R
\name{sdg_pneumo}
\alias{sdg_pneumo}
\title{Accelerate pneumo to SDG target}
\usage{
sdg_pneumo(df, ...)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate \code{pneumo} by aiming at 100 by 2030 if there are 2 values or more
reported. Otherwise, business as usual.
}
\seealso{
UHC SDG scenarios
\code{\link{sdg_anc4}()},
\code{\link{sdg_art}()},
\code{\link{sdg_beds}()},
\code{\link{sdg_bp}()},
\code{\link{sdg_doctors}()},
\code{\link{sdg_dtp3}()},
\code{\link{sdg_fh}()},
\code{\link{sdg_fpg}()},
\code{\link{sdg_fp}()},
\code{\link{sdg_hwf}()},
\code{\link{sdg_itn}()},
\code{\link{sdg_nurses}()},
\code{\link{sdg_tb}()},
\code{\link{sdg_uhc_sanitation}()},
\code{\link{sdg_uhc_tobacco}()}
}
\concept{uhc_sdg}
