% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_uhc.R
\name{sdg_uhc_tobacco}
\alias{sdg_uhc_tobacco}
\title{Accelerate uhc_tobacco to SDG target}
\usage{
sdg_uhc_tobacco(
  df,
  ind_ids = billion_ind_codes("uhc"),
  scenario_col = "scenario",
  ...
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{...}{additional parameters to be passed to scenario function}
}
\description{
Accelerate \code{uhc_tobacco} by aiming at 100 by 2030 if there are 2 values or more
reported. Otherwise, business as usual.
}
\seealso{
UHC SDG scenarios
\code{\link{sdg_anc4}()},
\code{\link{sdg_art}()},
\code{\link{sdg_beds}()},
\code{\link{sdg_bp}()},
\code{\link{sdg_doctors}()},
\code{\link{sdg_dtp3}()},
\code{\link{sdg_fh}()},
\code{\link{sdg_fpg}()},
\code{\link{sdg_fp}()},
\code{\link{sdg_hwf}()},
\code{\link{sdg_itn}()},
\code{\link{sdg_nurses}()},
\code{\link{sdg_pneumo}()},
\code{\link{sdg_tb}()},
\code{\link{sdg_uhc_sanitation}()}
}
\concept{uhc_sdg}
