% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdi_ratio}
\alias{sdi_ratio}
\title{Socio-Demographic Index data used in road safety transform.}
\format{
A data frame with 61 rows and 5 variables:
\describe{
\item{iso3}{Country ISO3 codes.}
\item{sdiratio}{SDI ratio.}
}
}
\usage{
sdi_ratio
}
\description{
Used internally to transform road safety data. Exact methods available in methods report
}
\keyword{datasets}
