% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_data.R
\name{style_data}
\alias{style_data}
\title{Style data according to its type}
\usage{
style_data(
  data_type,
  wb,
  sheet_name,
  rows,
  cols,
  no_show = TRUE,
  no_show_row = 0
)
}
\arguments{
\item{data_type}{character vector of class(es) of the data to be styled.
Can be one of "numeric", "integer", "Date", "character", "character_centered",
or "formula".}

\item{wb}{a \code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{rows}{Rows to apply style to.}

\item{cols}{columns to apply style to.}

\item{no_show}{Boolean with TRUE if data in \code{no_show_row} should be displayed}

\item{no_show_row}{integer identifying which row contains data no to show}
}
\description{
Style data according to its type
}
