% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_data.R
\name{style_data_single}
\alias{style_data_single}
\title{Style data single data column according to its type}
\usage{
style_data_single(
  data_type = c("numeric", "integer", "Date", "character", "character_centered"),
  wb,
  sheet_name,
  rows,
  col
)
}
\arguments{
\item{data_type}{character vector of class(es) of the data to be styled.
Can be one of "numeric", "integer", "Date", "character", or "character_centered"}

\item{wb}{\code{\link[openxlsx:createWorkbook]{openxlsx::createWorkbook()}} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{rows}{Rows to apply style to.}

\item{col}{Column to apply style to.}
}
\description{
Style data single data column according to its type
}
