% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_hpop_summary_sheet.R
\name{style_header_hpop_summary_sheet}
\alias{style_header_hpop_summary_sheet}
\title{Style HPOP summary worksheet header}
\usage{
style_header_hpop_summary_sheet(wb, sheet_name, start_row, start_col)
}
\arguments{
\item{wb}{a \code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{start_row}{integer start row of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}

\item{start_col}{integer start col of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{style_header_hpop_summary_sheet} styles the title and sub-title of the worksheet header.
}
