% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_hpop_datasheet.R
\name{style_hpop_billion_contribution}
\alias{style_hpop_billion_contribution}
\title{Style HPOP billion contribution all indicators summary box}
\usage{
style_hpop_billion_contribution(df, wb, sheet_name, start_row, start_col)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{\code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{start_row}{numeric start row of the table in the excel sheet}

\item{start_col}{numeric start row of the table in the excel sheet}
}
\description{
Style HPOP billion contribution all indicators summary box
}
