% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_hpop_timeseries.R
\name{style_hpop_timeseries}
\alias{style_hpop_timeseries}
\title{Styles HPOP timerserie}
\usage{
style_hpop_timeseries(
  df,
  wb,
  sheet_name,
  start_row,
  start_col,
  ind,
  year,
  type_col,
  df_wide,
  trans_value
)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{openxlsx workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{start_row}{numeric start row of the table in the excel sheet}

\item{start_col}{numeric start row of the table in the excel sheet}

\item{ind}{Column name of column with indicator names.}

\item{year}{Column name of column with years.}

\item{type_col}{Column name of column to store type.}
}
\description{
Styles HPOP timerserie
}
