% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_indicator_list.R
\name{style_indicator_list_sheet}
\alias{style_indicator_list_sheet}
\title{Styles indicator sheet}
\usage{
style_indicator_list_sheet(df, wb, sheet_name, start_row, start_col)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{openxlsx workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{start_row}{numeric start row of the table in the excel sheet}

\item{start_col}{numeric start row of the table in the excel sheet}
}
\description{
Applies styles to the data worksheet for country summary outputs
}
