% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_uhc_summary_sheet.R
\name{style_uhc_pillar}
\alias{style_uhc_pillar}
\title{Style UHC Pillars summary sheet}
\usage{
style_uhc_pillar(
  wb,
  sheet_name,
  boxes_bounds,
  data_type,
  pillar = c("RMNCH", "infec_diseases", "ncd", "service_cap_access"),
  no_show = TRUE,
  no_show_row = 2
)
}
\arguments{
\item{wb}{\code{\link[openxlsx:createWorkbook]{openxlsx::createWorkbook()}} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{boxes_bounds}{named list of bounds for data frame boxes to be written in sheet.}

\item{data_type}{named list with latest_reported and baseline_projection data
types. Passed to \code{style_data()}}

\item{pillar}{character Pillar identifying the pillar to style. Must be one of
\itemize{
\item RMNCH
\item infectious
\item NCD
\item service
}}

\item{no_show}{Boolean with TRUE if data in \code{no_show_row} should be displayed}

\item{no_show_row}{integer identifying which row contains data no to show}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{style_uhc_pillar()} styles the UHC Pillars (RMNCH, infectious diseases, etc.)
box/section of the UHC summary sheet. Used within \code{write_RMNCH_uhc_summary()}.
}
