% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_country.R
\name{summarize_hpop_billion_contribution}
\alias{summarize_hpop_billion_contribution}
\title{Summarize HPOP billion contributions for all indicators}
\usage{
summarize_hpop_billion_contribution(df, year, end_year, ind, contribution)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{year}{Column name of column with years.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2023.}

\item{ind}{Column name of column with indicator names.}

\item{contribution}{Glue specification used to transform \code{transform_value} column
names into contribution column names. Defaults to replacing the string
\code{"transform_value"} with \code{"contribution"}.}
}
\value{
data frame with summarized data
}
\description{
\code{summarize_hpop_data()} summarizes the data for the HPOP country summary
Excel file. It is used within \code{\link[=export_hpop_country_summary_xls]{export_hpop_country_summary_xls()}}.
}
\details{
For more details on the HPOP Billion calculation process and how this function
ties in with the rest, see the vignette:

\href{../doc/hpop.html}{\code{vignette("hpop", package = "billionaiRe")}}
}
