% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_country.R
\name{summarize_hpop_data}
\alias{summarize_hpop_data}
\title{Summarize data for HPOP country summary export}
\usage{
summarize_hpop_data(
  df,
  year = "year",
  iso3 = "iso3",
  ind = "ind",
  value = "value",
  transform_value = "transform_value",
  contribution = stringr::str_replace(transform_value, "transform_value",
    "contribution"),
  contribution_pct = paste0(contribution, "_percent"),
  contribution_pct_pop_total = paste0(contribution, "_percent_pop_total"),
  population = "population",
  scenario = NULL,
  type_col = "type",
  source_col = "source",
  ind_ids = billion_ind_codes("hpop"),
  start_year = 2018,
  end_year = 2019:2023,
  ind_df
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{year}{Column name of column with years.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{ind}{Column name of column with indicator names.}

\item{value}{Column name of column with indicator values.}

\item{transform_value}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{contribution}{Column name of column(s) to store contribution (population)
values. Must be the same length as \code{transform_value}.}

\item{contribution_pct}{Column name of column(s) to store contribution (percent)
values. Must be the same length as \code{transform_value}.}

\item{contribution_pct_pop_total}{Column name of column(s) to store contribution
(percent of total population of the country) values. Must be the same length
as \code{transform_value}.}

\item{population}{Column name of column with population figures.}

\item{scenario}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{type_col}{Column name of column to store type.}

\item{source_col}{Column name of column to store source information.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2023.}

\item{ind_df}{data frame containing the indicators in the correct order and format to be used.}
}
\value{
list of data frames to be used in the 'data' sheet of HPOP country summary
}
\description{
\code{summarize_hpop_data()} summarizes the data for the HPOP country summary
Excel file. It is used within \code{\link[=export_hpop_country_summary_xls]{export_hpop_country_summary_xls()}}.
}
