% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_style.R
\name{timeseries_style}
\alias{timeseries_style}
\title{Change font of time series values}
\usage{
timeseries_style(wb, iso, font_df, b_sheet)
}
\arguments{
\item{b_sheet}{}
}
\value{

}
\description{
\code{timeseries_style()} changes the font of time series values based on the
indicator type it is:
}
\details{
\itemize{
\item bold: reported
\item normal: estimated
\item faded: imputed or projected
}
}
