% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_hpop.R
\name{transform_hpop_single}
\alias{transform_hpop_single}
\title{Perform a transformation on a single column}
\usage{
transform_hpop_single(df, value_col, transform_col, ind_ids)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{value_col}{Column name of column with indicator values.}

\item{transform_col}{Column to put transformed values into}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
A single column data frame of transformed values.
}
\description{
This function is used within \code{\link[=transform_hpop_data]{transform_hpop_data()}} to generate transformed data
on a single column.
}
\seealso{
Other hpop: 
\code{\link{add_hpop_populations}()},
\code{\link{calculate_hpop_billion_change}()},
\code{\link{calculate_hpop_billion}()},
\code{\link{calculate_hpop_contributions}()},
\code{\link{hpop_df}},
\code{\link{transform_hpop_data}()},
\code{\link{untransform_hpop_data}()},
\code{\link{untransform_hpop_single}()}
}
\concept{hpop}
\keyword{internal}
