% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hep.R
\name{transform_prev_cmpgn_data}
\alias{transform_prev_cmpgn_data}
\title{Transform Prevent campaigns data}
\usage{
transform_prev_cmpgn_data(
  df,
  iso3,
  year,
  ind,
  value,
  type,
  ind_ids,
  cholera_latest_year = NULL,
  meningitis_latest_year = NULL,
  yellow_fever_latest_year = NULL,
  extrapolate_to = 2025
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{year}{Column name of column with year data.}

\item{ind}{Column name of column with indicator names.}

\item{value}{Column name of column with indicator values.}

\item{type}{Column name of column with type data.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{cholera_latest_year}{Latest year to calculate rolling average for cholera.
If \code{NULL}, uses the latest year with observations in the data.}

\item{meningitis_latest_year}{Latest year to calculate rolling average for meningitis
If \code{NULL}, uses the latest year with observations in the data.}

\item{yellow_fever_latest_year}{Latest year to calculate rolling average for meningitis
If \code{NULL}, uses the latest year with observations in the data.}

\item{extrapolate_to}{Year to extrapolate Prevent data to, defaults to 2025}
}
\description{
Prevent campaign data uses aggregates across years that a vaccine provides
protection against a specific pathogen. Thus, we want to do some specific
aggregation so that this analysis can be brought into the overall HEP
calculations. This function does just that. For each pathogen, we take the
data out to the latest year observed, or a separate year if provided. Then we
do the rolling sums for them, and ensure that the rows make sense (filtering
out for instance years before there were any pathogens reported for a country).
}
\details{
These transform values are then flat extrapolated from their latest year out to
a specific year, the default being 2025 If latest year values are provided
for a specific pathogen, those years are used for calculating the rolling
average out to, otherwise, the latest year with observed values is used.
}
