% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_hep.R
\name{transform_prev_routine_data}
\alias{transform_prev_routine_data}
\title{Transform Prevent routine data}
\usage{
transform_prev_routine_data(
  df,
  iso3,
  year,
  ind,
  value,
  transform_value,
  ind_ids
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{year}{Column name of column with year data.}

\item{ind}{Column name of column with indicator names.}

\item{value}{Column name of column with indicator values.}

\item{transform_value}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\description{
Prevent routine data is now stored raw using the percent coverage of the indicator.
We want to transform this back into a numerator value for use within \code{pathogen_calc}.
}
