% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhc.R
\name{transform_uhc_data}
\alias{transform_uhc_data}
\title{Transform Raw Indicator Values for UHC Billion}
\usage{
transform_uhc_data(
  df,
  ind = "ind",
  value = "value",
  ind_ids = billion_ind_codes("uhc")
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{ind}{Column name of column with indicator names.}

\item{value}{Column name of column with indicator values.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
Data frame in long format.
}
\description{
\code{transform_uhc_data()} applies transformations on UHC Billion indicators so
that transformed indicator values can be used within Billions calculations.
Details on the specific transformations applied can be found within the
Billions methods report.
}
\details{
For more details on the UHC Billion calculation process and how this function
ties in with the rest, see the vignette:

\href{../doc/uhc.html}{\code{vignette("uhc", package = "billionaiRe")}}
}
