% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_uhc_tobacco}
\alias{transform_uhc_tobacco}
\title{Transform UHC tobacco data}
\usage{
transform_uhc_tobacco(x)
}
\description{
Reverses the indicator and rescales from \link{29,100} to \link{0,100}
}
