% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_hpop.R
\name{untransform_hpop_data}
\alias{untransform_hpop_data}
\title{Untransform Indicator Values for HPOP Billion}
\usage{
untransform_hpop_data(
  df,
  transform_value_col = "transform_value_col",
  value_col = stringr::str_remove(transform_value_col, "transform_"),
  ind_ids = billion_ind_codes("hpop")
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{transform_value_col}{Column name(s) of column with transformed values to retrieve.}

\item{value_col}{Column name(s) of column to place untransformed values. Must be same
length as \code{transform_value_col}. If a column already exists, values are overwritten
wherever \code{ind} and \code{transform_value_col} are available to be untransformed for
this Billion, but otherwise, the column retains its
other values.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
Data frame in long format.
}
\description{
\code{untransform_hpop_data()} reverses transformations on HPOP Billion indicators to
return raw indicator values. Details on the specific transformations applied
can be found within the Billions methods report.
}
\details{
For more details on the HPOP Billion calculation process and how this function
ties in with the rest, see the vignette:

\href{../doc/hpop.html}{\code{vignette("hpop", package = "billionaiRe")}}
}
