% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhc.R
\name{untransform_uhc_data}
\alias{untransform_uhc_data}
\title{Untransform Indicator Values for UHC Billion}
\usage{
untransform_uhc_data(
  df,
  iso3 = "iso3",
  ind = "ind",
  transform_value = "transform_value",
  value = "value",
  ind_ids = billion_ind_codes("uhc")
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{ind}{Column name of column with indicator names.}

\item{transform_value}{Column name of column with transformed values to retrieve.}

\item{value}{Column name of column to put in untransformed values.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
Data frame in long format.
}
\description{
\code{untransform_uhc_data()} reverses transformations on UHC Billion indicators to
return raw indicator values. Details on the specific transformations applied
can be found within the Billions methods report.
}
\details{
For more details on the UHC Billion calculation process and how this function
ties in with the rest, see the vignette:

\href{../doc/uhc.html}{\code{vignette("uhc", package = "billionaiRe")}}
}
