% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataframes_uhc_summary.R
\name{write_data_headers_uhc_summary}
\alias{write_data_headers_uhc_summary}
\title{Write and style headers for data in UHC summary sheets}
\usage{
write_data_headers_uhc_summary(
  wb,
  sheet_name,
  value,
  boxes_bounds,
  start_year,
  end_year
)
}
\arguments{
\item{wb}{a \code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{value}{Column name of column with indicator values.}

\item{boxes_bounds}{named list of bounds for data frame boxes to be written in sheet.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2023.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
Used in \code{write_uhc_summary_sheet()}
}
