% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_hpop_datasheet.R
\name{write_hpop_datasheet}
\alias{write_hpop_datasheet}
\title{Write data sheet}
\usage{
write_hpop_datasheet(
  df,
  wb,
  sheet_name,
  iso,
  start_year = 2018,
  end_year = 2019:2023,
  value = "value",
  year = "year",
  iso3 = "iso3",
  ind = "ind",
  population = "population",
  scenario = NULL,
  ind_ids = billion_ind_codes("hpop"),
  transform_value = "transform_value",
  type_col = "type",
  source_col = "source",
  contribution = "contribution",
  contribution_pct = paste0(contribution, "_percent"),
  contribution_pct_pop_total = paste0(contribution, "_percent_pop_total"),
  ind_df
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{wb}{\code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{iso}{ISO3 code of country to summarize.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2023.}

\item{value}{Column name of column with indicator values.}

\item{year}{Column name of column with years.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{ind}{Column name of column with indicator names.}

\item{population}{Column name of column with population figures.}

\item{scenario}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{transform_value}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{type_col}{Column name of column to store type.}

\item{source_col}{Column name of column to store source information.}

\item{contribution}{Column name of column(s) to store contribution (population)
values. Must be the same length as \code{transform_value}.}

\item{contribution_pct}{Column name of column(s) to store contribution (percent)
values. Must be the same length as \code{transform_value}.}

\item{contribution_pct_pop_total}{Column name of column(s) to store contribution
(percent of total population of the country) values. Must be the same length
as \code{transform_value}.}

\item{ind_df}{data frame containing the indicators in the correct order and format to be used.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{write_hpop_datasheet} write all the content and styling found in the HPOP data sheet
}
