% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_hpop_inter.R
\name{write_hpop_inter}
\alias{write_hpop_inter}
\title{Write Inter sheet for HPOP}
\usage{
write_hpop_inter(
  wb,
  sheet_name,
  data_sheet_name,
  ind_df,
  start_year,
  end_year,
  start_col,
  start_row,
  transform_value,
  summary_bounds
)
}
\arguments{
\item{wb}{\code{\link[openxlsx:createWorkbook]{openxlsx::createWorkbook()}} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{data_sheet_name}{character Name of the data sheet}

\item{ind_df}{data frame containing the indicators in the correct order and format to be used.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2023.}

\item{start_col}{integer start col of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}

\item{start_row}{integer start row of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}

\item{transform_value}{Column name of column(s) with transformed indicator values,
used to calculate contributions.}

\item{summary_bounds}{named list with boxes bounds of the different data 'boxes'
present in the summary sheet. Used to identify where data sits in Excel sheet
dynamically.}
}
\description{
Write Inter sheet for HPOP
}
