% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_indicator_list.R
\name{write_indicator_list_sheet}
\alias{write_indicator_list_sheet}
\title{Write indicator list sheet}
\usage{
write_indicator_list_sheet(wb, sheet_name, billion, start_row, start_col)
}
\arguments{
\item{wb}{a \code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{billion}{Billion indicator names to return, either "hep", "hpop", "uhc", or "all".}

\item{start_row}{integer start row of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}

\item{start_col}{integer start col of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}
}
\description{
Write indicator list sheet
}
