% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_main_data.R
\name{write_main_df}
\alias{write_main_df}
\title{Write the main data frame to the data sheet}
\usage{
write_main_df(
  df,
  wb,
  start_row,
  start_col,
  start_year,
  end_year,
  sheet_name,
  value,
  transform_value,
  type_col,
  source_col
)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{openxlsx workbook to be edited.}

\item{start_row}{numeric start row of the table in the excel sheet}

\item{start_col}{numeric start row of the table in the excel sheet}

\item{sheet_name}{character name of the sheet to update}

\item{value}{Column name of column with indicator values.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
Write the main data frame to the data sheet
}
