% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_export.R
\name{write_main_df}
\alias{write_main_df}
\title{Write the main data frame to the data sheet}
\usage{
write_main_df(
  data_table,
  wb,
  start_row,
  start_col,
  start_year,
  end_year,
  sheet_name
)
}
\arguments{
\item{data_table}{data frame to be written into the main data frame}

\item{wb}{openxlsx workbook to to be edited.}

\item{start_row}{numeric start row of the table in the excel sheet}

\item{start_col}{numeric start row of the table in the excel sheet}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2023.}

\item{sheet_name}{character name of the sheet to update}
}
\value{

}
\description{
Write the main data frame to the data sheet
}
