% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_wrangling.R
\name{xmart_col_types}
\alias{xmart_col_types}
\title{Get the col_types for xMart columns}
\usage{
xmart_col_types()
}
\value{
a list with column specifications
}
\description{
A helper function for specifying the column types when reading/writing Triple
Billions csv files.
}
