% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_wrangling.R
\name{xmart_cols}
\alias{xmart_cols}
\title{Get names of xMart4 columns}
\usage{
xmart_cols(data_type = c("wrangled_data", "projected_data", "final_data"))
}
\arguments{
\item{data_type}{(string): the type of data
\itemize{
\item \code{wrangled_data} (default): raw data that has been wrangled into a suitable
form for analysis.
\item \code{projected_data}:  data that has been fully projected to the target year but
has not yet been transformed or calculated upon.
\item \code{final_data}: the complete set of billions data with transformed values,
contributions, and all calculations available.
}}
}
\value{
character vector
}
\description{
Returns a character vector with the names of all the columns expected by the
Triple Billions xMart tables.
}
\seealso{
Wrangle data functions
\code{\link{add_missing_xmart_rows}()},
\code{\link{get_data_lake_name}()},
\code{\link{get_whdh_path}()},
\code{\link{has_xmart_cols}()},
\code{\link{save_gho_backup_to_whdh}()},
\code{\link{save_wrangled_output}()},
\code{\link{wrangle_gho_data}()},
\code{\link{wrangle_gho_rural_urban_data}()},
\code{\link{wrangle_unsd_data}()},
\code{\link{xmart_col_types}()}
}
\concept{whdh}
