% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_wrangling.R
\name{xmart_cols}
\alias{xmart_cols}
\title{Get names of xMart4 columns}
\usage{
xmart_cols(data_type = c("wrangled_data", "projected_data", "final_data"))
}
\arguments{
\item{data_type}{(string): the type of data}
}
\value{
character vector
}
\description{
Returns a character vector with the names of all the columns expected by the
Triple Billions xMart tables.
}
